### 2026-08-12 2a3ba54
Address straightforward goodpractice recommendations - checkpoint
