+++ tests/testthat/test-utils-size-progress.R
@@ -1,506 +1,586 @@
-test_that("duration formatting covers invalid, seconds, minutes, and hours", {
-  format_duration <- camtrapReport:::.format_duration
-  
-  expect_identical(
-    format_duration(NULL),
-    "unknown time"
-  )
-  
-  expect_identical(
-    format_duration(NA_real_),
-    "unknown time"
-  )
-  
-  expect_identical(
-    format_duration(Inf),
-    "unknown time"
-  )
-  
-  expect_identical(
-    format_duration(-1),
-    "unknown time"
-  )
-  
-  expect_identical(
-    format_duration(0),
-    "0 sec"
-  )
-  
-  expect_identical(
-    format_duration(59.6),
-    "1 min 00 sec"
-  )
-  
-  expect_identical(
-    format_duration(61),
-    "1 min 01 sec"
-  )
-  
-  expect_identical(
-    format_duration(3599),
-    "59 min 59 sec"
-  )
-  
-  expect_identical(
-    format_duration(3600),
-    "1 h 0 min"
-  )
-  
-  expect_identical(
-    format_duration(7380),
-    "2 h 3 min"
-  )
-})
-
-
-test_that("file-size formatting covers all size units", {
-  format_size <- camtrapReport:::.format_file_size
-  
-  expect_identical(
-    format_size(NULL),
-    "unknown size"
-  )
-  
-  expect_identical(
-    format_size(character()),
-    "unknown size"
-  )
-  
-  expect_identical(
-    format_size(NA_real_),
-    "unknown size"
-  )
-  
-  expect_identical(
-    format_size(Inf),
-    "unknown size"
-  )
-  
-  expect_identical(
-    format_size(-1),
-    "unknown size"
-  )
-  
-  expect_identical(
-    format_size(500),
-    "500 B"
-  )
-  
-  expect_identical(
-    format_size(1536),
-    "1.5 KB"
-  )
-  
-  expect_identical(
-    format_size(2.5 * 1024^2),
-    "2.5 MB"
-  )
-  
-  expect_identical(
-    format_size(1.25 * 1024^3),
-    "1.25 GB"
-  )
-})
-
-
-test_that("size estimation handles missing input safely", {
-  estimate_size <- camtrapReport:::.estimate_camdata_size
-  
-  inputs <- list(
-    NULL,
-    character(),
-    NA_character_,
-    tempfile("missing-camtrap-data-")
-  )
-  
-  for (input in inputs) {
-    result <- estimate_size(input)
-    
-    expect_named(
-      result,
-      c(
-        "file_size",
-        "file_size_label",
-        "zip_uncompressed_size",
-        "zip_uncompressed_label",
-        "effective_size",
-        "effective_size_label",
-        "size_class"
-      )
+test_that(
+  "duration formatting covers invalid, seconds, minutes, and hours",
+  {
+    format_duration <- getFromNamespace(
+      ".format_duration",
+      "camtrapReport"
     )
-    
-    expect_true(is.na(result$file_size))
     expect_identical(
-      result$file_size_label,
+      format_duration(NULL),
+      "unknown time"
+    )
+
+    expect_identical(
+      format_duration(NA_real_),
+      "unknown time"
+    )
+
+    expect_identical(
+      format_duration(Inf),
+      "unknown time"
+    )
+
+    expect_identical(
+      format_duration(-1),
+      "unknown time"
+    )
+
+    expect_identical(
+      format_duration(0),
+      "0 sec"
+    )
+
+    expect_identical(
+      format_duration(59.6),
+      "1 min 00 sec"
+    )
+
+    expect_identical(
+      format_duration(61),
+      "1 min 01 sec"
+    )
+
+    expect_identical(
+      format_duration(3599),
+      "59 min 59 sec"
[…]
