### 2026-08-12 aeee418
Checkpoint before final goodpractice test cleanup
