+++ tests/testthat/test-gui.R
@@ -8,13 +8,13 @@ test_that("the GUI app is built and exposes its read-only server outputs", {
   shiny::testServer(app$serverFuncSource(), {
     session$flushReact()
 
-    expect_true(length(output$object_status) > 0L)
-    expect_true(length(output$years_ui) > 0L)
-    expect_true(length(output$existing_group_ui) > 0L)
-    expect_true(length(output$group_values_ui) > 0L)
-    expect_true(length(output$species_table) > 0L)
-    expect_true(length(output$sections_ui) > 0L)
-    expect_true(length(output$section_edit_select_ui) > 0L)
+    expect_gt(length(output$object_status), 0L)
+    expect_gt(length(output$years_ui), 0L)
+    expect_gt(length(output$existing_group_ui), 0L)
+    expect_gt(length(output$group_values_ui), 0L)
+    expect_gt(length(output$species_table), 0L)
+    expect_gt(length(output$sections_ui), 0L)
+    expect_gt(length(output$section_edit_select_ui), 0L)
     expect_match(output$section_list, "name")
     expect_match(output$cm_summary, cm$siteName, fixed = TRUE)
   })
