### 2026-08-12 99125c0
Checkpoint before safe lint cleanup
