+++ tests/testthat/test-camR-rem-branches.R
@@ -28,8 +28,9 @@ test_that("get_REM returns existing stored results without refitting", {
   
   valid <- which(!is.na(candidate_groups))
   
-  expect_true(
-    length(valid) > 0L
+  expect_gt(
+    length(valid),
+    0L
   )
   
   species <- candidate_species[valid[1]]
@@ -93,8 +94,9 @@ test_that("get_REM returns existing yearly results", {
   
   valid <- which(!is.na(candidate_groups))
   
-  expect_true(
-    length(valid) > 0L
+  expect_gt(
+    length(valid),
+    0L
   )
   
   species <- candidate_species[valid[1]]
