+++ tests/testthat/test-camData-datetime-helpers.R
@@ -30,7 +30,7 @@ test_that("parse_cam_datetime converts other POSIXt input", {
   
   expect_s3_class(result, "POSIXct")
   
-  expect_equal(
+  expect_identical(
     as.numeric(result),
     as.numeric(as.POSIXct(input))
   )
