+++ R/camR_functions.R
@@ -216,7 +216,7 @@
   }
   
   if ("Habitat_Type" %in% names(dep_loc)) {
-    dep_loc$Habitat_Type <- gsub("_", " ", dep_loc$Habitat_Type)
+    dep_loc$Habitat_Type <- gsub("_", " ", dep_loc$Habitat_Type, fixed = TRUE)
     dep_loc$Habitat_Type <- ifelse(
       dep_loc$Habitat_Type == "Other",
       "Unclassified Habitat",
@@ -557,9 +557,9 @@
   
   .n <- .nn <- colnames(df)[grepl('^vernacularName',colnames(df))]
   if (length(.n) > 0) {
-    .w <- which(sapply(.n,function(x) length(strsplit(x,'\\.')[[1]])) == 2)
+    .w <- which(sapply(.n,function(x) length(strsplit(x,'.', fixed = TRUE)[[1]])) == 2)
     if (length(.w) > 0) {
-      .nn[.w] <- paste0('species_list_',sapply(.n,function(x) strsplit(x,'\\.')[[1]][2]))
+      .nn[.w] <- paste0('species_list_',sapply(.n,function(x) strsplit(x,'.', fixed = TRUE)[[1]][2]))
     }
   }
   
@@ -755,11 +755,11 @@
   location_df2[["locationName"]][location_df2[["locationName"]] == ""] <- NA_character_
   
   location_df2[["longitude"]] <- suppressWarnings(
-    as.numeric(gsub(",", ".", trimws(as.character(location_df2[["longitude"]]))))
+    as.numeric(gsub(",", ".", trimws(as.character(location_df2[["longitude"]])), fixed = TRUE))
   )
   
   location_df2[["latitude"]] <- suppressWarnings(
-    as.numeric(gsub(",", ".", trimws(as.character(location_df2[["latitude"]]))))
+    as.numeric(gsub(",", ".", trimws(as.character(location_df2[["latitude"]])), fixed = TRUE))
   )
   
   complete_idx <- stats::complete.cases(
@@ -1801,7 +1801,7 @@
     pair_idx <- which(!is.na(lon_num) & !is.na(lat_num))
     if (!length(pair_idx)) return("")
     deg_like <- abs(lon_num[pair_idx]) <= 180 & abs(lat_num[pair_idx]) <= 90
-    if (all(deg_like) || all(!deg_like)) return("")
+    if (all(deg_like) || !any(deg_like)) return("")
     idx_deg    <- pair_idx[deg_like]
     idx_nondeg <- pair_idx[!deg_like]
     diff_rows  <- if (length(idx_deg) < length(idx_nondeg)) idx_deg else idx_nondeg
@@ -1944,7 +1944,7 @@
       if (is.na(col_count)) paste0(r, " Missing column: count")
     else .summ_status_counts(cm$data$observations[[col_count]], treat_blank = FALSE)
     
-    col_cb <- .pick_col(cm$data$observations, c("classifiedBy"))
+    col_cb <- .pick_col(cm$data$observations, "classifiedBy")
     if (is.na(col_cb)) {
       cm$data_status$Essentials$obs$classifiedBy_status <- paste0(r, " Missing column: classifiedBy")
     } else {
@@ -1972,7 +1972,7 @@
       n_animal <- sum(idx_animal)
       
       # taxonID: animals + unique
-      col_tax <- .pick_col(cm$data$observations, c("taxonID"))
+      col_tax <- .pick_col(cm$data$observations, "taxonID")
       if (is.na(col_tax)) {
         cm$data_status$Essentials$obs$taxonID <- paste0(r, " Missing column: taxonID")
       } else {
@@ -1998,13 +1998,13 @@
         obs_state$value[[key]] <- .animal_field_status(present_n, n_animal, label)
       }
       
[…]
