+++ tests/testthat/helper-cam-report.R
@@ -69,8 +69,9 @@ camtrap_test_report <- function() {
           ),
           warning = function(w) {
             if (grepl(
-              "chi\\^2 approximation may be inaccurate",
-              conditionMessage(w)
+              "chi^2 approximation may be inaccurate",
+              conditionMessage(w),
+              fixed = TRUE
             )) {
               invokeRestart("muffleWarning")
             }
