+++ R/sections.R
@@ -92,10 +92,8 @@ setMethod("section_names",signature(keep = "ANY"),
         }
         
         warning(
-          paste0(
-            "Several section/module names specified in 'keep' are not available: ",
-            .paste_comma_and(keep[!w])
-          )
+          "Several section/module names specified in 'keep' are not available: ",
+          .paste_comma_and(keep[!w])
         )
       }
       
@@ -122,10 +120,8 @@ setMethod("section_names",signature(keep = "ANY"),
         }
         
         warning(
-          paste0(
-            "Several section/module names specified in 'exclude' are not available: ",
-            .paste_comma_and(exclude[!w])
-          )
+          "Several section/module names specified in 'exclude' are not available: ",
+          .paste_comma_and(exclude[!w])
         )
       }
       
