+++ R/utils.R
@@ -920,7 +920,7 @@
   
   for (i in seq_along(.dtFormats)) {
     parsed <- suppressWarnings(as.POSIXct(x, format = .dtFormats[i], tz = "UTC"))
-    o[i] <- !any(is.na(parsed))
+    o[i] <- !anyNA(parsed)
   }
   
   if (any(o)) {
