### 2026-08-12 1ca431e
Add pkgcheck badge
