### 2026-08-13 4b20258
Remove triple-colon usage from tests
