+++ tests/testthat/test-camR-merged-spatial-summary.R
@@ -55,7 +55,7 @@ make_merged_summary_fixture <- function() {
 test_that("merged spatial summary returns the expected structure", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   expect_s3_class(result, "data.frame")
   expect_gt(nrow(result), 0L)
@@ -89,7 +89,7 @@ test_that("merged spatial summary returns the expected structure", {
 test_that("merged spatial summary calculates deployment counts", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   counts <- unique(
     result[
@@ -122,7 +122,7 @@ test_that("merged spatial summary calculates deployment counts", {
 test_that("merged spatial summary lists deployment identifiers", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   loc1 <- result[
     result$locationID == "loc-1",
@@ -152,7 +152,7 @@ test_that("merged spatial summary lists deployment identifiers", {
 test_that("merged spatial summary aggregates capture methods", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   loc1 <- result[
     result$locationID == "loc-1",
@@ -180,7 +180,7 @@ test_that("merged spatial summary aggregates capture methods", {
 test_that("merged spatial summary aggregates annotators", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   loc1 <- result[
     result$locationID == "loc-1",
@@ -208,7 +208,7 @@ test_that("merged spatial summary aggregates annotators", {
 test_that("merged spatial summary calculates photograph totals", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   totals <- unique(
     result[
@@ -241,7 +241,7 @@ test_that("merged spatial summary calculates photograph totals", {
 test_that("merged spatial summary creates species lists", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   loc1 <- result[
     result$locationID == "loc-1",
@@ -272,7 +272,7 @@ test_that("merged spatial summary creates species lists", {
 test_that("merged spatial summary converts habitat labels", {
   cm <- make_merged_summary_fixture()
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
   loc1_habitat <- unique(
     result$Habitat_Type[
@@ -302,7 +302,7 @@ test_that("merged spatial summary handles missing sequences", {
   cm <- make_merged_summary_fixture()
   cm$data$sequences <- NULL
   
-  result <- camtrapReport:::.camr_getMergedSummary(cm)
+  result <- ct_internal(".camr_getMergedSummary")(cm)
   
[…]
