+++ tests/testthat/test-supporting-functions.R
@@ -1,10 +1,10 @@
 test_that("species and class helpers summarise the bundled data", {
   cm <- camtrap_test_report()
 
-  species_counts <- camtrapReport:::.get_species(cm$data, count = TRUE)
-  species_names <- camtrapReport:::.get_species(cm$data, count = FALSE)
-  class_counts <- camtrapReport:::.get_classes(cm$data, count = TRUE)
-  class_names <- camtrapReport:::.get_classes(cm$data, count = FALSE)
+  species_counts <- ct_internal(".get_species")(cm$data, count = TRUE)
+  species_names <- ct_internal(".get_species")(cm$data, count = FALSE)
+  class_counts <- ct_internal(".get_classes")(cm$data, count = TRUE)
+  class_names <- ct_internal(".get_classes")(cm$data, count = FALSE)
 
   expect_s3_class(species_counts, "data.frame")
   expect_true(all(species_names %in% species_counts$scientificName))
@@ -15,7 +15,7 @@ test_that("species and class helpers summarise the bundled data", {
 
 test_that("merged data can retain nested media records", {
   cm <- camtrap_test_report()
-  merged <- camtrapReport:::.merge_data(cm$data, dropMedia = FALSE)
+  merged <- ct_internal(".merge_data")(cm$data, dropMedia = FALSE)
 
   expect_s3_class(merged, "data.frame")
   expect_true("media" %in% names(merged))
@@ -25,27 +25,27 @@ test_that("merged data can retain nested media records", {
 
 test_that("effort helpers aggregate deployments and draw with base graphics", {
   cm <- camtrap_test_report()
-  total <- camtrapReport:::.calc_effort(cm$data, unit = "day")
-  by_location <- camtrapReport:::.calc_effort(
+  total <- ct_internal(".calc_effort")(cm$data, unit = "day")
+  by_location <- ct_internal(".calc_effort")(
     cm$data,
     by = "locationID",
     unit = "hour"
   )
-  step_table <- camtrapReport:::.effort_table(cm, startend = TRUE)
+  step_table <- ct_internal(".effort_table")(cm, startend = TRUE)
 
   expect_gt(total$effort, 0)
   expect_identical(total$unit, "day")
   expect_true(all(c("locationID", "effort") %in% names(by_location)))
   expect_true(all(c("time", "nrCams") %in% names(step_table)))
   expect_error(
-    camtrapReport:::.calc_effort(cm$data, by = "missing_column"),
+    ct_internal(".calc_effort")(cm$data, by = "missing_column"),
     "Grouping columns"
   )
 
   file <- tempfile(fileext = ".pdf")
   grDevices::pdf(file)
   on.exit(grDevices::dev.off(), add = TRUE)
-  expect_null(camtrapReport:::.plot_effort(cm, dynamic = FALSE))
+  expect_null(ct_internal(".plot_effort")(cm, dynamic = FALSE))
 })
 
 test_that("base left join handles equal and differently named keys", {
@@ -53,13 +53,13 @@ test_that("base left join handles equal and differently named keys", {
   right <- data.frame(id = 2, extra = "x")
   right2 <- data.frame(other_id = 2, extra = "x")
 
-  joined <- camtrapReport:::.left_join(left, right, "id")
-  joined2 <- camtrapReport:::.left_join(left, right2, c("id", "other_id"))
+  joined <- ct_internal(".left_join")(left, right, "id")
+  joined2 <- ct_internal(".left_join")(left, right2, c("id", "other_id"))
 
   expect_identical(nrow(joined), 2L)
   expect_identical(nrow(joined2), 2L)
   expect_true(is.na(joined$extra[joined$id == 1]))
-  expect_error(camtrapReport:::.left_join(left, right, "missing"))
[…]
