+++ tests/testthat/test-report-output-paths.R
@@ -1,28 +1,33 @@
 library(testthat)
-
 capture_report_conditions <- function(expr) {
-  messages <- character()
-  warnings <- character()
+  state <- new.env(parent = emptyenv())
+  state$messages <- character()
+  state$warnings <- character()
   
   value <- withCallingHandlers(
     expr,
     message = function(condition) {
-      messages <<- c(messages, conditionMessage(condition))
+      state$messages <- c(
+        state$messages,
+        conditionMessage(condition)
+      )
       invokeRestart("muffleMessage")
     },
     warning = function(condition) {
-      warnings <<- c(warnings, conditionMessage(condition))
+      state$warnings <- c(
+        state$warnings,
+        conditionMessage(condition)
+      )
       invokeRestart("muffleWarning")
     }
   )
   
   list(
     value = value,
-    messages = messages,
-    warnings = warnings
+    messages = state$messages,
+    warnings = state$warnings
   )
 }
-
 
 test_that("report returns rendering errors without crashing", {
   original <- camtrap_test_report()
@@ -128,11 +133,12 @@ test_that("report does not invoke viewer after rendering failure", {
   dir.create(output_dir)
   
   old_viewer <- getOption("viewer")
-  viewer_called <- FALSE
+  viewer_state <- new.env(parent = emptyenv())
+  viewer_state$called <- FALSE
   
   options(
     viewer = function(path) {
-      viewer_called <<- TRUE
+      viewer_state$called <- TRUE
     }
   )
   
@@ -157,5 +163,5 @@ test_that("report does not invoke viewer after rendering failure", {
   )
   
   expect_s3_class(captured$value, "try-error")
-  expect_false(viewer_called)
+  expect_false(viewer_state$called)
 })
