+++ tests/testthat/test-install-all-branches.R
@@ -93,7 +93,8 @@ test_that("install_All rejects invalid arguments before installation", {
 
 
 test_that("install_All reports when all requested packages are installed", {
-  checked_packages <- character()
+  state <- new.env(parent = emptyenv())
+  state$checked_packages <- character()
   
   testthat::local_mocked_bindings(
     .getPackageList = function() {
@@ -105,9 +106,9 @@ test_that("install_All reports when all requested packages are installed", {
     .is.installed = function(n) {
       n <- as.character(n)
       
-      checked_packages <<- unique(
+      state$checked_packages <- unique(
         c(
-          checked_packages,
+          state$checked_packages,
           n
         )
       )
@@ -132,7 +133,7 @@ test_that("install_All reports when all requested packages are installed", {
   
   expect_true(
     all(
-      c("methods", "stats") %in% checked_packages
+      c("methods", "stats") %in% state$checked_packages
     )
   )
 })
@@ -200,7 +201,8 @@ test_that("install_All attempts only missing CRAN packages", {
 
 
 test_that("install_All reports CRAN installation failures safely", {
-  installation_calls <- character()
+  state <- new.env(parent = emptyenv())
+  state$installation_calls <- character()
   
   testthat::local_mocked_bindings(
     .getPackageList = function() {
@@ -217,8 +219,8 @@ test_that("install_All reports CRAN installation failures safely", {
       result
     },
     install.packages = function(pkgs, ...) {
-      installation_calls <<- c(
-        installation_calls,
+      state$installation_calls <- c(
+        state$installation_calls,
         as.character(pkgs)
       )
       
@@ -236,7 +238,7 @@ test_that("install_All reports CRAN installation failures safely", {
   expect_null(result)
   
   expect_setequal(
-    installation_calls,
+    state$installation_calls,
     c("missingOne", "missingTwo")
   )
 })
@@ -245,7 +247,8 @@ test_that("install_All reports CRAN installation failures safely", {
 test_that(
   "install_All handles missing GitHub packages without network access",
   {
-    github_calls <- character()
+    state <- new.env(parent = emptyenv())
+    state$github_calls <- character()
     
     testthat::local_mocked_bindings(
       .getPackageList = function() {
@@ -264,8 +267,8 @@ test_that(
         result
       },
       .installGitHub = function(repository) {
-        github_calls <<- c(
-          github_calls,
+        state$github_calls <- c(
+          state$github_calls,
           repository
         )
         
@@ -283,7 +286,7 @@ test_that(
     expect_null(result)
     
     expect_identical(
-      github_calls,
+      state$github_calls,
       "example/exampleRepository"
     )
   }
@@ -291,7 +294,8 @@ test_that(
 
 
 test_that("install_All counts successful mocked GitHub installations", {
-  github_installed <- FALSE
+  state <- new.env(parent = emptyenv())
+  state$github_installed <- FALSE
   
   testthat::local_mocked_bindings(
     .getPackageList = function() {
@@ -309,7 +313,7 @@ test_that("install_All counts successful mocked GitHub installations", {
         n,
         function(package) {
           identical(package, "githubPackage") &&
-            github_installed
+            state$github_installed
         },
         logical(1)
       )
@@ -323,7 +327,7 @@ test_that("install_All counts successful mocked GitHub installations", {
         "example/exampleRepository"
       )
       
-      github_installed <<- TRUE
+      state$github_installed <- TRUE
       TRUE
     },
     .package = "camtrapReport"
@@ -364,8 +368,9 @@ test_that(
 
 
 test_that("install_All update mode reinstalls mocked CRAN packages", {
-  installation_calls <- character()
[…]
