### 2026-08-14 66175fd
Simplify code-section matching
