### 2026-08-14 e746e3e
Remove unnecessary empty concatenations
