### 2026-08-14 e666999
Use exact expectations in tests
