### 2026-08-14 864cd27
Simplify string concatenation
