### 2026-08-14 5400d3d
Handle intentional testthat library usage
