+++ tests/testthat.R
@@ -6,7 +6,7 @@
 # * https://r-pkgs.org/testing-design.html#sec-tests-files-overview
 # * https://testthat.r-lib.org/articles/special-files.html
 
-library(testthat)
-library(camtrapReport)
+library(testthat) # nolint: undesirable_function_linter.
+library(camtrapReport) # nolint: undesirable_function_linter.
 
 test_check("camtrapReport")
