### 2026-08-14 421c8b2
Make string factor handling explicit
