### 2026-08-14 eb86bf2
Simplify conditional branches
