### 2026-08-14 823a3bd
Suppress intentional lint exceptions
