### 2026-08-14 0b12005
Use toString in supporting functions
