### 2026-08-14 890ce7a
Use toString in module messages
