### 2026-08-14 d90d87c
Use toString in GUI summaries
