### 2026-08-14 3951e16
Address goodpractice lint recommendations
