+++ R/camR_functions.R
@@ -47,14 +47,14 @@
     if (length(missing_x) > 0) {
       stop(
         "Missing join column(s) in x: ",
-        paste(missing_x, collapse = ", ")
+        toString(missing_x)
       )
     }
     
     if (length(missing_y) > 0) {
       stop(
         "Missing join column(s) in y: ",
-        paste(missing_y, collapse = ", ")
+        toString(missing_y)
       )
     }
     
@@ -576,10 +576,10 @@
       .w <- which(df$observation_Year == .years[i])
       .df$total_observations[i] <- length(.w)
       .df$total_species[i] <- length(unique(df$scientificName[.w]))
-      .df$species_list_scientificName[i] <- paste(sort(unique(df$scientificName[.w])), collapse = ", ")
+      .df$species_list_scientificName[i] <- toString(sort(unique(df$scientificName[.w])))
       if (length(.nn) > 0) {
         for (j in seq_along(.nn)) {
-          .df[[.nn[j]]] <- paste(sort(unique(df[[.n[j]]][.w])), collapse = ", ")
+          .df[[.nn[j]]] <- toString(sort(unique(df[[.n[j]]][.w])))
         }
       }
     }
@@ -776,7 +776,7 @@
   } else {
     paste0(
       r, " ", length(missing_rows), " rows with missing data: [",
-      paste(missing_rows, collapse = ", "), "]"
+      toString(missing_rows), "]"
     )
   }
   
@@ -960,7 +960,7 @@
         " High-risk (",
         out$num_highrisk_outliers,
         "): ",
-        paste(high_names, collapse = ", ")
+        toString(high_names)
       )
     }
     
@@ -972,7 +972,7 @@
         " Medium-risk (",
         out$num_mediumrisk_outliers,
         "): ",
-        paste(med_names, collapse = ", ")
+        toString(med_names)
       )
     }
     
@@ -984,7 +984,7 @@
         " Low-risk (",
         out$num_lowrisk_outliers,
         "): ",
-        paste(low_names, collapse = ", ")
+        toString(low_names)
       )
     }
     
@@ -1506,7 +1506,7 @@
     
     out_years <- setdiff(y, clusters[[cand]])
     if (length(out_years)) {
-      paste0("Years: ", paste(out_years, collapse = ", "), " ", ic$yellow)
+      paste0("Years: ", toString(out_years), " ", ic$yellow)
     } else {
       paste("None", ic$green)
     }
@@ -1527,7 +1527,7 @@
     parts <- vapply(runs, function(r) {
       if (length(r) == 1) as.character(r[1]) else paste0(r[1], "-", r[length(r)])
     }, character(1))
-    paste(parts, collapse = ", ")
+    toString(parts)
   }
   
   .years_message <- function(years_chr) {
@@ -1561,7 +1561,7 @@
       else paste0(.month_abb(r[1]), "-", .month_abb(r[length(r)]))
     }, character(1))
     
-    paste(parts, collapse = ", ")
+    toString(parts)
   }
   
   .dep_month_coverage <- function(dep_ints, years_keep_chr) {
@@ -1628,8 +1628,8 @@
       paste0(ic$green, " Years in observations and deployments are the same")
     } else {
       parts2 <- character(0)
-      if (length(dep_not_obs)) parts2 <- c(parts2, paste0("Deployments exist, but observations are missing for: ", paste(dep_not_obs, collapse = ", ")))
-      if (length(obs_not_dep)) parts2 <- c(parts2, paste0("Observations exist, but deployments are missing for: ", paste(obs_not_dep, collapse = ", ")))
[…]
