### 2026-08-14 bb25e90
Replace type-stable sapply calls with vapply
