+++ R/show.R
@@ -19,7 +19,7 @@ setMethod ('show' , '.textSection',
 setMethod ('show' , 'camInfo',
            function ( object ) {
              if (length(object) > 0) {
-               .cls <- sapply(object,function(x) class(x)[1])
+               .cls <- vapply(object,function(x) class(x)[1], character(1))
                w <- which(.cls == 'character')
                cat('===========================================================','\n')
                if (length(w) > 0) {
