### 2026-08-14 4bbb1bb
Address goodpractice lint recommendations
