+++ R/camR.R
@@ -975,7 +975,7 @@ camR <- setRefClass(
       
       #=============================
       # Attach summary counts to species_summary object
-      .self$species_summary$count <- data.frame(Group=names(.self$species_summary),Count=sapply(.self$species_summary,function(x) nrow(x$site_list)))
+      .self$species_summary$count <- data.frame(Group=names(.self$species_summary),Count=vapply(.self$species_summary,function(x) nrow(x$site_list), integer(1)))
       
       # add a Name column to be used in the report text:
       .self$species_summary$count$Name <- .self$species_summary$count$Group
@@ -1387,7 +1387,7 @@ camR <- setRefClass(
         .added <- FALSE
         if (x@parent %in% names(.self$reportObjects)) {
           if (is.list(.self$reportObjects[[x@parent]])) {
-            .tmp <- sapply(.self$reportObjects[[x@parent]],function(.x) .x@name)
+            .tmp <- simplify2array(lapply(.self$reportObjects[[x@parent]], function(.x) .x@name))
             .w <- which(.tmp == x@parent)
             if (length(.w) > 1) {
               if (is.null(.self$reportObjects[[x@parent]][[.w[1]]]@Rchunk)) {
@@ -1396,7 +1396,7 @@ camR <- setRefClass(
               } else {
                 if (is.list(.self$reportObjects[[x@parent]][[.w[1]]]@Rchunk)) {
                   if (length(.self$reportObjects[[x@parent]][[.w[1]]]@Rchunk) > 0) {
-                    .n <- sapply(.self$reportObjects[[x@parent]][[.w[1]]]@Rchunk,function(x) x@name)
+                    .n <- simplify2array(lapply(.self$reportObjects[[x@parent]][[.w[1]]]@Rchunk, function(x) x@name))
                     if (x@name %in% .n) {
                       .self$reportObjects[[x@parent]][[.w[1]]]@Rchunk[[match(x@name, .n)]] <- x
                       .added <- TRUE
@@ -1429,7 +1429,7 @@ camR <- setRefClass(
             } else {
               if (is.list(.self$reportObjects[[x@parent]]@Rchunk)) {
                 if (length(.self$reportObjects[[x@parent]]@Rchunk) > 0) {
-                  .n <- sapply(.self$reportObjects[[x@parent]]@Rchunk,function(x) x@name)
+                  .n <- simplify2array(lapply(.self$reportObjects[[x@parent]]@Rchunk, function(x) x@name))
                   if (x@name %in% .n) {
                     .self$reportObjects[[x@parent]]@Rchunk[[match(x@name, .n)]] <- x
                     .added <- TRUE
@@ -1460,7 +1460,7 @@ camR <- setRefClass(
             if (is.list(.self$reportObjects[[.n]])) {
               if (x@parent %in% names(.self$reportObjects[[.n]])) {
                 if (is.list(.self$reportObjects[[.n]][[x@parent]])) {
-                  .tmp <- sapply(.self$reportObjects[[.n]][[x@parent]],function(.x) .x@name)
+                  .tmp <- simplify2array(lapply(.self$reportObjects[[.n]][[x@parent]], function(.x) .x@name))
                   .w <- which(.tmp == x@parent)
                   if (length(.w) > 1) {
                     if (is.null(.self$reportObjects[[.n]][[x@parent]][[.w[1]]]@Rchunk)) {
@@ -1470,7 +1470,7 @@ camR <- setRefClass(
                     } else {
                       if (is.list(.self$reportObjects[[.n]][[x@parent]][[.w[1]]]@Rchunk)) {
[…]
