### 2026-08-14 6990c28
Reduce line-length lint warnings
