### 2026-08-14 22e3d8c
Wrap report section conditions for lint
