+++ R/reportSection.R
@@ -225,7 +225,11 @@
   
   section_name <- x@name
   
-  if (is.null(section_name) || length(section_name) == 0 || is.na(section_name[1])) {
+  if (
+    is.null(section_name) ||
+      length(section_name) == 0 ||
+      is.na(section_name[1])
+  ) {
     section_name <- "unknown"
   }
   
@@ -237,11 +241,20 @@
   
   title <- x@title
   
-  if (!is.null(title) && length(title) > 0 && !is.na(title[1]) && nzchar(title[1])) {
+  if (
+    !is.null(title) &&
+      length(title) > 0 &&
+      !is.na(title[1]) &&
+      nzchar(title[1])
+  ) {
     
     head_level <- x@headLevel
     
-    if (is.null(head_level) || length(head_level) == 0 || is.na(head_level[1])) {
+    if (
+      is.null(head_level) ||
+        length(head_level) == 0 ||
+        is.na(head_level[1])
+    ) {
       head_level <- 1
     }
     
