### 2026-08-14 f8261bd
Wrap module and utility tests for lint
