### 2026-08-14 0499bdc
Wrap remaining test and vignette lines for lint
