+++ tests/testthat/test-update-api.R
@@ -1,4 +1,9 @@
-test_that("report-section matching supports names, titles, and informative errors", {
+test_that(
+  paste0(
+    "report-section matching supports names, titles, ",
+    "and informative errors"
+  ),
+  {
   catalog <- data.frame(
     name = c("intro", "sampling", "sampling_effort"),
     title = c("Introduction", "Sampling", "Sampling effort"),
@@ -8,7 +13,10 @@ test_that("report-section matching supports names, titles, and informative error
   match_section <- ct_internal(".matchReportSection")
 
   expect_identical(match_section(catalog, "INTRO")$name, "intro")
-  expect_identical(match_section(catalog, "Introduction", by = "title")$name, "intro")
+  expect_identical(
+    match_section(catalog, "Introduction", by = "title")$name,
+    "intro"
+  )
   expect_identical(
     suppressWarnings(match_section(catalog, "effort", by = "name"))$name,
     "sampling_effort"
@@ -18,11 +26,22 @@ test_that("report-section matching supports names, titles, and informative error
     "intro"
   )
   expect_error(match_section(catalog, ""), "empty")
-  expect_error(suppressWarnings(match_section(catalog, "unknown")), "No report section")
-  expect_error(suppressWarnings(match_section(catalog, "amp", by = "title")), "More than one")
+  expect_error(
+    suppressWarnings(match_section(catalog, "unknown")),
+    "No report section"
+  )
+  expect_error(
+    suppressWarnings(match_section(catalog, "amp", by = "title")),
+    "More than one"
+  )
 })
 
-test_that("code and chunk settings are captured without evaluation side effects", {
+test_that(
+  paste0(
+    "code and chunk settings are captured without ",
+    "evaluation side effects"
+  ),
+  {
   capture_code <- ct_internal(".capture_code_text")
   capture_setting <- ct_internal(".capture_setting_text")
   env <- list2env(list(code_value = c("x <- 1", "x + 1")), parent = baseenv())
@@ -166,5 +185,8 @@ test_that("updateReportSection updates title, text, code, and packages", {
   expect_error(updateReportSection(updated, section = 1), "single character")
   empty <- camR$new()
   empty$reportObjects <- list()
-  expect_error(updateReportSection(empty, "missing", text = "x"), "No report sections")
+  expect_error(
+    updateReportSection(empty, "missing", text = "x"),
+    "No report sections"
+  )
 })
