+++ R/camData.R
@@ -112,7 +112,8 @@
 
 
 .getSequences <- function(media) {
-  if (!.require('data.table')) stop('The data.table package is not installed...!')
+  if (!.require('data.table'))
+    stop('The data.table package is not installed...!')
 
   sequences <- .eval('media |>
     dplyr::distinct() |>
@@ -265,21 +266,42 @@
     if (!is.null(path) && is.character(path)) {
       .path <- file.path(
         path.expand(path),
-        gsub(basename(file), pattern = ".zip", replacement = "", ignore.case = TRUE)
+        gsub(
+          basename(file),
+          pattern = ".zip",
+          replacement = "",
+          ignore.case = TRUE
+        )
       )
     } else {
-      .path <- gsub(basename(file), pattern = ".zip", replacement = "", ignore.case = TRUE)
+      .path <- gsub(
+        basename(file),
+        pattern = ".zip",
+        replacement = "",
+        ignore.case = TRUE
+      )
     }
 
     file <- utils::unzip(file, exdir = .path)
 
   } else if (dir.exists(file)) {
-    if (all(c("datapackage.json", "deployments.csv", "observations.csv") %in% tolower(dir(file)))) {
+    if (all(
+      c("datapackage.json", "deployments.csv", "observations.csv") %in%
+        tolower(dir(file))
+    )) {
       .path <- file
       file <- dir(file, full.names = TRUE)
     } else {
-      if (any(c("datapackage.json", "deployments.csv", "observations.csv") %in% tolower(dir(file)))) {
-        required_files <- c("datapackage.json", "deployments.csv", "observations.csv", "media.csv")
+      if (any(
+        c("datapackage.json", "deployments.csv", "observations.csv") %in%
+          tolower(dir(file))
+      )) {
+        required_files <- c(
+          "datapackage.json",
+          "deployments.csv",
+          "observations.csv",
+          "media.csv"
+        )
         .w <- !required_files %in% tolower(dir(file))
 
         stop(
@@ -288,7 +310,9 @@
           ") are not available in the specified folder."
         )
       } else {
-        stop("The specified folder does not have the standard Camtrap DP files.")
+        stop(
+          "The specified folder does not have the standard Camtrap DP files."
+        )
       }
     }
   } else {
@@ -324,11 +348,17 @@
   }
 
   if ("deploymentStart" %in% names(.d$deployments)) {
-    .d$deployments$deploymentStart <- .parse_cam_datetime(.d$deployments$deploymentStart, tz = tz)
+    .d$deployments$deploymentStart <- .parse_cam_datetime(
+      .d$deployments$deploymentStart,
+      tz = tz
+    )
   }
 
   if ("deploymentEnd" %in% names(.d$deployments)) {
-    .d$deployments$deploymentEnd <- .parse_cam_datetime(.d$deployments$deploymentEnd, tz = tz)
+    .d$deployments$deploymentEnd <- .parse_cam_datetime(
+      .d$deployments$deploymentEnd,
+      tz = tz
+    )
   }
 
   if ("timestamp" %in% names(.d$media)) {
@@ -336,11 +366,17 @@
   }
 
   if ("eventStart" %in% names(.d$observations)) {
-    .d$observations$eventStart <- .parse_cam_datetime(.d$observations$eventStart, tz = tz)
+    .d$observations$eventStart <- .parse_cam_datetime(
+      .d$observations$eventStart,
+      tz = tz
+    )
   }
 
   if ("eventEnd" %in% names(.d$observations)) {
[…]
