### 2026-08-14 770c02b
Wrap supporting function lines for lint
