+++ R/supporting_functions.R
@@ -6,7 +6,11 @@
 # Effort = deployment duration
 # Returns both effort in requested units and raw seconds
 # ------------------------------------------------------------------
-.calc_effort <- function(x, by = NULL, unit = c("day", "hour", "minute", "second", "week")) {
+.calc_effort <- function(
+  x,
+  by = NULL,
+  unit = c("day", "hour", "minute", "second", "week")
+) {
   unit <- match.arg(unit)
   
   dep <- x$deployments
@@ -15,17 +19,22 @@
   # start_col <- .first_existing(dep, c("deploymentStart", "start"))
   # end_col   <- .first_existing(dep, c("deploymentEnd", "end"))
   
+  # nolint start: line_length_linter.
   dep <- .eval('dep |>
     dplyr::mutate(
       effort_seconds = as.numeric(difftime(.data[["deploymentEnd"]], .data[["deploymentStart"]], units = "secs"))
     ) |>
     dplyr::filter(!is.na(.data$effort_seconds), .data$effort_seconds >= 0)',environment())
+  # nolint end
   
   if (!is.null(x$locations) &&
       nrow(x$locations) > 0 &&
       "locationID" %in% names(dep) &&
       "locationID" %in% names(x$locations)) {
-    dep <- .eval('dplyr::left_join(dep, x$locations, by = "locationID", multiple = "all")',environment())
+    dep <- .eval(
+      'dplyr::left_join(dep, x$locations, by = "locationID", multiple = "all")',
+      environment()
+    )
   }
   
   divisor <- c(
@@ -72,7 +81,8 @@
 
 
 #---------------
-# a customised version of the merige_tibbles in the ctdp package (developed by: Henjo de Knegt)
+# a customised version of the merige_tibbles in the ctdp package
+# (developed by: Henjo de Knegt)
 .merge_data <- function(x, dropMedia = TRUE) {
   
   keys <- list(
@@ -232,28 +242,49 @@
   
   #---------------- select columns ----------------
   
-  omitKeys <- unlist(keys[!names(keys) %in% linkStructure$media], use.names = FALSE)
+  omitKeys <- unlist(
+    keys[!names(keys) %in% linkStructure$media],
+    use.names = FALSE
+  )
   y1 <- .drop_any_of(x$media, as.character(omitKeys))
   
-  omitKeys <- unlist(keys[!names(keys) %in% linkStructure$sequences], use.names = FALSE)
+  omitKeys <- unlist(
+    keys[!names(keys) %in% linkStructure$sequences],
+    use.names = FALSE
+  )
   y2 <- .drop_any_of(x$sequences, as.character(omitKeys))
   
-  omitKeys <- unlist(keys[!names(keys) %in% linkStructure$observations], use.names = FALSE)
+  omitKeys <- unlist(
+    keys[!names(keys) %in% linkStructure$observations],
+    use.names = FALSE
+  )
   y3 <- .drop_any_of(x$observations, as.character(omitKeys))
   
-  omitKeys <- unlist(keys[!names(keys) %in% linkStructure$deployments], use.names = FALSE)
+  omitKeys <- unlist(
+    keys[!names(keys) %in% linkStructure$deployments],
+    use.names = FALSE
+  )
   y4 <- .drop_any_of(x$deployments, as.character(omitKeys))
   
-  omitKeys <- unlist(keys[!names(keys) %in% linkStructure$locations], use.names = FALSE)
+  omitKeys <- unlist(
+    keys[!names(keys) %in% linkStructure$locations],
+    use.names = FALSE
+  )
   y5 <- .drop_any_of(x$locations, as.character(omitKeys))
   
-  omitKeys <- unlist(keys[!names(keys) %in% linkStructure$taxonomy], use.names = FALSE)
+  omitKeys <- unlist(
[…]
