### 2026-08-14 2ae6dfc
Finish supporting function line wrapping
