+++ R/supporting_functions.R
@@ -848,12 +848,19 @@
     timeshift <- pi - mean(suntimes[, 1] + suntimes[, 3]/2) * pi/12
     
     #obs$solartime <- .eval('obs |> with(activity::solartime(timestamp,latitude, longitude, 0)) |> .$solar |> + timeshift |> activity::wrap()',environment()) # nolint: line_length_linter.
+    # nolint start: line_length_linter.
     obs$solartime <- .eval('with(obs,
           activity::wrap(
             activity::solartime(timestamp, latitude, longitude, 0)$solar + timeshift
           ))',environment())
+    # nolint end
     
-    .eval('activity::fitact(obs$solartime, adj = 1.5, sample = "data", reps = reps)',environment())
+    # nolint start: line_length_linter.
+    .eval(
+      'activity::fitact(obs$solartime, adj = 1.5, sample = "data", reps = reps)',
+      environment()
+    )
+    # nolint end
   }
   else NULL
 }
@@ -1047,7 +1054,14 @@
 }
 #-------
 # adjusted from the package camtrapDensity:
-.get_parameter_table <- function (traprate_data, radius_model, angle_model, speed_model, activity_model, reps = 999) {
+.get_parameter_table <- function (
+  traprate_data,
+  radius_model,
+  angle_model,
+  speed_model,
+  activity_model,
+  reps = 999
+) {
   rad <- radius_model$edd
   ang <- angle_model$edd * 2
   spd <- dplyr::select(speed_model$estimate, dplyr::all_of(c("est", "se")))
@@ -1063,9 +1077,24 @@
   res$n <- c(nrow(radius_model$data), nrow(angle_model$data), 
              nrow(speed_model$data), length(activity_model@data), 
              NA)
-  res$unit <- c(radius_model$unit, angle_model$unit, speed_model$unit, "none", speed_model$unit)
-  rownames(res) <- c("radius", "angle", "active_speed", "activity_level", "overall_speed")
-  traprate <- .eval('camtrapDensity::get_trap_rate(traprate_data, strata=NULL, reps)',environment())
+  res$unit <- c(
+    radius_model$unit,
+    angle_model$unit,
+    speed_model$unit,
+    "none",
+    speed_model$unit
+  )
+  rownames(res) <- c(
+    "radius",
+    "angle",
+    "active_speed",
+    "activity_level",
+    "overall_speed"
+  )
+  traprate <- .eval(
+    'camtrapDensity::get_trap_rate(traprate_data, strata=NULL, reps)',
+    environment()
+  )
   j <- c("estimate", "se", "lcl95", "ucl95")
   traprate[, j] <- traprate[, j] * radius_model$proportion_used
   res <- rbind(res, traprate)
@@ -1119,21 +1148,32 @@
 .rem <- function (parameters) {
   required_rows <- c("trap_rate", "overall_speed", "radius", "angle")
   required_cols <- c("estimate", "se", "unit")
-  if (!all(required_rows %in% rownames(parameters)) | !all(required_cols %in% colnames(parameters))) 
+  if (
+    !all(required_rows %in% rownames(parameters)) |
+      !all(required_cols %in% colnames(parameters))
+  )
     stop(
       "parameters must have (at least) row names: ",
       toString(required_rows),
       " ;\nand (at least) column names: ",
       toString(required_cols)
     )
-  param <- .eval("camtrapDensity::convert_units(parameters[required_rows, ])",environment())
+  param <- .eval(
+    "camtrapDensity::convert_units(parameters[required_rows, ])",
+    environment()
+  )
   wtd_est <- param$estimate + c(0, 0, 0, 2)
   pwr_est <- wtd_est^c(1, -1, -1, -1)
[…]
