### 2026-08-14 d829780
Wrap camR helper lines for lint
