+++ R/camR_functions.R
@@ -249,7 +249,10 @@
   
   #---------------- deployment-to-location mapping ----------------
   
-  deployment_to_location <- dep_loc[, c("deploymentID", "locationID"), drop = FALSE]
+  deployment_to_location <- dep_loc[
+    , c("deploymentID", "locationID"),
+    drop = FALSE
+  ]
   deployment_to_location <- .camr_unique_rows(
     deployment_to_location,
     cols = c("deploymentID", "locationID")
@@ -263,8 +266,15 @@
     all(c("deploymentID", "captureMethod") %in% names(cm$data$sequences))
   ) {
     
-    seq_cap <- cm$data$sequences[, c("deploymentID", "captureMethod"), drop = FALSE]
-    seq_cap <- .camr_base_left_join(seq_cap, deployment_to_location, by = "deploymentID")
+    seq_cap <- cm$data$sequences[
+      , c("deploymentID", "captureMethod"),
+      drop = FALSE
+    ]
+    seq_cap <- .camr_base_left_join(
+      seq_cap,
+      deployment_to_location,
+      by = "deploymentID"
+    )
     
     capture_methods_per_location <- .camr_summary_by_location(
       seq_cap,
@@ -298,7 +308,10 @@
     all(c("sequenceID", "deploymentID") %in% names(cm$data$sequences))
   ) {
     
-    sequence_to_deployment <- cm$data$sequences[, c("sequenceID", "deploymentID"), drop = FALSE]
+    sequence_to_deployment <- cm$data$sequences[
+      , c("sequenceID", "deploymentID"),
+      drop = FALSE
+    ]
     sequence_to_deployment <- .camr_unique_rows(
       sequence_to_deployment,
       cols = c("sequenceID", "deploymentID")
@@ -495,7 +508,11 @@
   .d <- dep_loc
   
   .d <- .camr_base_left_join(.d, deployments_per_location, by = "locationID")
-  .d <- .camr_base_left_join(.d, capture_methods_per_location, by = "locationID")
+  .d <- .camr_base_left_join(
+    .d,
+    capture_methods_per_location,
+    by = "locationID"
+  )
   .d <- .camr_base_left_join(.d, setup_per_location, by = "locationID")
   .d <- .camr_base_left_join(.d, Classify_per_location, by = "locationID")
   .d <- .camr_base_left_join(.d, bait_use_per_location, by = "locationID")
@@ -510,29 +527,56 @@
 #############
 
 
-.summarize_species <- function(cm,df, class = NULL,order=NULL,domestic=FALSE,scientificName=NULL,.filterCount=TRUE,observationType=NULL) {
-  # To summarise, either scientificName, OR other criteria (one or combination of class, order, etc.) is provided!
-  # .filterCount = T -> the count threshold (if specified in cm$filterCount) is applied
-  
-  if (!is.null(observationType) && is.character(observationType)) df <- df[df$observationType %in% observationType,]
+.summarize_species <- function(
+  cm,
+  df,
+  class = NULL,
+  order = NULL,
+  domestic = FALSE,
+  scientificName = NULL,
+  .filterCount = TRUE,
+  observationType = NULL
+) {
+  # To summarise, either scientificName, OR other criteria
+  # (one or combination of class, order, etc.) is provided!
+  # .filterCount = T -> the count threshold
+  # (if specified in cm$filterCount) is applied
+
+  if (!is.null(observationType) && is.character(observationType))
+    df <- df[df$observationType %in% observationType,]
   
   if (is.null(scientificName) || length(scientificName) == 0) {
     df <- df[grepl("\\s", df$scientificName),]
     df <- df[!grepl(" sp.$", df$scientificName),]
     
     
[…]
