### 2026-08-14 23cd023
Continue camR helper line wrapping
