+++ R/camR_functions.R
@@ -813,8 +813,13 @@
     out$total_locationsrow <- 0L
     out$total_unique_locations <- 0L
     out$message_missing <- paste0(r, " No valid cm$data$locations table found.")
-    out$status_spatial <- paste0(w, " Too few locations to detect a spatial pattern")
+    out$status_spatial <- paste0(
+      w,
+      " Too few locations to detect a spatial pattern"
+    )
+    # nolint start: line_length_linter.
     out$status_MCArea <- "No valid camera-location coordinates were available, so study-area size could not be estimated."
+    # nolint end
     
     cm$data_status$Spatial <- out
     return(out)
@@ -830,18 +835,40 @@
   
   location_df2[["row"]] <- seq_len(nrow(location_df2))
   
-  location_df2[["locationID"]] <- trimws(as.character(location_df2[["locationID"]]))
-  location_df2[["locationID"]][location_df2[["locationID"]] == ""] <- NA_character_
+  location_df2[["locationID"]] <- trimws(
+    as.character(location_df2[["locationID"]])
+  )
+  location_df2[["locationID"]][
+    location_df2[["locationID"]] == ""
+  ] <- NA_character_
   
-  location_df2[["locationName"]] <- trimws(as.character(location_df2[["locationName"]]))
-  location_df2[["locationName"]][location_df2[["locationName"]] == ""] <- NA_character_
+  location_df2[["locationName"]] <- trimws(
+    as.character(location_df2[["locationName"]])
+  )
+  location_df2[["locationName"]][
+    location_df2[["locationName"]] == ""
+  ] <- NA_character_
   
   location_df2[["longitude"]] <- suppressWarnings(
-    as.numeric(gsub(",", ".", trimws(as.character(location_df2[["longitude"]])), fixed = TRUE))
+    as.numeric(
+      gsub(
+        ",",
+        ".",
+        trimws(as.character(location_df2[["longitude"]])),
+        fixed = TRUE
+      )
+    )
   )
   
   location_df2[["latitude"]] <- suppressWarnings(
-    as.numeric(gsub(",", ".", trimws(as.character(location_df2[["latitude"]])), fixed = TRUE))
+    as.numeric(
+      gsub(
+        ",",
+        ".",
+        trimws(as.character(location_df2[["latitude"]])),
+        fixed = TRUE
+      )
+    )
   )
   
   complete_idx <- stats::complete.cases(
@@ -895,10 +922,19 @@
   #---------------- 3. duplicated coordinates ----------------
   
   if (nrow(location_cleaned) > 0) {
-    location_cleaned[["lon_round"]] <- round(location_cleaned[["longitude"]], coord_round)
-    location_cleaned[["lat_round"]] <- round(location_cleaned[["latitude"]], coord_round)
+    location_cleaned[["lon_round"]] <- round(
+      location_cleaned[["longitude"]],
+      coord_round
+    )
+    location_cleaned[["lat_round"]] <- round(
+      location_cleaned[["latitude"]],
+      coord_round
+    )
     
-    ckey <- .coord_key(location_cleaned[["lon_round"]], location_cleaned[["lat_round"]])
+    ckey <- .coord_key(
+      location_cleaned[["lon_round"]],
+      location_cleaned[["lat_round"]]
+    )
     ctab <- table(ckey, useNA = "ifany")
     
     dup_coord_groups <- sum(ctab > 1)
@@ -927,8 +963,15 @@
   #---------------- 4. unique locations ----------------
   
   if (nrow(location_cleaned) > 0) {
-    xy_key <- .coord_key(location_cleaned[["longitude"]], location_cleaned[["latitude"]])
[…]
