### 2026-08-15 477b502
Continue camR helper line wrapping
