+++ R/camR_functions.R
@@ -1317,7 +1317,10 @@
       )
       
       out$MCArea <- area_km2
-      out$MCArea_method <- paste0(buffer_m / 1000, " km buffered camera-location area")
+      out$MCArea_method <- paste0(
+        buffer_m / 1000,
+        " km buffered camera-location area"
+      )
       
       out$MCArea_text <- paste0(
         .format_area(area_km2),
@@ -1326,6 +1329,7 @@
         " km buffers around camera locations)"
       )
       
+      # nolint start: line_length_linter.
       out$status_MCArea <- paste0(
         "The ",
         n_area_locations,
@@ -1337,6 +1341,7 @@
         .format_area(area_km2),
         "."
       )
+      # nolint end
       
     } else {
       
@@ -1344,6 +1349,7 @@
       out$MCArea_method <- "Minimum convex polygon"
       out$MCArea_text <- .format_area(area_km2)
       
+      # nolint start: line_length_linter.
       out$status_MCArea <- paste0(
         "The ",
         n_area_locations,
@@ -1351,6 +1357,7 @@
         out$MCArea_text,
         "."
       )
+      # nolint end
     }
   }
   
@@ -1358,7 +1365,9 @@
   
   if (!is.null(loc)) {
     country_values <- unique(terra::extract(wrld, loc)[["name"]])
-    country_values <- country_values[!is.na(country_values) & country_values != ""]
+    country_values <- country_values[
+      !is.na(country_values) & country_values != ""
+    ]
     out$country <- .paste_comma_and(country_values)
   } else {
     out$country <- NA_character_
@@ -1416,7 +1425,10 @@
   #---------------- 9. spatial pattern detection ----------------
   
   if (nrow(total_unique_locations_df) > 0) {
-    coords_xy <- total_unique_locations_df[, c("longitude", "latitude"), drop = FALSE]
+    coords_xy <- total_unique_locations_df[
+      , c("longitude", "latitude"),
+      drop = FALSE
+    ]
     coords_xy <- coords_xy[stats::complete.cases(coords_xy), , drop = FALSE]
     
     xy_key <- .coord_key(coords_xy[["longitude"]], coords_xy[["latitude"]])
@@ -1504,7 +1516,10 @@
     }
     
   } else {
-    out$status_spatial <- paste0(w, " Too few locations to detect a spatial pattern")
+    out$status_spatial <- paste0(
+      w,
+      " Too few locations to detect a spatial pattern"
+    )
   }
   
   # Keep original behavior: assignment returns out
@@ -1550,8 +1565,16 @@
     di <- .trim_chr(deployments_df$deployment_interval)
     parts <- strsplit(ifelse(is.na(di), "", di), "--", fixed = TRUE)
     
-    start_raw <- vapply(parts, function(z) if (length(z) >= 1) trimws(z[1]) else NA_character_, character(1))
-    end_raw   <- vapply(parts, function(z) if (length(z) >= 2) trimws(z[2]) else NA_character_, character(1))
+    start_raw <- vapply(
+      parts,
+      function(z) if (length(z) >= 1) trimws(z[1]) else NA_character_,
+      character(1)
+    )
+    end_raw <- vapply(
+      parts,
+      function(z) if (length(z) >= 2) trimws(z[2]) else NA_character_,
+      character(1)
+    )
     
     start_d <- .as_date(start_raw)
     end_d   <- .as_date(end_raw)
@@ -1566,7 +1589,12 @@
[…]
