### 2026-08-15 ed1a685
Continue camR temporal line wrapping
