+++ R/camR_functions.R
@@ -1854,19 +1854,46 @@
       substr(trimws(z[1]), 1, 4) else NA_character_,
     character(1)
   )
-  dep_end_year   <- vapply(parts, function(z) if (length(z) >= 2) substr(trimws(z[2]), 1, 4) else NA_character_, character(1))
+  dep_end_year <- vapply(
+    parts,
+    function(z) if (length(z) >= 2)
+      substr(trimws(z[2]), 1, 4) else NA_character_,
+    character(1)
+  )
   
-  cm$data_status$Temporal$dep_years <- sort(unique(c(dep_start_year, dep_end_year)))
-  cm$data_status$Temporal$dep_years <- cm$data_status$Temporal$dep_years[grepl("^[0-9]{4}$", cm$data_status$Temporal$dep_years)]
+  cm$data_status$Temporal$dep_years <- sort(
+    unique(c(dep_start_year, dep_end_year))
+  )
+  cm$data_status$Temporal$dep_years <-
+    cm$data_status$Temporal$dep_years[
+      grepl("^[0-9]{4}$", cm$data_status$Temporal$dep_years)
+    ]
   
-  cm$data_status$Temporal$obs_years <- sort(unique(.year4(cm$data$observations$timestamp)))
-  cm$data_status$Temporal$obs_years <- cm$data_status$Temporal$obs_years[!is.na(cm$data_status$Temporal$obs_years)]
+  cm$data_status$Temporal$obs_years <- sort(
+    unique(.year4(cm$data$observations$timestamp))
+  )
+  cm$data_status$Temporal$obs_years <-
+    cm$data_status$Temporal$obs_years[
+      !is.na(cm$data_status$Temporal$obs_years)
+    ]
   
-  cm$data_status$Temporal$dep_years_message <- .years_message(cm$data_status$Temporal$dep_years)
-  cm$data_status$Temporal$obs_years_message <- .years_message(cm$data_status$Temporal$obs_years)
+  cm$data_status$Temporal$dep_years_message <-
+    .years_message(cm$data_status$Temporal$dep_years)
+  cm$data_status$Temporal$obs_years_message <-
+    .years_message(cm$data_status$Temporal$obs_years)
   
-  cm$data_status$Temporal$years_in_dep_not_obs <- sort(setdiff(cm$data_status$Temporal$dep_years, cm$data_status$Temporal$obs_years))
-  cm$data_status$Temporal$years_in_obs_not_dep <- sort(setdiff(cm$data_status$Temporal$obs_years, cm$data_status$Temporal$dep_years))
+  cm$data_status$Temporal$years_in_dep_not_obs <- sort(
+    setdiff(
+      cm$data_status$Temporal$dep_years,
+      cm$data_status$Temporal$obs_years
+    )
+  )
+  cm$data_status$Temporal$years_in_obs_not_dep <- sort(
+    setdiff(
+      cm$data_status$Temporal$obs_years,
+      cm$data_status$Temporal$dep_years
+    )
+  )
   
   cm$data_status$Temporal$temporal_inconsistency <- {
     dep_not_obs <- cm$data_status$Temporal$years_in_dep_not_obs
@@ -1876,21 +1903,57 @@
       paste0(ic$green, " Years in observations and deployments are the same")
     } else {
       parts2 <- character(0)
-      if (length(dep_not_obs)) parts2 <- c(parts2, paste0("Deployments exist, but observations are missing for: ", toString(dep_not_obs)))
-      if (length(obs_not_dep)) parts2 <- c(parts2, paste0("Observations exist, but deployments are missing for: ", toString(obs_not_dep)))
-      paste0(ic$red, " Temporal inconsistency (", paste(parts2, collapse = " | "), ")")
+      if (length(dep_not_obs))
+        parts2 <- c(
+          parts2,
+          paste0(
[…]
