### 2026-08-15 bb3dbe9
Continue camR essentials line wrapping
