+++ R/camR_functions.R
@@ -2019,16 +2019,26 @@
     years_keep_chr = cm$data_status$Temporal$dep_years
   )
   
-  cm$data_status$Temporal$dep_month_coverage_lines <- if (nrow(cm$data_status$Temporal$dep_month_coverage) == 0) {
+  cm$data_status$Temporal$dep_month_coverage_lines <- if (
+    nrow(cm$data_status$Temporal$dep_month_coverage) == 0
+  ) {
     character(0)
   } else {
-    paste0(cm$data_status$Temporal$dep_month_coverage$Year, ": ", cm$data_status$Temporal$dep_month_coverage$MonthSpan)
+    paste0(
+      cm$data_status$Temporal$dep_month_coverage$Year,
+      ": ",
+      cm$data_status$Temporal$dep_month_coverage$MonthSpan
+    )
   }
   
   # D) Observations QA: invalid format + future timestamps
   
   ts_raw <- cm$data$observations$timestamp
-  bad_fmt_idx <- which(!is.na(ts_raw) & nzchar(.trim_chr(ts_raw)) & !.is_iso_prefix(ts_raw))
+  bad_fmt_idx <- which(
+    !is.na(ts_raw) &
+      nzchar(.trim_chr(ts_raw)) &
+      !.is_iso_prefix(ts_raw)
+  )
   
   cm$data_status$Temporal$invalid_timestamp_format <- if (length(bad_fmt_idx)) {
     paste0(length(bad_fmt_idx), " timestamp(s) have invalid format ", ic$red,
@@ -2105,8 +2115,10 @@
   .status_rows <- function(bad_idx, total, all_empty = FALSE) {
     if (is.na(total) || total <= 0) return(paste0(y, " No data"))
     if (!length(bad_idx)) return(paste0(g, " Complete"))
-    if (length(bad_idx) == total && all_empty) return(paste0(r, " Incomplete (all rows empty)"))
-    if (length(bad_idx) == total) return(paste0(r, " Incomplete (all rows missing/invalid)"))
+    if (length(bad_idx) == total && all_empty)
+      return(paste0(r, " Incomplete (all rows empty)"))
+    if (length(bad_idx) == total)
+      return(paste0(r, " Incomplete (all rows missing/invalid)"))
     paste0(y, " Partial (row: ", .row_list(bad_idx), ")")
   }
   
@@ -2127,7 +2139,11 @@
     bad_idx    <- sort(unique(c(empty_idx, nonnum_idx, range_idx)))
     
     list(
-      status = .status_rows(bad_idx, total, all_empty = (length(empty_idx) == total)),
+      status = .status_rows(
+        bad_idx,
+        total,
+        all_empty = (length(empty_idx) == total)
+      ),
       num    = x_num
     )
   }
@@ -2139,12 +2155,21 @@
     if (all(deg_like) || !any(deg_like)) return("")
     idx_deg    <- pair_idx[deg_like]
     idx_nondeg <- pair_idx[!deg_like]
-    diff_rows  <- if (length(idx_deg) < length(idx_nondeg)) idx_deg else idx_nondeg
+    diff_rows <- if (
+      length(idx_deg) < length(idx_nondeg)
+    ) idx_deg else idx_nondeg
     plural <- if (length(diff_rows) == 1) "row" else "rows"
-    paste0(" | ", plural, " ", .row_list(diff_rows), " use a different coordinate system")
+    paste0(
+      " | ",
+      plural,
+      " ",
+      .row_list(diff_rows),
+      " use a different coordinate system"
+    )
   }
   
[…]
