### 2026-08-15 ebba045
Continue camR validation line wrapping
