+++ R/camR_functions.R
@@ -2441,23 +2441,42 @@
       obs_state <- new.env(
         parent = emptyenv()
       ); obs_state$value <- cm$data_status$Essentials$obs
-      .set_animal_field <- function(key, candidates, label, type = c("chr","num")) {
+      .set_animal_field <- function(
+        key, candidates, label,
+        type = c("chr", "num")
+      ) {
         type <- match.arg(type)
         col <- .pick_col(cm$data$observations, candidates)
         if (is.na(col)) {
-          obs_state$value[[key]] <- paste0(r, " Missing column: ", paste(candidates, collapse = " / "))
+          obs_state$value[[key]] <- paste0(
+            r,
+            " Missing column: ",
+            paste(candidates, collapse = " / ")
+          )
           return()
         }
         x <- cm$data$observations[[col]][idx_animal]
         present_n <- if (type == "num") .present_num(x) else .present_chr(x)
-        obs_state$value[[key]] <- .animal_field_status(present_n, n_animal, label)
+        obs_state$value[[key]] <- .animal_field_status(
+          present_n,
+          n_animal,
+          label
+        )
       }
       
       .set_animal_field("behavior", "behavior", "behavior","chr")
       .set_animal_field("sex", "sex","sex", "chr")
       .set_animal_field("lifeStage", "lifeStage", "lifeStage", "chr")
-      .set_animal_field("angle", c("individualPositionAngle","angle"), "angle", "num")
-      .set_animal_field("radius", c("individualPositionRadius","radius"), "radius", "num")
+      .set_animal_field(
+        "angle",
+        c("individualPositionAngle", "angle"),
+        "angle", "num"
+      )
+      .set_animal_field(
+        "radius",
+        c("individualPositionRadius", "radius"),
+        "radius", "num"
+      )
       .set_animal_field("speed", c("individualSpeed","speed"), "speed", "num")
       .set_animal_field("individualID", "individualID", "individualID", "chr")
       cm$data_status$Essentials$obs <- obs_state$value
@@ -2466,15 +2485,30 @@
   #------
   # DEPLOYMENTS 
   
-  if (!("deployments" %in% names(cm$data)) || !is.data.frame(cm$data$deployments)) {
-    cm$data_status$Essentials$dep$status <- paste0(r, " Missing table: cm$data$deployments")
+  if (
+    !("deployments" %in% names(cm$data)) ||
+      !is.data.frame(cm$data$deployments)
+  ) {
+    cm$data_status$Essentials$dep$status <- paste0(
+      r,
+      " Missing table: cm$data$deployments"
+    )
   } else {
-    dep_state <- new.env(parent = emptyenv()); dep_state$value <- cm$data_status$Essentials$dep
-    add_dep <- function(key, candidates, kind = c("counts","timestamp","dep_interval"), treat_blank = TRUE) {
+    dep_state <- new.env(parent = emptyenv())
+    dep_state$value <- cm$data_status$Essentials$dep
+    add_dep <- function(
+      key, candidates,
+      kind = c("counts", "timestamp", "dep_interval"),
+      treat_blank = TRUE
+    ) {
       kind <- match.arg(kind)
       col <- .pick_col(cm$data$deployments, candidates)
       if (is.na(col)) {
-        dep_state$value[[key]] <- paste0(r, " Missing column: ", paste(candidates, collapse=" / "))
[…]
