+++ R/camR_functions.R
@@ -3073,8 +3073,10 @@
     scientificName = .trim_chr(Keep_sp$scientificName),
     family = if ("family" %in% names(Keep_sp))
       .trim_chr(Keep_sp$family) else NA_character_,
-    order = if ("order"  %in% names(Keep_sp)) .trim_chr(Keep_sp$order)  else NA_character_,
-    class = if ("class"  %in% names(Keep_sp)) .trim_chr(Keep_sp$class)  else NA_character_,
+    order = if ("order" %in% names(Keep_sp))
+      .trim_chr(Keep_sp$order) else NA_character_,
+    class = if ("class" %in% names(Keep_sp))
+      .trim_chr(Keep_sp$class) else NA_character_,
     obs_records_count = 0,
     n_sequences  = 0L,
     stringsAsFactors = FALSE
@@ -3276,8 +3278,10 @@
   
   if (length(txt) > 0 && any(grepl("\\beow\\b", txt, ignore.case = TRUE))) {
     cm$reportTextElements$name <- "EOW"
+    # nolint start: line_length_linter.
     cm$reportTextElements$message <- "This survey is part of the [European Observatory of Wildlife](https://wildlifeobservatory.org/), an international project in which institutions monitor protected areas across European countries."
     cm$reportTextElements$Intro_text <- "The [European Observatory of Wildlife (EOW)](https://wildlifeobservatory.org/) is a standardized camera-trapping network operating across more than 100 study areas in Europe. It is coordinated by the [ENETWILD](https://enetwild.com/) consortium and funded by the [European Food Safety Authority (EFSA)](https://www.efsa.europa.eu/en). Images collected by camera traps within this network are processed and archived in Agouti, exported in the [Camtrap DP](https://camtrap-dp.tdwg.org/) standard format, and included in annual monitoring reports submitted to EFSA. A key feature of the EOW protocol is the use of the [Random Encounter Model (REM)](https://github.com/MarcusRowcliffe/camtrapDensity) to estimate population density from camera-trap detections (*Rowcliffe et al., 2014*). As a result, the entire workflow follows an established and standardized framework."
+    # nolint end
     cm$info[["is.EOW"]] <- TRUE
   } else {
     cm$reportTextElements$name <- "Non-EOW"
@@ -3313,7 +3317,9 @@
       )
     } else {
       cm$reportTextElements$habitat_text <- paste0(
-        "The area is a mosaic of ", .paste_comma_and(hab_vals), " habitat types."
+        "The area is a mosaic of ",
+        .paste_comma_and(hab_vals),
+        " habitat types."
       )
     }
   }
@@ -3337,14 +3343,19 @@
       drop = FALSE
     ]
     
-    sp_table <- sp_table[order(sp_table$captures, decreasing = TRUE), , drop = FALSE]
+    sp_table <- sp_table[
+      order(sp_table$captures, decreasing = TRUE),
+      ,
+      drop = FALSE
+    ]
     
     top5_names <- head(sp_table$scientificName, 5)
     
     cm$data_status$Species$most_observed_sp <- top5_names
     
     top5_names_italic <- paste0("*", top5_names, "*")
-    cm$reportTextElements$most_observed_sp_text <- .paste_comma_and(top5_names_italic)
+    cm$reportTextElements$most_observed_sp_text <-
+      .paste_comma_and(top5_names_italic)
     
   } else {
     cm$data_status$Species$most_observed_sp <- character(0)
@@ -3352,7 +3363,9 @@
   }
   #-------- Image processing source
   if (!is.null(cm$info$json$sources) &&
[…]
