### 2026-08-15 1e22f1c
Continue camR line wrapping
