+++ R/camR.R
@@ -65,7 +65,9 @@
     paste0(
       "```{r setup, include=FALSE}\n",
       "pkgs <- c(", pkg_txt, ")\n",
+      # nolint start: line_length_linter.
       "missing_pkgs <- pkgs[!vapply(pkgs, requireNamespace, logical(1), quietly = TRUE)]\n",
+      # nolint end
       "if (length(missing_pkgs) > 0) {\n",
       "  stop('Missing package(s): ', paste(missing_pkgs, collapse = ', '))\n",
       "}\n",
@@ -78,7 +80,9 @@
     paste0(
       "```{r setup, include=FALSE}\n",
       "pkgs <- c(", pkg_txt, ")\n",
+      # nolint start: line_length_linter.
       "missing_pkgs <- pkgs[!vapply(pkgs, requireNamespace, logical(1), quietly = TRUE)]\n",
+      # nolint end
       "if (length(missing_pkgs) > 0) {\n",
       "  stop('Missing package(s): ', paste(missing_pkgs, collapse = ', '))\n",
       "}\n",
@@ -185,9 +189,12 @@
 
 #---------
 
-# in group_definition, the groups like large_mammals, wild_mammals can be defined
-# each group can be specified based on data columns (e.g., order, class, scientificNames)
-# domestic group can be defined based on which the wild_animals also defined (species NOT in domestic)
+# in group_definition, the groups like large_mammals, wild_mammals
+# can be defined
+# each group can be specified based on data columns
+# (e.g., order, class, scientificNames)
+# domestic group can be defined based on which the wild_animals also
+# defined (species NOT in domestic)
 
 camR <- setRefClass(
   "camReport",
@@ -229,11 +236,14 @@ camR <- setRefClass(
     rem = "list",
     .rem_params  = "list",
     .act_models = "list",
-    .any_data_for_rem  = 'logical', # a vector of species names for which REM could not be fitted!
+    # a vector of species names for which REM could not be fitted!
+    .any_data_for_rem  = 'logical',
     .tempObjects  = "list",
     reportTextElements = "list",
     reportObjectElements = "list",
-    statusReportObjects = "list", # a list of .txtSection and .Rchunk objects (for data_status report!)
+    # a list of .txtSection and .Rchunk objects
+    # (for data_status report!)
+    statusReportObjects = "list",
     reportObjects    = "list" # a list of .txtSection and .Rchunk objects
   ),
   
@@ -241,16 +251,23 @@ camR <- setRefClass(
     initialize = function() {
       #.loadlib()
       
-      .self$setting <- list(locationLegend = FALSE,color=c("#CA6A28","#6C9100","#00A383","#008ADF","#D44CBF"))
+      .self$setting <- list(
+        locationLegend = FALSE,
+        color = c("#CA6A28", "#6C9100", "#00A383", "#008ADF", "#D44CBF")
+      )
       
       
       .self$filterDuration <- 5
       
       .self$siteName <-  "**an unspecified location**"
       
+      # nolint start: line_length_linter.
       .self$description <- "No discription has been provided for this site. The report object can be updated by overriding the 'description' field!"
+      # nolint end
       
+      # nolint start: line_length_linter.
       #.self$sampling <- "Description of sampling method has NOT been provided. The 'sampling' field can be updated by the user. More details on camera trap deployments can be found in the summarized information table below (Table 1)!"
+      # nolint end
       .self$title <- 'Report generated using the camtrapReport package'
       .self$authors <- "Monitoring team"
       .self$logoPath <- ''
@@ -263,9 +280,12 @@ camR <- setRefClass(
[…]
