### 2026-08-15 d4bfb59
Continue camR line wrapping
