+++ R/camR.R
@@ -1207,48 +1207,166 @@ camR <- setRefClass(
           ) {
             if ('class' %in% names(.self$group_definition[[n]])) {
               if ('order' %in% names(.self$group_definition[[n]])) {
-                if ('domestic' %in% names(.self$group_definition[[n]]) && is.logical(.self$group_definition[[n]]$domestic)) {
-                  if ("observationType" %in% names(.self$group_definition[[n]])) {
-                    .self$species_summary[[n]] <- .summarize_species(.self,.tax_obs,class = .self$group_definition[[n]]$class,order = .self$group_definition[[n]]$order,domestic = .self$group_definition[[n]]$domestic,observationType=.self$group_definition[[n]]$observationType)
+                if (
+                  'domestic' %in%
+                    names(.self$group_definition[[n]]) &&
+                  is.logical(
+                    .self$group_definition[[n]]$domestic
+                  )
+                ) {
+                  if (
+                    "observationType" %in%
+                      names(.self$group_definition[[n]])
+                  ) {
+                    .self$species_summary[[n]] <-
+                      .summarize_species(
+                        .self,
+                        .tax_obs,
+                        class = .self$group_definition[[n]]$class,
+                        order = .self$group_definition[[n]]$order,
+                        domestic =
+                          .self$group_definition[[n]]$domestic,
+                        observationType =
+                          .self$group_definition[[n]]$observationType
+                      )
                   } else {
-                    .self$species_summary[[n]] <- .summarize_species(.self,.tax_obs,class = .self$group_definition[[n]]$class,order = .self$group_definition[[n]]$order,domestic = .self$group_definition[[n]]$domestic)
+                    .self$species_summary[[n]] <-
+                      .summarize_species(
+                        .self,
+                        .tax_obs,
+                        class = .self$group_definition[[n]]$class,
+                        order = .self$group_definition[[n]]$order,
+                        domestic =
+                          .self$group_definition[[n]]$domestic
+                      )
                   }
                   
                 } else {
-                  if ("observationType" %in% names(.self$group_definition[[n]])) {
-                    .self$species_summary[[n]] <- .summarize_species(.self,.tax_obs,class = .self$group_definition[[n]]$class,order = .self$group_definition[[n]]$order,observationType=.self$group_definition[[n]]$observationType)
+                  if (
+                    "observationType" %in%
+                      names(.self$group_definition[[n]])
+                  ) {
+                    .self$species_summary[[n]] <-
+                      .summarize_species(
+                        .self,
+                        .tax_obs,
+                        class = .self$group_definition[[n]]$class,
+                        order = .self$group_definition[[n]]$order,
+                        observationType =
+                          .self$group_definition[[n]]$observationType
+                      )
                   } else {
-                    .self$species_summary[[n]] <- .summarize_species(.self,.tax_obs,class = .self$group_definition[[n]]$class,order = .self$group_definition[[n]]$order)
+                    .self$species_summary[[n]] <-
+                      .summarize_species(
+                        .self,
+                        .tax_obs,
+                        class = .self$group_definition[[n]]$class,
+                        order = .self$group_definition[[n]]$order
+                      )
                   }
                 }
               } else {
-                if ('domestic' %in% names(.self$group_definition[[n]]) && is.logical(.self$group_definition[[n]]$domestic)) {
-                  if ("observationType" %in% names(.self$group_definition[[n]])) {
[…]
