### 2026-08-15 757a9c1
Continue camR line wrapping
