+++ R/camR.R
@@ -2313,15 +2313,25 @@ camR <- setRefClass(
                               break
                             }
                           } else {
+                            # nolint start: line_length_linter.
                             if (.self$reportObjects[[.n]][[.nn]][[x@parent]]@Rchunk@name == x@name) {
+                            # nolint end
+                              # nolint start: line_length_linter.
                               .self$reportObjects[[.n]][[.nn]][[x@parent]]@Rchunk <- x
+                              # nolint end
                               .added <- TRUE
                               break
                             } else {
                               .tmp <- list()
+                              # nolint start: line_length_linter.
                               .tmp[[.self$reportObjects[[.n]][[.nn]][[x@parent]]@Rchunk@name]] <- .self$reportObjects[[.n]][[.nn]][[x@parent]]@Rchunk
+                              # nolint end
+                              # nolint start: line_length_linter.
                               .self$reportObjects[[.n]][[.nn]][[x@parent]]@Rchunk <- .tmp
+                              # nolint end
+                              # nolint start: line_length_linter.
                               .self$reportObjects[[.n]][[.nn]][[x@parent]]@Rchunk[[x@name]] <- x
+                              # nolint end
                               .added <- TRUE
                               break
                             }
@@ -2341,9 +2351,13 @@ camR <- setRefClass(
                 } else {
                   if (is.list(.self$reportObjects[[.n]]@Rchunk)) {
                     if (length(.self$reportObjects[[.n]]@Rchunk) > 0) {
+                      # nolint start: line_length_linter.
                       .nn <- simplify2array(lapply(.self$reportObjects[[.n]]@Rchunk, function(x) x@name))
+                      # nolint end
                       if (x@name %in% .nn) {
+                        # nolint start: line_length_linter.
                         .self$reportObjects[[.n]]@Rchunk[[.nn[.nn == x@name]]] <- x
+                        # nolint end
                         .added <- TRUE
                         break
                       } else {
@@ -2363,7 +2377,9 @@ camR <- setRefClass(
                       break
                     } else {
                       .tmp <- list()
+                      # nolint start: line_length_linter.
                       .tmp[[.self$reportObjects[[.n]]@Rchunk@name]] <- .self$reportObjects[[.n]]@Rchunk
+                      # nolint end
                       .self$reportObjects[[.n]]@Rchunk <- .tmp
                       .self$reportObjects[[.n]]@Rchunk[[x@name]] <- x
                       .added <- TRUE
@@ -2376,7 +2392,9 @@ camR <- setRefClass(
           }
         }
         #----
+        # nolint start: line_length_linter.
         # if (!.added) message('\nthe parent .textSection is not found... The R chunk is NOT added!!')
+        # nolint end
         # else message('\n...Done!')
       }
       
@@ -2387,7 +2405,9 @@ camR <- setRefClass(
       if (inherits(x,'.textSection')) {
         if (is.null(x@parent) || x@parent == "" || x@parent == ".root") {
           x@headLevel <- 1
+          # nolint start: line_length_linter.
           if (is.list(.self$statusReportObjects[[x@name]])) .self$statusReportObjects[[x@name]][[x@name]] <- x
+          # nolint end
           else .self$statusReportObjects[[x@name]] <- x
           .done <- TRUE
         } else {
@@ -2398,7 +2418,9 @@ camR <- setRefClass(
               .done <- TRUE
             } else {
               .tmp <- list()
+              # nolint start: line_length_linter.
               .tmp[[.self$statusReportObjects[[x@parent]]@name]] <- .self$statusReportObjects[[x@parent]]
+              # nolint end
               .self$statusReportObjects[[x@parent]] <- .tmp
[…]
