+++ R/sections.R
@@ -89,19 +89,15 @@ setMethod("section_names",signature(keep = "ANY"),
       if (!all(w)) {
         if (!any(w)) {
           stop(
-            paste0(
-              "None of the specified section/module names in 'keep' ",
-              "are available; use section_names() to get a list ",
-              "of existing modules."
-            )
+            "None of the specified section/module names in 'keep' ",
+            "are available; use section_names() to get a list ",
+            "of existing modules."
           )
         }
         
         warning(
-          paste0(
-            "Several section/module names specified in 'keep' ",
-            "are not available: "
-          ),
+          "Several section/module names specified in 'keep' ",
+          "are not available: ",
           .paste_comma_and(keep[!w])
         )
       }
@@ -124,19 +120,15 @@ setMethod("section_names",signature(keep = "ANY"),
       if (!all(w)) {
         if (!any(w)) {
           stop(
-            paste0(
-              "None of the specified section/module names in 'exclude' ",
-              "are available; use section_names() to get a list ",
-              "of existing modules."
-            )
+            "None of the specified section/module names in 'exclude' ",
+            "are available; use section_names() to get a list ",
+            "of existing modules."
           )
         }
         
         warning(
-          paste0(
-            "Several section/module names specified in 'exclude' ",
-            "are not available: "
-          ),
+          "Several section/module names specified in 'exclude' ",
+          "are not available: ",
           .paste_comma_and(exclude[!w])
         )
       }
@@ -194,30 +186,24 @@ setMethod("sections",signature(x = "camReport"),
       if (all(n %in% x$reportObjectElements$Modules_info$name)) {
         
         message(
-          paste0(
-            "\nSome of the specified sections are excluded because ",
-            "their test results were problematic."
-          )
+          "\nSome of the specified sections are excluded because ",
+          "their test results were problematic."
         )
         
       } else {
         
         if (!any(n %in% nn)) {
           stop(
-            paste0(
-              "None of the specified section names are known. ",
-              "Use section_names() to get the correct names ",
-              "of available sections."
-            )
+            "None of the specified section names are known. ",
+            "Use section_names() to get the correct names ",
+            "of available sections."
           )
         }
 
         message(
-          paste0(
-            "\nSome of the specified section names are unknown ",
-            "and ignored. Use section_names() to get the correct ",
-            "names of available sections."
-          )
+          "\nSome of the specified section names are unknown ",
+          "and ignored. Use section_names() to get the correct ",
+          "names of available sections."
         )
       }
     }
