+++ R/info.R
@@ -99,10 +99,8 @@ setMethod(
       
       if (length(name) == 0) {
         warning(
-          paste0(
-            "The specified name(s) are not identified or available ",
-            "in the camReport object; the default fields are used."
-          )
+          "The specified name(s) are not identified or available ",
+          "in the camReport object; the default fields are used."
         )
         name <- .default_cam_info_names
       }
@@ -183,10 +181,8 @@ setReplaceMethod(
       
       if (length(name) == 0) {
         stop(
-          paste0(
-            "The specified name is not identified or available ",
-            "in the camReport object."
-          )
+          "The specified name is not identified or available ",
+          "in the camReport object."
         )
       }
       
