### 2026-08-15 05de2ec
Suppress nested pipe style lints
