### 2026-08-15 3889a4f
Add spatial and report coverage tests
