+++ tests/testthat/test-spatial-summary-coverage.R
@@ -1,0 +1,277 @@
+run_spatial_summary_coverage <- function(
+  locations,
+    unavailable = "lutz"
+) {
+  cm <- camtrap_test_report()$copy(shallow = FALSE)
+
+  cm$data$locations <- locations
+
+  if (is.null(cm$data$settings)) {
+    cm$data$settings <- list()
+  }
+
+  cm$data$settings$tz <- "Europe/Amsterdam"
+
+  original_require <- ct_internal(".require")
+
+  testthat::with_mocked_bindings(
+    ct_internal(".summarize_spatial")(cm),
+    .require = function(x) {
+      package <- as.character(x)[1]
+
+      if (
+        !is.na(package) &&
+          package %in% unavailable
+      ) {
+        return(FALSE)
+      }
+
+      original_require(x)
+    },
+    .package = "camtrapReport"
+  )
+}
+
+
+test_that("spatial summary handles a missing locations table", {
+  out <- run_spatial_summary_coverage(NULL)
+
+  expect_identical(out$total_locationsrow, 0L)
+  expect_identical(out$total_unique_locations, 0L)
+
+  expect_match(
+    out$message_missing,
+    "No valid cm$data$locations table found",
+    fixed = TRUE
+  )
+
+  expect_match(
+    out$status_MCArea,
+    "No valid camera-location coordinates",
+    fixed = TRUE
+  )
+
+  expect_match(
+    out$status_spatial,
+    "Too few locations",
+    fixed = TRUE
+  )
+})
+
+
+test_that("spatial summary handles an empty locations table", {
+  locations <- data.frame(
+    locationID = character(),
+    locationName = character(),
+    longitude = numeric(),
+    latitude = numeric(),
+    stringsAsFactors = FALSE
+  )
+
+  out <- run_spatial_summary_coverage(locations)
+
+  expect_identical(out$total_locationsrow, 0L)
+  expect_identical(out$total_unique_locations, 0L)
+  expect_identical(out$number_missing_rows, 0L)
+  expect_identical(out$num_duplicated_coordinate, 0L)
+
+  expect_true(is.na(out$MCArea))
+  expect_identical(out$MCArea_method, "Not estimated")
+
+  expect_match(
+    out$status_MCArea,
+    "No valid camera-location coordinates",
+    fixed = TRUE
+  )
+
+  expect_match(
+    out$outliers_status,
+    "No valid locations available",
+    fixed = TRUE
+  )
+})
+
+
+test_that("spatial summary handles one camera location", {
+  locations <- data.frame(
+    locationID = "loc-1",
+    locationName = "Utrecht",
+    longitude = 5.1214,
+    latitude = 52.0907,
+    stringsAsFactors = FALSE
+  )
+
+  out <- run_spatial_summary_coverage(locations)
+
+  expect_identical(out$total_unique_locations, 1L)
+
+  expect_true(is.finite(out$MCArea))
+  expect_gt(out$MCArea, 0)
+
+  expect_identical(
+    out$MCArea_method,
+    "1 km buffer around one camera location"
+  )
+
+  expect_true(is.na(out$mean_distance_cam))
+  expect_true(is.na(out$min_distance_cam))
+  expect_true(is.na(out$max_distance_cam))
+
+  expect_match(
+    out$status_MCArea,
+    "one distinct camera location",
+    fixed = TRUE
+  )
+
+  expect_match(
+    out$status_spatial,
+    "Too few locations",
+    fixed = TRUE
+  )
+})
+
+
+test_that("spatial summary handles two camera locations", {
+  locations <- data.frame(
+    locationID = c("loc-1", "loc-2"),
[…]
