### 2026-08-15 85daf73
Add setup coverage tests
