+++ tests/testthat/test-setup-coverage.R
@@ -1,0 +1,343 @@
+run_setup_coverage <- function(cm, tz = NULL) {
+  original_require <- ct_internal(".require")
+
+  fake_summary <- function(...) {
+    list(
+      site_list = data.frame(
+        scientificName = character()
+      )
+    )
+  }
+
+  suppressWarnings(
+    testthat::with_mocked_bindings(
+      cm$setup(tz = tz),
+      .summarize_species = fake_summary,
+      .require = function(x) {
+        package <- as.character(x)[1]
+
+        if (
+          !is.na(package) &&
+            identical(package, "suncalc")
+        ) {
+          return(FALSE)
+        }
+
+        original_require(x)
+      },
+      .package = "camtrapReport"
+    )
+  )
+
+  invisible(cm)
+}
+
+
+test_that(
+  "setup handles supported class and order group definitions",
+  {
+    cm <- camtrap_test_report()$copy(
+      shallow = FALSE
+    )
+
+    data <- cm$data
+
+    data$taxonomy$class <- rep(
+      "TestClass",
+      nrow(data$taxonomy)
+    )
+
+    data$taxonomy$order <- rep(
+      "TestOrder",
+      nrow(data$taxonomy)
+    )
+
+    cm$data <- data
+
+    custom_groups <- list(
+      class_order_dom_obs = list(
+        class = "TestClass",
+        order = "TestOrder",
+        domestic = FALSE,
+        observationType = "animal"
+      ),
+      class_order_dom = list(
+        class = "TestClass",
+        order = "TestOrder",
+        domestic = FALSE
+      ),
+      class_order_obs = list(
+        class = "TestClass",
+        order = "TestOrder",
+        observationType = "animal"
+      ),
+      class_order_plain = list(
+        class = "TestClass",
+        order = "TestOrder"
+      ),
+      class_dom_obs = list(
+        class = "TestClass",
+        domestic = FALSE,
+        observationType = "animal"
+      ),
+      class_dom = list(
+        class = "TestClass",
+        domestic = FALSE
+      ),
+      class_obs = list(
+        class = "TestClass",
+        observationType = "animal"
+      ),
+      class_plain = list(
+        class = "TestClass"
+      ),
+      order_dom_obs = list(
+        order = "TestOrder",
+        domestic = FALSE,
+        observationType = "animal"
+      ),
+      order_dom = list(
+        order = "TestOrder",
+        domestic = FALSE
+      ),
+      order_obs = list(
+        order = "TestOrder",
+        observationType = "animal"
+      ),
+      order_plain = list(
+        order = "TestOrder"
+      )
+    )
+
+    cm$group_definition <- custom_groups
+
+    setting <- cm$setting
+    setting$focus_groups <- NULL
+    cm$setting <- setting
+
+    cm$filterDuration <- 0
+
+    run_setup_coverage(
+      cm,
+      tz = "UTC"
+    )
+
+    expect_identical(
+      cm$setting$tz,
+      "UTC"
+    )
+
+    expected <- names(custom_groups)
+
+    expect_true(
+      all(
+        expected %in%
+          cm$species_summary$count$Group
+      )
+    )
+
+    counts <- cm$species_summary$count$Count[
+      match(
+        expected,
+        cm$species_summary$count$Group
+      )
+    ]
+
+    expect_true(
+      all(counts == 0)
+    )
+  }
+)
+
+
+test_that(
+  "setup creates standard taxonomic groups when taxa support them",
+  {
+    cm <- camtrap_test_report()$copy(
[…]
