### 2026-08-15 936f2db
Add data status coverage tests
