+++ tests/testthat/test-status-coverage.R
@@ -1,0 +1,921 @@
+run_temporal_coverage <- function(deployments, observations) {
+  cm <- camR$new()
+
+  cm$data <- list(
+    deployments = deployments,
+    observations = observations
+  )
+
+  cm$data_status <- list()
+
+  ct_internal(".Temporal")(cm)
+
+  cm$data_status$Temporal
+}
+
+
+run_essentials_coverage <- function(data) {
+  cm <- camR$new()
+
+  cm$data <- data
+  cm$data_status <- list()
+
+  ct_internal(".Essentials")(cm)
+
+  cm$data_status$Essentials
+}
+
+
+test_that(
+  "Temporal handles a single deployment without gaps",
+  {
+    deployments <- data.frame(
+      deploymentID = "d1",
+      deployment_interval = "2020-01-01--2020-01-03",
+      deploymentStart = "2020-01-01 00:00:00",
+      deploymentEnd = "2020-01-03 00:00:00",
+      stringsAsFactors = FALSE
+    )
+
+    observations <- data.frame(
+      timestamp = "2020-01-02 12:00:00",
+      stringsAsFactors = FALSE
+    )
+
+    out <- run_temporal_coverage(
+      deployments,
+      observations
+    )
+
+    expect_identical(
+      out$dep_years,
+      "2020"
+    )
+
+    expect_match(
+      out$dep_calendar_coverage,
+      "3 of 3 days"
+    )
+
+    expect_match(
+      out$dep_max_gap,
+      "None"
+    )
+
+    expect_match(
+      out$dep_min_gap,
+      "None"
+    )
+
+    expect_gt(
+      length(out$dep_month_coverage_lines),
+      0
+    )
+
+    expect_match(
+      out$message_first_last,
+      "on/after"
+    )
+  }
+)
+
+
+test_that(
+  "Temporal identifies reversed intervals, gaps, and temporal problems",
+  {
+    deployments <- data.frame(
+      deploymentID = c(
+        "d1",
+        "d2",
+        "d3"
+      ),
+      deployment_interval = c(
+        "2020-01-03--2020-01-01",
+        "2020-01-10--2020-01-10",
+        "2040-02-01--2040-03-01"
+      ),
+      deploymentStart = c(
+        "2020-01-01 00:00:00",
+        "2020-01-10 00:00:00",
+        "2040-02-01 00:00:00"
+      ),
+      deploymentEnd = c(
+        "2020-01-03 00:00:00",
+        "2020-01-10 00:00:00",
+        "2040-03-01 00:00:00"
+      ),
+      stringsAsFactors = FALSE
+    )
+
+    observations <- data.frame(
+      timestamp = c(
+        "2019-12-31 12:00:00",
+        "2020-01-02 12:00:00",
+        "2020-01-05",
+        rep(
+          "2100-01-01 12:00:00",
+          11
+        )
+      ),
+      stringsAsFactors = FALSE
+    )
+
+    out <- run_temporal_coverage(
+      deployments,
+      observations
+    )
+
+    expect_match(
+      out$dep_missing_intervals,
+      "1 invalid"
+    )
+
+    expect_match(
+      out$dep_zero_length,
+      "1 zero-length"
+    )
+
+    expect_match(
+      out$dep_max_gap,
+      "days"
+    )
+
+    expect_false(
+      grepl(
+        "Same as max gap",
+        out$dep_min_gap,
+        fixed = TRUE
+      )
+    )
+
+    expect_match(
+      out$dep_years_message,
+      "missing:"
+    )
+
+    expect_match(
+      out$temporal_outliers,
+      "Years:"
+    )
+
+    expect_match(
+      out$temporal_inconsistency,
+      "Deployments exist"
+    )
+
[…]
