### 2026-08-15 4fc23db
Add safe summary coverage tests
