+++ tests/testthat/test-safe-summary-coverage.R
@@ -1,0 +1,661 @@
+make_sampling_cm <- function(
+  sampling_design = character(),
+  camera_model = character(),
+  bait = character(),
+  height = numeric(),
+  capture_method = character(),
+  individual_animals = character(),
+  is_eow = FALSE
+) {
+  cm <- camR$new()
+
+  cm$data <- list(
+    deployments = data.frame(
+      cameraModel = camera_model,
+      baitUse = bait,
+      cameraHeight = height,
+      stringsAsFactors = FALSE
+    )
+  )
+
+  cm$info <- list(
+    is.EOW = is_eow,
+    json = list(
+      project = list(
+        samplingDesign = sampling_design,
+        captureMethod = capture_method,
+        individualAnimals = individual_animals
+      )
+    )
+  )
+
+  cm$reportTextElements <- list()
+
+  cm
+}
+
+
+test_that(
+  "sampling text describes a single standard design",
+  {
+    cm <- make_sampling_cm(
+      sampling_design = "simpleRandom",
+      camera_model = "Model A",
+      bait = "FALSE",
+      height = 0.5,
+      capture_method = "activityDetection",
+      individual_animals = "FALSE",
+      is_eow = TRUE
+    )
+
+    ct_internal(".get_sampling_text")(cm)
+
+    text <- cm$reportTextElements$sampling
+
+    expect_match(
+      text,
+      "EOW camera-trap protocol",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "simple random",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "Model A",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "No bait was used",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "0.5 m",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "activity detection",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "broader wildlife monitoring",
+      fixed = TRUE
+    )
+  }
+)
+
+
+test_that(
+  "sampling text handles multiple survey settings",
+  {
+    cm <- make_sampling_cm(
+      sampling_design = c(
+        "simpleRandom",
+        "targeted",
+        "opportunistic"
+      ),
+      camera_model = c(
+        "A",
+        "B",
+        "C",
+        "D"
+      ),
+      bait = c(
+        "TRUE",
+        "FALSE"
+      ),
+      height = c(
+        0.2,
+        0.5,
+        1,
+        1.5
+      ),
+      capture_method = c(
+        "motionDetection",
+        "timeLapse",
+        "audio"
+      ),
+      individual_animals = c(
+        "TRUE",
+        "FALSE"
+      )
+    )
+
+    ct_internal(".get_sampling_text")(cm)
+
+    text <- cm$reportTextElements$sampling
+
+    expect_match(
+      text,
+      "combines the following approaches",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "Multiple camera models",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "mixture of baited and unbaited",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "ranging from 0.2 to 1.5",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "motion detection",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "time-lapse",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "audio recording",
+      fixed = TRUE
+    )
+
+    expect_match(
+      text,
+      "both the identification of individual animals",
[…]
