### 2026-08-16 d6fa03a
Correct software citation metadata
