### 2026-08-17 #799 @maelle
- [x ] **License:** The package has a CRAN or OSI accepted license.
---

Thank you for your submission!

I saw an issue related to [MEE editor comments](https://github.com/spatialecology/camtrapReport/issues/2), have you already submitted the package/paper to MEE?
The rOpenSci software peer-review should take place before the MEE review.

My comments are below, please don't hesitate to ask any question or add any information in a new comment!

##### Use

Has the package been used in recent ecological studies?
Or has it been developed with the hope that it will be used?
If it has already been used "in the wild", the README could mention that.

The README examples should be executed.
For that, the paths should be paths to example data, not fake paths.
The README and Get started links should feature a fully worked example, a simple one, but that one could follow on one's machine to try out the package.
I would like to have a new look as editor at the package once that's been added, because right now it'd be difficult for me to get started.

##### Automatic checks

I agree with Adam that goodpractice's items should be handled or discussed.
Furthermore, running [jarl](https://jarl.etiennebacher.com/) on the package I was in particular worried about

```r
~/Documents/ropensci/SOFTWARE-REVIEW/camtrapReport$ jarl check . -s duplicated_function_definition
warning: duplicated_function_definition
  --> R/utils.R:63:1
   |
63 | .require <- function(x) {
   | -------- `.require` is defined more than once in this package.
   |
   = help: Other definition at R/installAll.R:35:1

warning: duplicated_function_definition
   --> R/utils.R:297:1
    |
297 | .eval <- function(x, env) {
    | ----- `.eval` is defined more than once in this package.
    |
    = help: Other definition at R/installAll.R:7:1

warning: duplicated_function_definition
   --> R/utils.R:385:1
    |
385 | .make_render_env <- function(object) {
    | ---------------- `.make_render_env` is defined more than once in this package.
    |
    = help: Other definition at R/camR_functions.R:668:1

warning: duplicated_function_definition
   --> R/utils.R:875:1
    |
875 | .is.projected <- function(x) {
    | ------------- `.is.projected` is defined more than once in this package.
    |
    = help: Other definition at R/utils.R:797:1

── Summary ──────────────────────────────────────
Found 4 errors.
```

Other jarl results might also be worth looking at:

```r
jarl check . --statistics
   10 [ ] implicit_assignment
    4 [ ] duplicated_function_definition
    4 [ ] vector_logic
    1 [*] string_boundary
    1 [ ] true_false_symbol
```

##### Dependencies

Why does the package depend on data.table as well as dplyr and tidyr?
I'd suggest picking either data.table or dplyr if possible.

The package has many suggested dependencies.
For what kind of functionality are they needed? The interactive report?

Installation instructions should use [pak](https://pak.r-lib.org/) rather than remotes.
The custom `install_All()` function should be removed, unless I missed something it does that isn't achievable through calls to pak functions.
I think that would remove some of the duplicated functions e.g. `.require` wouldn't be needed at all.
Out of curiosity, was `install_All()` a suggestion by a coding assistant?
I was also surprised by how it was defined (using `setGeneric`), it seemed overcomplicated (but again, I might be missing something).

##### Design

The README and the "Package overview" vignettes are aimed at potential users, not developers, of the package.
Therefore I'd recommend

- Not mentioning Reference Classes in the README.
- Not mentioning implementation details like the "Container class" in the diagram.

Why use Reference Classes by the way?
This information should feature in the contributing guide as it's something relevant and important for potential contributors.

###### Interface

Why use camelCase for function names?
We strongly recommend using snake_case for alignment with many R packages.

##### Comparison with other packages

The comparison with other packages should be condensed.
I do not think the column called "most existing tools" makes much sense, that title is too vague.

##### Reproducibility

In https://spatialecology.github.io/camtrapReport/articles/Package-Overview.html#reproducibility-interpretation-and-scope rather than mentioning `sessionInfo()`, the docs could link to external guidance on reproducibility.
It is not the job of the package docs to teach reproducibility as a whole. :grin:

##### Documentation

In the Package Overview, "What camtrapReport is designed to do" repeats parts of what's been said above.
In general, I'd recommend trying to condense the documentation to make it easier to read.

This section of the reference index is not needed: https://spatialecology.github.io/camtrapReport/reference/index.html#package-overview
The corresponding manual page should use `@keywords internal` to be excluded from the pkgdown reference index.

https://spatialecology.github.io/camtrapReport/reference/index.html#core-classes should come at the very end of the reference index: do users really need to learn about core classes first? That'd be a bit intimidating. :smile_cat:

https://spatialecology.github.io/camtrapReport/articles/resources.html is a nice page to have but the workshop links are not clickable?

Because the index.md is different from the README.md, the pkgdown landing page lacks important information such as the author list, the link to the GitHub repository, etc.
That might be suprising for potential users used to the common pkgdown langing pages.

##### Code

I do not understand these lines [R/camData.R#L60-L74](https://github.com/maelle/camtrapReport/blob/7d6ea4131a9fcca4e74041ee0854cf4f33c4c8df/R/camData.R#L60-L74):

```r
suppressWarnings(
      .eval('lubridate::parse_date_time(
        x_chr[missing_i],
        orders = c(
          "ymd HMS z", "ymd HMS",
          "ymd HM z",  "ymd HM",
          "ymd z",     "ymd",
          "Ymd HMS z", "Ymd HMS",
          "ymdT HMS z", "ymdT HMS",
          "ymdT HM z",  "ymdT HM"
        ),
        tz = tz,
        quiet = TRUE
      )',environment())
    )
```

Why is a function called `.eval()` needed at all in the package?

The `R/utils.R` file is very long.
If it's still very long after removing duplicate function definitions, it might be worth splitting it by topic e.g. `R/utils-spatial.R`.

What is `.make_render_env()`?
It should be documented for potential contributors (and for reviewers :grin:).

For all uses of `tempdir()` in package code and tests, please make sure the temporary directory is removed afterwards.
The withr package might be useful.
Why is the cache location a temporary directory as opposed to an user directory? https://blog.r-hub.io/2021/07/30/cache/#storing-on-disk (or as opposed to the use of the memoise package for instance)

##### Tests

From jarl:

```sh
jarl check . -s TESTTHAT
warning: expect_no_match
   --> tests/testthat/test-camR-render-helpers.R:249:3
    |
249 | /   expect_false(
250 | |     grepl(
...   |
254 | |     )
255 | |   )
    | |___- `expect_false(grepl(...))` is not as clear as `expect_no_match(...)`.
    |
    = help: Use `expect_no_match(...)` instead.

warning: expect_no_match
   --> tests/testthat/test-camR-render-helpers.R:257:3
    |
257 | /   expect_false(
258 | |     grepl(
...   |
262 | |     )
263 | |   )
    | |___- `expect_false(grepl(...))` is not as clear as `expect_no_match(...)`.
    |
    = help: Use `expect_no_match(...)` instead.

warning: expect_no_match
   --> tests/testthat/test-camR-render-helpers.R:340:3
    |
340 | /   expect_false(
341 | |     grepl(
...   |
345 | |     )
346 | |   )
    | |___- `expect_false(grepl(...))` is not as clear as `expect_no_match(...)`.
    |
    = help: Use `expect_no_match(...)` instead.

── Summary ──────────────────────────────────────
Found 3 errors.
3 fixable with the `--fix` option.
```

It doesn't seem test files are aligned with R files.
It is not an obligation but can be very handy: https://r-pkgs.org/testing-basics.html#create-a-test

Are there tests for the Shiny app?

---
