### 2026-08-17 f85b2ad
Fix duplicate GitHub installation in install_All
