### 2026-08-17 38e6d05
Use suppressible messages in install_All
