### 2026-08-17 6968f6f
Align package citation metadata
