+++ CITATION.cff
@@ -1,4 +1,3 @@
-# ------------------------------------------------
 # CITATION.cff file for camtrapReport
 # ------------------------------------------------
 
@@ -24,7 +23,6 @@ authors:
 
 repository-code: https://github.com/spatialecology/camtrapReport
 url: https://spatialecology.github.io/camtrapReport/
-date-released: '2026-08-09'
 
 contact:
   - family-names: Ebrahimi
@@ -39,4 +37,3 @@ keywords:
   - modular-r-package
   - reproducible-workflow
   - wildlife-monitoring
-
