### 2026-08-17 8621297
Improve REM failure handling
