### 2026-08-17 e2c0b1a
Update bundled Leuven example dataset
