+++ tests/testthat/test-camData.R
@@ -1,46 +1,135 @@
-test_that("the bundled Camtrap DP toy dataset is available and linked", {
+test_that("the bundled Leuven Camtrap DP example dataset is available and linked", {
   path <- camtrap_test_dataset()
-  deployments <- data.table::fread(file.path(path, "deployments.csv"))
-  media <- data.table::fread(file.path(path, "media.csv"))
-  observations <- data.table::fread(file.path(path, "observations.csv"))
-  media_event_ids <- sub("^sequenceID:", "", media$mediaComments)
-
+  
+  deployments <- data.table::fread(
+    file.path(path, "deployments.csv")
+  )
+  
+  media <- data.table::fread(
+    file.path(path, "media.csv")
+  )
+  
+  observations <- data.table::fread(
+    file.path(path, "observations.csv")
+  )
+  
+  media_event_ids <- sub(
+    "^sequenceID:",
+    "",
+    media$mediaComments
+  )
+  
   expect_true(dir.exists(path))
-  expect_true(all(file.exists(file.path(
-    path,
-    c("datapackage.json", "deployments.csv", "media.csv", "observations.csv")
-  ))))
-  expect_identical(data.table::uniqueN(deployments$locationID), 8L)
+  
+  expect_true(
+    all(
+      file.exists(
+        file.path(
+          path,
+          c(
+            "datapackage.json",
+            "deployments.csv",
+            "media.csv",
+            "observations.csv"
+          )
+        )
+      )
+    )
+  )
+  
+  expect_identical(
+    data.table::uniqueN(deployments$locationID),
+    8L
+  )
+  
+  expect_setequal(
+    unique(deployments$locationID),
+    c(
+      "LEUVEN_416",
+      "LEUVEN_881",
+      "LEUVEN_930",
+      "LEUVEN_1210",
+      "LEUVEN_1304",
+      "LEUVEN_1434",
+      "LEUVEN_1473",
+      "LEUVEN_1713"
+    )
+  )
+  
   expect_setequal(
     unique(substr(deployments$deploymentStart, 1L, 4L)),
-    c("2022", "2023", "2024")
+    c(
+      "2018",
+      "2019",
+      "2020",
+      "2021",
+      "2022",
+      "2023"
+    )
   )
-  expect_false(anyDuplicated(deployments$deploymentID) > 0L)
-  expect_false(anyDuplicated(media$mediaID) > 0L)
-  expect_false(anyDuplicated(observations$observationID) > 0L)
-  expect_true(all(media$deploymentID %in% deployments$deploymentID))
-  expect_true(all(observations$deploymentID %in% deployments$deploymentID))
-  expect_true(all(
-    observations$mediaID[observations$observationLevel == "media"] %in%
-      media$mediaID
-  ))
-  expect_true(all(observations$eventID %in% media_event_ids))
+  
+  expect_false(
+    anyDuplicated(deployments$deploymentID) > 0L
+  )
+  
+  expect_false(
+    anyDuplicated(media$mediaID) > 0L
+  )
+  
+  expect_false(
+    anyDuplicated(observations$observationID) > 0L
+  )
+  
+  expect_true(
+    all(
+      media$deploymentID %in%
+        deployments$deploymentID
+    )
+  )
+  
+  expect_true(
+    all(
+      observations$deploymentID %in%
+        deployments$deploymentID
+    )
+  )
+  
+  expect_true(
+    all(
+      observations$mediaID[
+        observations$observationLevel == "media"
+      ] %in% media$mediaID
+    )
+  )
+  
+  expect_true(
+    all(
+      observations$eventID %in%
+        media_event_ids
+    )
+  )
+  
[…]
