+++ inst/external/README-Leuven-dataset.md
@@ -1,0 +1,181 @@
+# Leuven example data for camtrapReport
+
+## Purpose
+
+The bundled Leuven data are a compact derived subset of the
+**GMU8_LEUVEN** Camtrap DP dataset, prepared for `camtrapReport` examples,
+automated tests, and software demonstrations.
+
+The subset is intended to exercise package functionality across multiple
+years, repeated camera deployments, multiple species, and REM-related fields.
+It should not be treated as a dataset prepared for ecological inference.
+
+## Source dataset
+
+The subset was derived from:
+
+Casaer J, Boone N, Vercammen J, Devisscher S, Pallemaerts L, Rutten A,
+Bollen M, Desmet P, Govaert S (2025). *GMU8_LEUVEN - Camera trap
+observations in natural habitats south of Leuven (Belgium).* Version 1.
+Camtrap DP dataset.
+
+The source metadata specifies:
+
+- data licence: CC0-1.0
+- media licence: CC-BY-4.0
+
+These licence declarations are retained in the derived Camtrap DP metadata.
+
+## Subset composition
+
+The bundled subset contains:
+
+- 8 derived example location identifiers
+- 87 deployments
+- deployment-start years from 2018 through 2023
+- 11,092 raw Camtrap DP observation records
+- 10,182 unique events
+- 15,492 retained media records
+
+All observation records associated with the selected deployments were
+retained. Observations were not selected on the basis of species or
+observation type.
+
+## Location identifiers
+
+The source dataset uses deployment-specific `locationID` values, while
+`locationName` values contain a trailing site code, for example
+`2020_7_1473`.
+
+For this derived software fixture, deployments sharing the same trailing
+site code were assigned a common location identifier so that repeated
+deployments can be handled consistently in package examples and tests.
+
+The identifiers used in the fixture are:
+
+- LEUVEN_416
+- LEUVEN_881
+- LEUVEN_930
+- LEUVEN_1210
+- LEUVEN_1304
+- LEUVEN_1434
+- LEUVEN_1473
+- LEUVEN_1713
+
+These identifiers are specific to the derived `camtrapReport` fixture and
+should not be interpreted as official location identifiers from the source
+dataset.
+
+The original `deploymentID` values are retained unchanged.
+
+Coordinates associated with the same derived location showed small
+differences among deployments. For reproducibility within the fixture,
+latitude and longitude were standardized to the median coordinate for each
+derived location and rounded to 0.001 degrees.
+
+| locationID | latitude | longitude |
+|---|---:|---:|
+| LEUVEN_416 | 50.769 | 4.638 |
+| LEUVEN_881 | 50.803 | 4.635 |
+| LEUVEN_930 | 50.805 | 4.634 |
+| LEUVEN_1210 | 50.816 | 4.674 |
+| LEUVEN_1304 | 50.821 | 4.649 |
+| LEUVEN_1434 | 50.828 | 4.638 |
+| LEUVEN_1473 | 50.829 | 4.636 |
+| LEUVEN_1713 | 50.845 | 4.649 |
+
+## Media reduction
+
+The source dataset contains substantially more media records than are needed
+for package testing.
+
+To keep the installed package reasonably small, the derived fixture retains:
+
+- the first media record associated with each selected event;
+- the last media record associated with each selected event; and
+- media records explicitly referenced by retained observations.
+
+Intermediate media records were omitted.
+
+Consequently, the number of media records associated with an event in this
+fixture should not be interpreted as the original number of photographs in
+the source dataset.
+
+## Habitat example
+
+A small location-level habitat table is provided in:
+
+`inst/external/habitat/habitat.csv`
+
+The habitat information was derived from habitat information associated with
+the selected Leuven deployments.
+
[…]
