### 2026-08-18 55eabb9
Address easy rOpenSci goodpractice findings
