+++ tests/testthat/test-camData.R
@@ -1,4 +1,6 @@
-test_that("the bundled Leuven Camtrap DP example dataset is available and linked", {
+test_that(
+  "the bundled Leuven Camtrap DP example dataset is available and linked",
+  {
   path <- camtrap_test_dataset()
   
   deployments <- data.table::fread(
@@ -119,10 +121,13 @@ test_that("the bundled Leuven Camtrap DP example dataset is available and linked
       "unknown"
     )
   )
-})
+  }
+)
 
 
-test_that("camData reads the bundled Leuven dataset into a complete camReport", {
+test_that(
+  "camData reads the bundled Leuven dataset into a complete camReport",
+  {
   cm <- camtrap_test_report()
   
   expect_s4_class(
@@ -192,7 +197,8 @@ test_that("camData reads the bundled Leuven dataset into a complete camReport", 
   expect_true(
     all(file.exists(cache_files))
   )
-})
+  }
+)
 
 
 test_that("camData reuses its saved camReport object", {
@@ -216,5 +222,5 @@ test_that("camData reuses its saved camReport object", {
     nrow(cached$data$observations),
     nrow(cm$data$observations)
   )
-})
-
+}
+)
