+++ tests/testthat/test-report-api.R
@@ -48,7 +48,6 @@ test_that("section selectors expose and update valid modules", {
   selected <- attached[[1]]
 
   expect_message(
-    # nolint next: implicit_assignment_linter.
     result <- sections(cm, selected),
     "report sections are updated",
     fixed = TRUE
