+++ tests/testthat/helper-cam-report.R
@@ -60,7 +60,6 @@ camtrap_test_report <- function() {
     
     invisible(
       capture.output(
-        # nolint next: implicit_assignment_linter.
         object <- withCallingHandlers(
           suppressMessages(
             camData(
