### 2026-08-18 a73317b
Remove duplicate helper definitions
