### 2026-08-18 f9ec183
Simplify camData evaluation and datetime parsing
