### 2026-08-18 d1aaf0a
Remove dynamic evaluation from camData
