### 2026-08-18 7dd1897
Remove dynamic evaluation from camR
