+++ R/camR.R
@@ -311,7 +311,7 @@ camR <- setRefClass(
       
       if (!any(
         c('order', 'class', 'scientificName', 'observationType') %in%
-          names(x)
+        names(x)
       ))
         stop(
           # nolint start: line_length_linter.
@@ -502,8 +502,8 @@ camR <- setRefClass(
       .rich
     },
     spatial_density = function(
-      x = NULL, species = NULL, year = NULL,
-      .crs = NULL, .ext = NULL, .crop = NULL
+    x = NULL, species = NULL, year = NULL,
+    .crs = NULL, .ext = NULL, .crop = NULL
     ) {
       # spatial density raster map
       # x: the output of species_summary_by_location wit cor_matrix=F
@@ -534,9 +534,9 @@ camR <- setRefClass(
         if (!is.null(.crop)) {
           if (
             xr[1] < .ext[1] |
-              xr[2] > .ext[2] |
-              yr[1] < .ext[3] |
-              yr[2] > .ext[4]
+            xr[2] > .ext[2] |
+            yr[1] < .ext[3] |
+            yr[2] > .ext[4]
           ) {
             .ww <- NULL
             .w <- xr[1] - .ext[1]
@@ -593,18 +593,16 @@ camR <- setRefClass(
       yr[1] <- yr[1] - .w
       yr[2] <- yr[2] + .w
       #------
-      .win <- .eval(
-        "spatstat.geom::owin(xrange = xr, yrange = yr)",
-        env = environment()
+      .win <- spatstat.geom::owin(
+        xrange = xr,
+        yrange = yr
       )
-      .ppp_obj <- .eval(
-        "spatstat.geom::ppp(
-    x = x$x,
-    y = x$y,
-    window = .win,
-    marks = x$total_observations
-  )",
-        env = environment()
+
+      .ppp_obj <- spatstat.geom::ppp(
+        x = x$x,
+        y = x$y,
+        window = .win,
+        marks = x$total_observations
       )
       den <- density(.ppp_obj, weights = x$total_observations)
       den$v <- den$v / max(den$v, na.rm = TRUE)
@@ -616,8 +614,8 @@ camR <- setRefClass(
       r
     },
     species_summary_by_location = function(
-      year = NULL, spList = NULL,
-      cor_matrix = TRUE, PA = TRUE
+    year = NULL, spList = NULL,
+    cor_matrix = TRUE, PA = TRUE
     ) {
       # cor_matrix = T -> cor of co-occurrence matrix across
       # locations will be returned
@@ -727,7 +725,7 @@ camR <- setRefClass(
         
         if (
           is.null(.self$.act_models[[sp]]) ||
-            !is.list(.self$.act_models[[sp]])
+          !is.list(.self$.act_models[[sp]])
         ) {
           x <- try({
             deployment_locations <- .self$data$deployments |>
@@ -859,21 +857,24 @@ camR <- setRefClass(
                 reps = 10
               )
               .density_estimates <- .rem(.parameters)
-              # nolint start: line_length_linter.
-              .density_estimates <- .eval("camtrapDensity::convert_units(.density_estimates,radius_unit = \"m\",angle_unit = \"degree\",active_speed_unit = \"km/hour\",overall_speed_unit = \"km/day\")",
-              # nolint end
-                                          environment())
+              .density_estimates <- camtrapDensity::convert_units(
+                .density_estimates,
+                radius_unit = "m",
+                angle_unit = "degree",
+                active_speed_unit = "km/hour",
+                overall_speed_unit = "km/day"
+              )
               if ("vernacularNames.eng" %in% colnames(dat$taxonomy)) {
                 english_name <-
                   dat$taxonomy$vernacularNames.eng[
                     dat$taxonomy$scientificName ==
[…]
