### 2026-08-18 85a2deb
Remove dynamic evaluation from camR functions
