### 2026-08-18 bc1019a
Remove dynamic evaluation from supporting functions
