+++ R/supporting_functions.R
@@ -20,20 +20,31 @@
   # end_col   <- .first_existing(dep, c("deploymentEnd", "end"))
   
   # nolint start: line_length_linter.
-  dep <- .eval('dep |>
+  dep <- dep |>
     dplyr::mutate(
-      effort_seconds = as.numeric(difftime(.data[["deploymentEnd"]], .data[["deploymentStart"]], units = "secs"))
+      effort_seconds = as.numeric(
+        difftime(
+          .data[["deploymentEnd"]],
+          .data[["deploymentStart"]],
+          units = "secs"
+        )
+      )
     ) |>
-    dplyr::filter(!is.na(.data$effort_seconds), .data$effort_seconds >= 0)',environment())
+    dplyr::filter(
+      !is.na(.data$effort_seconds),
+      .data$effort_seconds >= 0
+    )
   # nolint end
   
   if (!is.null(x$locations) &&
       nrow(x$locations) > 0 &&
       "locationID" %in% names(dep) &&
       "locationID" %in% names(x$locations)) {
-    dep <- .eval(
-      'dplyr::left_join(dep, x$locations, by = "locationID", multiple = "all")',
-      environment()
+    dep <- dplyr::left_join(
+      dep,
+      x$locations,
+      by = "locationID",
+      multiple = "all"
     )
   }
   
@@ -45,17 +56,19 @@
     week   = 604800
   )[[unit]]
   
-  dep <- .eval('dep |>
-    dplyr::mutate(effort = .data$effort_seconds / divisor)',environment())
+  dep <- dep |>
+    dplyr::mutate(
+      effort = .data$effort_seconds / divisor
+    )
   
   if (is.null(by)) {
-    out <- .eval('dep |>
+    out <- dep |>
       dplyr::summarise(
         effort_seconds = sum(.data$effort_seconds, na.rm = TRUE),
         effort = sum(.data$effort, na.rm = TRUE),
         unit = unit,
         .groups = "drop"
-      )',environment())
+      )
   } else {
     by <- unique(by)
     missing_by <- setdiff(by, names(dep))
@@ -66,14 +79,16 @@
       )
     }
     
-    out <- .eval('dep |>
-      dplyr::group_by(dplyr::across(dplyr::all_of(by))) |>
+    out <- dep |>
+      dplyr::group_by(
+        dplyr::across(dplyr::all_of(by))
+      ) |>
       dplyr::summarise(
         effort_seconds = sum(.data$effort_seconds, na.rm = TRUE),
         effort = sum(.data$effort, na.rm = TRUE),
         unit = unit,
         .groups = "drop"
-      )',environment())
+      )
   }
   
   out
@@ -744,10 +759,7 @@
   
   mod <- suppressWarnings(
     suppressMessages(
-      .eval(
-        'do.call(Distance::ds,args)$ddf',
-        environment()
-      )
+      do.call(Distance::ds, args)$ddf
     )
   )
   
@@ -805,7 +817,7 @@
   if (nrow(dat) == 0) stop("There are no usable speed data")
   
   args <- c(list(formula = formula,data=dat),list(...))
-  res <- .eval('do.call(sbd::sbm,args)',environment())
+  res <- do.call(sbd::sbm, args)
   
   res$unit <- paste0(distUnit, "/", timeUnit) # nolint
   res
@@ -841,24 +853,34 @@
   obs <- obs[i, ]
   
   if (nrow(obs) > 1) {
-    suntimes <- .eval(
-      'activity::get_suntimes(obs$timestamp, obs$latitude, obs$longitude, 0)',
-      environment()
+    suntimes <- activity::get_suntimes(
+      obs$timestamp,
+      obs$latitude,
+      obs$longitude,
+      0
[…]
