### 2026-08-18 2016a39
Remove obsolete evaluation helper
