### 2026-08-20 7cab0b3
Address rOpenSci editor comments
