+++ tests/testthat/test-report-output-paths.R
@@ -32,8 +32,9 @@ test_that("report returns rendering errors without crashing", {
   original <- camtrap_test_report()
   report_object <- original$copy(shallow = FALSE)
   
-  output_dir <- tempfile("camtrapReport-report-error-")
-  dir.create(output_dir)
+  output_dir <- withr::local_tempdir(
+    pattern = "camtrapReport-report-error-"
+  )
   
   on.exit(
     unlink(output_dir, recursive = TRUE, force = TRUE),
@@ -77,8 +78,9 @@ test_that("report warns and falls back when output directory is missing", {
   original <- camtrap_test_report()
   report_object <- original$copy(shallow = FALSE)
   
-  fallback_dir <- tempfile("camtrapReport-fallback-")
-  dir.create(fallback_dir)
+  fallback_dir <- withr::local_tempdir(
+    pattern = "camtrapReport-fallback-"
+  )
   
   on.exit(
     unlink(fallback_dir, recursive = TRUE, force = TRUE),
@@ -128,8 +130,9 @@ test_that("report does not invoke viewer after rendering failure", {
   original <- camtrap_test_report()
   report_object <- original$copy(shallow = FALSE)
   
-  output_dir <- tempfile("camtrapReport-viewer-")
-  dir.create(output_dir)
+  output_dir <- withr::local_tempdir(
+    pattern = "camtrapReport-viewer-"
+  )
   
   old_viewer <- getOption("viewer")
   viewer_state <- new.env(parent = emptyenv())
