+++ R/camtrapReport-package.R
@@ -18,7 +18,7 @@
 #' * <https://github.com/spatialecology/camtrapReport>
 #' * Report bugs at <https://github.com/spatialecology/camtrapReport/issues>
 #'
-#' @keywords package
+#' @keywords internal
 #' @importFrom methods new setClass setGeneric setMethod
 #' @importFrom methods setReplaceMethod signature
 #' @importFrom graphics abline legend par points lines boxplot
