### 2026-08-20 76a54c4
Finalize README and documentation
