+++ README.Rmd
@@ -1,0 +1,365 @@
+<p align="center">
+  <img src="inst/report-assets/camtrapReport-logo.png" width="200" alt="camtrapReport"/>
+</p>
+
+<h1 align="center">camtrapReport</h1>
+
+<p align="center">
+  <em>A modular R package for automating camera-trap data reporting in wildlife monitoring.</em>
+</p>
+
+<p align="center">
+  <a href="https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml"><img src="https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml/badge.svg" alt="R-CMD-check"></a>
+  <a href="https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml"><img src="https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml/badge.svg" alt="pkgcheck"></a>
+  <a href="https://app.codecov.io/github/spatialecology/camtrapReport"><img src="https://codecov.io/github/spatialecology/camtrapReport/graph/badge.svg?token=9VBXAR9XOD" alt="Codecov test coverage"></a>
+  <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project Status: Active"></a>
+  <a href="https://www.r-project.org/"><img src="https://img.shields.io/badge/R-%E2%89%A5%204.1.0-276DC3?logo=r&logoColor=white" alt="R ≥ 4.1.0"></a>
+  <a href="https://github.com/spatialecology/camtrapReport/blob/main/LICENSE.md"><img src="https://img.shields.io/badge/license-MIT-yellow.svg" alt="MIT"></a>
+  <a href="https://spatialecology.github.io/camtrapReport/"><img src="https://img.shields.io/badge/docs-pkgdown-orange.svg" alt="pkgdown"></a>
+</p>
+
+---
+
+`camtrapReport` turns standardised camera-trap datasets into structured, reproducible ecological reports through an automated workflow. Drop in a [Camtrap DP](https://camtrap-dp.tdwg.org/) ZIP file, and the package will diagnose data quality, run a suite of ecological analyses, and compile narrative text, figures, maps, and tables into a single article-style HTML document.
+
+<p align="center">
+  <img src="vignettes/articles/figures/workflow.png" width="900" alt="Schematic overview of the camtrapReport workflow"/>
+  <br>
+  <em>Schematic overview of the automated <code>camtrapReport</code> workflow.</em>
+</p>
+
+## Use in practice
+
+`camtrapReport` has been developed and evaluated using multiple real-world
+camera-trap datasets from different parts of the world, spanning a wide range of
+geographic regions, monitoring contexts, survey designs, and data structures.
+A substantial part of this evaluation has involved datasets from across Europe,
+particularly through the European Observatory of Wildlife (EOW) network. The
+package has also been presented at conferences, workshops, and training events,
+and is increasingly being used and tested by external users.
+
+Testing the package across these heterogeneous datasets has helped identify
+practical issues related to data structure, metadata completeness, spatial and
+temporal coverage, and report generation. Feedback from users working with these
+datasets has been incorporated into the continued development and refinement of
+the package.
+
+## Installation
+
+`camtrapReport` requires R 4.1.0 or later. Install the development version from GitHub with [`pak`](https://pak.r-lib.org/):
+
+```r
+if (!requireNamespace("pak", quietly = TRUE)) {
+  install.packages("pak")
+}
+
+pak::pkg_install(
+  "spatialecology/camtrapReport",
+  dependencies = TRUE
+)
+
+library(camtrapReport)
+```
+
[…]
