### 2026-08-20 eeada29
Fix CI dependencies and complete documentation rebuild
