+++ pkgdown/index.md
@@ -1,183 +1,0 @@
----
-title: "Get started"
----
-
-# Get started
-
-## Package aims
-
-[`camtrapReport`](https://github.com/spatialecology/camtrapReport) provides a
-modular workflow for turning standardised camera-trap data into reproducible
-wildlife-monitoring reports. It combines data-quality assessment, ecological
-analysis, visualisation, and report generation within a single workflow.
-
-The package produces two main outputs:
-
-- **Data Status Check**, which evaluates data quality, completeness, and
-  spatiotemporal structure before ecological analysis; and
-- **Ecological Report**, which assembles selected ecological analyses,
-  summaries, tables, figures, and maps into a structured report.
-
-For more detail on the package scope and its relationship to other
-camera-trap tools, see the
-[Package Overview](articles/Package-Overview.html).
-
-## Data format
-
-`camtrapReport` uses datasets in the [Camtrap DP format](https://camtrap-dp.tdwg.org/), a community-developed data exchange format for camera-trap data. Camera-trap data management systems including [Agouti](https://agouti.eu/) and [TRAPPER](https://os-conservation.org/trapper/) can export Camtrap DP datasets.
-
-Data from other sources, including manually managed datasets and Wildlife Insights exports, can also be used after conversion to Camtrap DP.
-
-## Installation
-
-`camtrapReport` requires R 4.1.0 or later. Install the development version from GitHub with [`pak`](https://pak.r-lib.org/):
-
-```r
-if (!requireNamespace("pak", quietly = TRUE)) {
-  install.packages("pak")
-}
-
-pak::pkg_install(
-  "spatialecology/camtrapReport",
-  dependencies = TRUE
-)
-
-library(camtrapReport)
-```
-
-Using `dependencies = TRUE` also installs optional packages required by some report modules and examples.
-
-## Quick start with the bundled example data
-
-`camtrapReport` includes a compact example dataset derived from the **GMU8_LEUVEN** Camtrap DP dataset, together with example habitat data and a study-area boundary. These files are bundled with the package for examples, automated tests, and software demonstrations.
-
-Because `camData()` stores a processed `camReport` object beside the input dataset, the bundled dataset is first copied to a temporary working directory so that the installed package files remain unchanged.
-
-```r
-library(camtrapReport)
-
-# Copy the bundled example dataset to a temporary working directory
-example_dir <- tempfile("camtrapReport-")
-dir.create(example_dir)
-
-file.copy(
-  system.file("external", "dataset", package = "camtrapReport"),
-  example_dir,
-  recursive = TRUE
-)
-
-# Create a camReport object
-cm <- camData(file.path(example_dir, "dataset"))
-
-cm
-```
-
-The bundled Leuven fixture is intended for demonstrating and testing the
-software rather than for ecological inference. Its provenance and the
-transformations used to prepare it are documented in the
-[bundled dataset documentation](https://github.com/spatialecology/camtrapReport/blob/main/inst/external/README-Leuven-dataset.md).
-
-## Using your own data
-
-The only required input to `camData()` is a Camtrap DP dataset, supplied either as a ZIP archive or as an extracted Camtrap DP directory.
-
-Optional habitat data and a study-area boundary can also be supplied to add spatial context to maps, summaries, and analyses. Habitat data should include `locationName` and `Habitat`; the study area can be provided as a spatial file path, `SpatVector`, or `sf` object.
-
-```r
-# Required input
-dataset <- "path/to/your/dataset.zip"
-
-# Create a camReport object using only the required input
[…]
