### 2026-08-20 7b422bc
Fix camtraptor dependency resolution
