### 2026-08-20 c16335a
Finish addressing rOpenSci editor feedback
