### 2026-08-20 75756ff
Use unique labels in generated report chunks
