+++ R/testSection.R
@@ -32,7 +32,8 @@
   # Reuse the same package-loader helper used by generateReport()
   .env$pkg_chunk <- .make_package_loader_chunk(
     pkgs = module_pkgs,
-    core = "knitr"
+    core = "knitr",
+    label = "camtrap-quick-test-packages"
   )
   
   rmd_template <- glue::glue(
@@ -106,7 +107,8 @@ output:
   # Reuse the same package-loader helper used by generateReport()
   .env$pkg_chunk <- .make_package_loader_chunk(
     pkgs = module_pkgs,
-    core = "knitr"
+    core = "knitr",
+    label = "camtrap-section-test-packages"
   )
   
   rmd_template <- glue::glue(
