### 2026-08-20 6426cea
Address rOpenSci editor feedback
