+++ tests/testthat/test-install-all-branches.R
@@ -1,553 +1,0 @@
-test_that("installation helpers validate and classify package names", {
-  is_installed <- ct_internal(".is.installed")
-  load_lib <- ct_internal(".loadLib")
-  
-  empty_result <- is_installed(character())
-  
-  expect_type(empty_result, "logical")
-  expect_length(empty_result, 0L)
-  expect_named(empty_result, character())
-  
-  installed <- is_installed(
-    c(
-      "methods",
-      "stats",
-      "",
-      NA_character_,
-      "a_package_that_does_not_exist_12345"
-    )
-  )
-  
-  expect_true(unname(installed["methods"]))
-  expect_true(unname(installed["stats"]))
-  
-  expect_false(
-    unname(installed[3])
-  )
-  
-  expect_false(
-    unname(installed[4])
-  )
-  
-  expect_false(
-    unname(
-      installed["a_package_that_does_not_exist_12345"]
-    )
-  )
-  
-  expect_identical(
-    unname(
-      load_lib(
-        list(
-          "methods",
-          c("stats", "methods")
-        )
-      )
-    ),
-    c(TRUE, TRUE)
-  )
-  
-  expect_false(
-    unname(
-      load_lib(
-        list(
-          "a_package_that_does_not_exist_12345"
-        )
-      )
-    )[1]
-  )
-})
-
-
-test_that("install_All rejects invalid arguments before installation", {
-  expect_error(
-    install_All(update = NA),
-    "'update' must be TRUE or FALSE",
-    fixed = TRUE
-  )
-  
-  expect_error(
-    install_All(update = 1),
-    "'update' must be TRUE or FALSE",
-    fixed = TRUE
-  )
-  
-  expect_error(
-    install_All(update = c(TRUE, FALSE)),
-    "'update' must be TRUE or FALSE",
-    fixed = TRUE
-  )
-  
-  expect_error(
-    install_All(pkgs = 1),
-    "'pkgs' must be NULL or a character vector",
-    fixed = TRUE
-  )
-  
-  expect_error(
-    install_All(pkgs = c("methods", NA_character_)),
-    "'pkgs' must be NULL or a character vector",
-    fixed = TRUE
-  )
-})
-
-
-test_that("install_All reports when all requested packages are installed", {
-  state <- new.env(parent = emptyenv())
-  state$checked_packages <- character()
-  
-  testthat::local_mocked_bindings(
-    .getPackageList = function() {
-      c("methods", "stats")
-    },
-    .getPackageGitHubList = function() {
-      character()
-    },
-    .is.installed = function(n) {
-      n <- as.character(n)
-      
-      state$checked_packages <- unique(
-        c(
-          state$checked_packages,
-          n
-        )
-      )
-      
-      result <- rep(TRUE, length(n))
-      names(result) <- n
-      result
-    },
-    .package = "camtrapReport"
-  )
-  
-  expect_message(
-    # nolint next: implicit_assignment_linter.
-    result <- install_All(
-      pkgs = c(" methods ", "", "stats"),
-      update = FALSE
-    ),
-    "All required packages have already been installed",
-    fixed = TRUE
-  )
-  
-  expect_null(result)
-  
-  expect_true(
-    all(
-      c("methods", "stats") %in% state$checked_packages
-    )
-  )
-})
-
-
-test_that("install_All attempts only missing CRAN packages", {
-  state <- new.env(parent = emptyenv())
-  state$installed <- FALSE
-  state$installation_calls <- character()
-  
-  testthat::local_mocked_bindings(
-    .getPackageList = function() {
-      c("methods", "missingPackage")
-    },
-    .getPackageGitHubList = function() {
-      character()
-    },
-    .is.installed = function(n) {
-      n <- as.character(n)
-      
[…]
