### 2026-08-21 c55a32a
Restore module installer and fix report rendering
