+++ NAMESPACE
@@ -7,6 +7,7 @@ export(camR)
 export(empty_trash)
 export(gui)
 export(info)
+export(install_all)
 export(listReportSections)
 export(list_Modules)
 export(move_Module)
