### 2026-08-21 bde5b9e
Finalize documentation and pkgdown articles
