+++ README.md
@@ -1,11 +1,6 @@
 
 <!-- README.md is generated from README.Rmd. Edit README.Rmd, not README.md. -->
 
-<p align="center">
-
-<img src="inst/report-assets/camtrapReport-logo.png" width="200" alt="camtrapReport logo"/>
-</p>
-
 <h1 align="center">
 
 camtrapReport
@@ -29,47 +24,50 @@ camtrapReport
 
 ------------------------------------------------------------------------
 
-`camtrapReport` turns camera-trap data in [Camtrap
+`camtrapReport` converts camera-trap data in [Camtrap
 DP](https://camtrap-dp.tdwg.org/) format into two coordinated HTML
 outputs:
 
 - a **Data Status Check**, which identifies issues in data completeness,
   consistency, annotation, and spatiotemporal coverage; and
-- an **Ecological Report**, which brings selected ecological summaries,
-  analyses, figures, tables, maps, metadata, and explanatory text into a
-  configurable report.
+- an **Ecological Report**, which combines selected ecological
+  summaries, analyses, figures, tables, maps, metadata, and explanatory
+  text in a configurable report.
 
-The package focuses on the reporting workflow itself. It complements
-tools for camera-trap data management and specialised ecological
-modelling rather than replacing them.
+The package focuses on the reporting workflow. It complements tools for
+camera-trap data management and specialised ecological modelling rather
+than replacing them.
 
 <p align="center">
 
-<img src="vignettes/figures/package-architecture.png" width="900" alt="Workflow from Camtrap DP input through the camReport object to the Data Status Check and Ecological Report"/>
-<br> <em>Overview of the camtrapReport workflow.</em>
+<img src="vignettes/figures/package-architecture.png"
+       width="900"
+       alt="Workflow from Camtrap DP input through the camReport object to the Data Status Check and Ecological Report"><br>
+<em>Overview of the camtrapReport workflow.</em>
 </p>
 
 ## Use in practice
 
-`camtrapReport` has been developed and evaluated with real camera-trap
-monitoring datasets that differ in geography, survey design, duration,
-and data structure. These include datasets from the European Observatory
-of Wildlife network and datasets used by external teams in Spain,
-Slovenia, and Germany. The package has also been tested in workshops at
-Utrecht University and the Research Institute for Nature and Forest in
-Belgium.
+`camtrapReport` was developed and evaluated using camera-trap monitoring
+datasets from regions worldwide, including North America, Asia, Europe,
+Africa, and Australia. The package was tested both by its developer and
+by users who voluntarily applied it to their own datasets and provided
+feedback. Evaluation in Europe has benefited particularly from datasets
+available through camera-trap research networks, including the [European
+Observatory of Wildlife network](https://enetwild.com/), which spans
+diverse geographic settings and monitoring designs.
 
-These applications have primarily supported software validation, report
-generation, and user feedback. Experience with heterogeneous datasets
-has informed checks for metadata completeness, temporal and spatial
-coverage, and robust report generation. Publications, presentations, and
-training activities are listed on the [Resources
+Experience across these heterogeneous datasets has informed checks for
+metadata completeness and spatiotemporal coverage, improvements to
+report generation, and revisions based on user feedback. Related
+publications, presentations, and training activities are listed on the
+[Resources
 page](https://spatialecology.github.io/camtrapReport/articles/resources.html).
 
[…]
