+++ vignettes/articles/data-status-report.Rmd
@@ -43,8 +43,34 @@ knitr::include_graphics("figures/data-status/data-status-workflow.png")
 ```{r data-status-basic-code, eval=FALSE}
 library(camtrapReport)
 
-# Load camera-trap data
-cm <- camData("cameratrap.zip")
+# Copy the bundled example dataset to a writable temporary directory
+example_root <- tempfile("camtrapReport-example-")
+dir.create(example_root, recursive = TRUE)
+
+example_source <- system.file(
+  "external",
+  "dataset",
+  package = "camtrapReport",
+  mustWork = TRUE
+)
+
+copied <- file.copy(
+  from = example_source,
+  to = example_root,
+  recursive = TRUE,
+  copy.mode = FALSE
+)
+
+stopifnot(
+  isTRUE(copied),
+  dir.exists(file.path(example_root, "dataset"))
+)
+
+# Load the example camera-trap data
+cm <- camData(
+  file.path(example_root, "dataset"),
+  update = TRUE
+)
 
 # Generate and view the Data Status Report
 status(cm, view = TRUE)
@@ -58,6 +84,12 @@ the R session's temporary directory. Use `filename` to give another file stem
 or a path inside an existing output directory:
 
 ```{r data-status-output-path, eval=FALSE}
+dir.create(
+  "reports",
+  recursive = TRUE,
+  showWarnings = FALSE
+)
+
 status_file <- status(
   cm,
   filename = file.path("reports", "data_status_2026"),
@@ -191,7 +223,7 @@ Below are previews of Data Status Reports generated with `camtrapReport` for dif
 
 ## Exploring data-status results in R
 
-The results of the data-status checks are also stored inside the `camReport` object. Users can explore these results directly in R through the `data_status` slot. This is useful when users want to inspect the underlying tables, extract summaries, or use the quality-control outputs in another workflow.
+The results of the data-status checks are also stored inside the `camReport` object. Users can explore these results directly in R through `cm$data_status`. This is useful when users want to inspect the underlying tables, extract summaries, or use the quality-control outputs in another workflow.
 
 ```{r data-status-explore-object, eval=FALSE}
 # Explore all data-status outputs stored in the camReport object
