### 2026-08-21 a237421
Restore module installer and finalize documentation
