+++ README.Rmd
@@ -85,8 +85,7 @@ if (!requireNamespace("pak", quietly = TRUE)) {
 }
 
 pak::pkg_install(
-  "spatialecology/camtrapReport",
-  dependencies = TRUE
+  "spatialecology/camtrapReport"
 )
 ```
 
