+++ README.Rmd
@@ -165,7 +165,7 @@ directory. Habitat data and a study-area boundary are optional.
 
 ### Generate a Data Status Check
 
-```{r data-status}
+```{r generate-data-status, message=FALSE, warning=FALSE, results='hide'}
 status_file <- status(
   cm,
   view = FALSE
@@ -179,7 +179,7 @@ stopifnot(
 
 ### Generate an Ecological Report
 
-```{r ecological-report}
+```{r generate-ecological-report, message=FALSE, warning=FALSE, results='hide'}
 report_file <- report(
   cm,
   view = FALSE
