+++ README.md
@@ -150,7 +150,7 @@ cm <- camData(
 #> Dataset size: 6.6 MB.
 #> File size looks modest, but full object creation may still take several minutes depending on the number of records.
 #> Setup is done!
-#> Data loaded successfully in 1 min 20 sec.
+#> Data loaded successfully in 1 min 06 sec.
 #> camReport object is ready for GMU8 LEUVEN.
 
 cm
@@ -175,8 +175,6 @@ status_file <- status(
   cm,
   view = FALSE
 )
-#> Rendering R Markdown data_status report ...
-#> Data_Status Report generated at: C:\Users\ebrah010\AppData\Local\Temp\Rtmpc5CA1e\data_status.html
 
 stopifnot(
   !inherits(status_file, "try-error"),
@@ -191,8 +189,6 @@ report_file <- report(
   cm,
   view = FALSE
 )
-#> Rendering R Markdown report ...
-#> Report generated at: C:\Users\ebrah010\AppData\Local\Temp\Rtmpc5CA1e\report.html
 
 stopifnot(
   !inherits(report_file, "try-error"),
