### 2026-08-22 e1dc1d1
Split utility helpers into topic-specific files
