### 2026-08-22 74e3569
Complete documentation updates for rOpenSci review
