### 2026-08-22 d03d4c5
Refine README examples and output
