+++ README.Rmd
@@ -94,11 +94,11 @@ This executed example uses a compact Camtrap DP dataset derived from
 **GMU8_LEUVEN**, together with bundled habitat data and a study-area boundary.
 The dataset is copied to a temporary directory because `camData()` may store
 processed files beside writable input data.
-
 ```{r quick-start}
 library(camtrapReport)
 
 example_root <- tempfile("camtrapReport-example-")
+
 dir.create(
   example_root,
   recursive = TRUE
@@ -133,20 +133,25 @@ study_area <- system.file(
   package = "camtrapReport",
   mustWork = TRUE
 )
+```
 
+```{r validate-example-data, include=FALSE}
 stopifnot(
   isTRUE(copied),
-  dir.exists(file.path(example_root, "dataset")),
-  file.exists(study_area)
+  dir.exists(file.path(example_root, "dataset"))
 )
+```
 
+```{r create-camreport-object, message=FALSE, warning=FALSE, results='hide'}
 cm <- camData(
   data = file.path(example_root, "dataset"),
   habitat = habitat,
   study_area = study_area,
   update = TRUE
 )
+```
 
+```{r show-camreport-object}
 cm
 ```
 
@@ -160,9 +165,12 @@ status_file <- status(
   cm,
   view = FALSE
 )
+```
 
+```{r validate-data-status, include=FALSE}
 stopifnot(
-  !inherits(status_file, "try-error"),
+  is.character(status_file),
+  length(status_file) == 1L,
   file.exists(status_file)
 )
 ```
@@ -174,9 +182,12 @@ report_file <- report(
   cm,
   view = FALSE
 )
+```
 
+```{r validate-ecological-report, include=FALSE}
 stopifnot(
-  !inherits(report_file, "try-error"),
+  is.character(report_file),
+  length(report_file) == 1L,
   file.exists(report_file)
 )
 ```
@@ -185,8 +196,15 @@ stopifnot(
 
 Inspect report metadata and choose which sections to include:
 
-```{r configure-report}
-info(cm, name = c("title", "authors"))
+## Configure a report
+
+Inspect report metadata and choose which sections to include:
+
+```{r configure-report, message=FALSE, warning=FALSE, results='hide'}
+info(
+  cm,
+  name = c("title", "authors")
+)
 
 section_names()
 
@@ -200,7 +218,10 @@ selected_sections <- section_names(
   )
 )
 
-sections(cm, selected_sections)
+sections(
+  cm,
+  selected_sections
+)
 ```
 
 Reports can also be configured by taxonomic group, survey year, and
