### 2026-08-22 9fbcd0f
Refine package overview and README
