### 2026-08-22 2f3859f
Split utility helpers by topic and improve documentation
* Split utility helpers into topic-specific files

* Complete documentation updates for rOpenSci review

* Refine README examples and output

* Refine package overview and README
