### 2026-08-25 #799 @maelle
@spatialecology Do you know approximately when you'll answer in this thread? (I see activity in the GitHub repo so am not too worried 😸 ) Thanks!
