### 2026-08-25 #799 @maelle
@spatialecology when preparing the rOpenSci newsletter, where we summarize all current submissions, we noticed the URL in your GitHub profile, `https://www.wur.nl/en/persons/elham-ebrahimi.htm`, is broken.
