### 2026-08-25 #799 @spatialecology
> [@spatialecology](https://github.com/spatialecology) Do you know approximately when you'll answer in this thread? (I see activity in the GitHub repo so am not too worried 😸 ) Thanks!

Hi @maelle ,
I’m finalizing everything and expect to post an update in this thread by the end of today. Thank you for checking in!

Best,
Elham
