+++ DESCRIPTION
@@ -29,7 +29,6 @@ Imports:
 Suggests:
     activity,
     camtrapDensity,
-    camtraptor,
     corrplot,
     Distance,
     dygraphs,
@@ -39,16 +38,13 @@ Suggests:
     htmltools,
     htmlwidgets,
     iNEXT,
-    jpeg,
     knitr,
     leaflet,
     lutz,
     magick,
     pak,
-    plotly,
     readr,
     rmarkdown,
-    roxygen2,
     sbd,
     scales,
     sf,
@@ -59,10 +55,8 @@ Suggests:
     testthat (>= 3.1.7),
     tidyr,
     withr,
-    xml2,
     xts
 Remotes:
-    inbo/camtraptor,
     MarcusRowcliffe/camtrapDensity
 VignetteBuilder: knitr
 Encoding: UTF-8
