### 2026-08-25 #799 @spatialecology
> [@spatialecology](https://github.com/spatialecology) when preparing the rOpenSci newsletter, where we summarize all current submissions, we noticed the URL in your GitHub profile, `https://www.wur.nl/en/persons/elham-ebrahimi.htm`, is broken.

@maelle, thanks for noticing and letting me know! I have updated the URL in my GitHub profile. My current profile pages are:

https://www.uu.nl/staff/EEbrahimi
https://research.wur.nl/en/persons/elham-ebrahimi/
