+++ DESCRIPTION
@@ -29,6 +29,7 @@ Imports:
 Suggests:
     activity,
     camtrapDensity,
+    camtraptor,
     corrplot,
     Distance,
     dygraphs,
@@ -57,6 +58,7 @@ Suggests:
     withr,
     xts
 Remotes:
+    inbo/camtraptor,
     MarcusRowcliffe/camtrapDensity
 VignetteBuilder: knitr
 Encoding: UTF-8
