### 2026-08-25 6b2b529
Add rOpenSci pkgcheck report
