+++ ropensci-pkgcheck-report.md
@@ -1,0 +1,444 @@
+## Checks for [camtrapReport (v1.0.47)](https://github.com/spatialecology/camtrapReport)
+
+git hash: [d0fb7d6a](https://github.com/spatialecology/camtrapReport/tree/d0fb7d6a29253f4f19f84b4ed074596fe1e19530)
+
+- :heavy_check_mark: Package name is available
+- :heavy_check_mark: has a 'contributing' file.
+- :heavy_check_mark: uses 'roxygen2'.
+- :heavy_check_mark: 'DESCRIPTION' has a URL field.
+- :heavy_check_mark: 'DESCRIPTION' has a BugReports field.
+- :heavy_check_mark: Package has at least one HTML vignette
+- :heavy_check_mark: All functions have examples.
+- :heavy_check_mark: Repository has a website
+- :heavy_check_mark:  Package has continuous integration checks.
+- :heavy_check_mark: Package coverage is 78.4%.
+- :heavy_check_mark: All goodpractice linters passed.
+- :heavy_check_mark: R CMD check found no errors.
+- :heavy_check_mark: R CMD check found no warnings.
+- :eyes: Function names are duplicated in other packages
+
+(Checks marked with :eyes: may be optionally addressed.)
+
+
+Package License: MIT + file LICENSE
+
+---
+
+
+### 1. Package Dependencies
+
+<details>
+<summary>Details of Package Dependency Usage (click to open)</summary>
+<p>
+
+The table below tallies all function calls to all packages ('ncalls'), both internal (r-base + recommended, along with the package itself), and external (imported and suggested packages). 'NA' values indicate packages to which no identified calls to R functions could be found. Note that these results are generated by an automated code-tagging system which may not be entirely accurate.
+
+|type       |package          | ncalls|
+|:----------|:----------------|------:|
+|internal   |base             |   1815|
+|internal   |stats            |    113|
+|internal   |graphics         |     37|
+|internal   |grDevices        |     35|
+|internal   |camtrapReport    |     34|
+|internal   |utils            |     27|
+|internal   |Matrix           |     14|
+|internal   |mgcv             |      8|
+|internal   |tools            |      4|
+|imports    |methods          |     30|
+|imports    |terra            |     25|
+|imports    |dplyr            |     23|
+|imports    |glue             |      5|
+|imports    |spatstat.geom    |      4|
+|imports    |lubridate        |      3|
+|imports    |spatstat.explore |      2|
+|imports    |data.table       |      1|
+|imports    |jsonlite         |      1|
+|suggests   |shiny            |      5|
+|suggests   |rmarkdown        |      4|
+|suggests   |camtrapDensity   |      2|
+|suggests   |activity         |      1|
+|suggests   |knitr            |      1|
+|suggests   |sbd              |      1|
+|suggests   |sf               |      1|
+|suggests   |suncalc          |      1|
+|suggests   |xts              |      1|
+|suggests   |camtraptor       |     NA|
+|suggests   |corrplot         |     NA|
+|suggests   |Distance         |     NA|
+|suggests   |dygraphs         |     NA|
+|suggests   |ggplot2          |     NA|
+|suggests   |ggrepel          |     NA|
+|suggests   |gt               |     NA|
+|suggests   |htmltools        |     NA|
+|suggests   |htmlwidgets      |     NA|
+|suggests   |iNEXT            |     NA|
[…]
