+++ NEWS.md
@@ -1,30 +1,138 @@
-# camtrapReport 1.0.47 (2026-08-09)
+# camtrapReport 1.0.47
 
-## Improvements
+## rOpenSci review preparation
 
-- Prepared the package for rOpenSci software peer review.
-- Improved package structure, namespace management, and internal code quality.
-- Strengthened automated testing and package-check workflows.
-- Improved robustness of report generation and supporting utilities.
-- Replaced the former S4 `install_All()` implementation with the simpler
-  snake-case `install_all()` function. It automatically discovers dependencies
-  declared by bundled and registered user-managed report modules, and delegates
-  dependency resolution and installation to `pak`.
+This release prepares `camtrapReport` for rOpenSci software review and for a possible Methods in Ecology and Evolution software-article route.
 
-## Documentation
+`camtrapReport` provides a reproducible workflow for processing camera-trap data in Camtrap DP format and generating data-status and ecological reports for wildlife monitoring, biodiversity assessment, and conservation decision support.
 
-- Expanded and revised the package documentation to clarify the scope, workflow, and differences from existing camera-trap tools.
-- Improved the pkgdown website, reference documentation, vignettes, and reporting resources.
-- Standardised package metadata, citations, links, and release information.
+## Package quality and review readiness
 
-## Maintenance
+- Prepared the package for rOpenSci review using `pkgcheck`.
+- Added a `ropensci-pkgcheck-report.md` file documenting the package-review status.
+- Confirmed that `R CMD check` reports no errors and no warnings.
+- Confirmed that GitHub Actions continuous integration is available and passing.
+- Confirmed package test coverage of approximately 78%.
+- Confirmed that all exported functions have examples.
+- Confirmed that the package website is available through `pkgdown`.
+- Added or revised GitHub issue templates to support clearer bug reports, feature requests, and user feedback.
+- Added or revised community files, including contributing guidance and code-of-conduct information.
 
-- Removed obsolete and unused package assets and temporary files.
-- Resolved non-ASCII source-code issues and other package-check findings.
-- Updated CRAN package links to their canonical form.
+## Documentation and website
 
-This release primarily prepares camtrapReport for rOpenSci software peer review and improves package robustness, documentation, reproducibility, and maintainability.
+- Revised the README to clarify the package scope, intended users, main workflows, and relationship to Camtrap DP.
+- Expanded the package overview to better explain how `camtrapReport` supports reproducible camera-trap data reporting.
+- Improved the pkgdown website structure, including the Get started page, Articles, Reference, Resources, GitHub links, discussion links, and issue-reporting links.
+- Added and revised article pages for:
+  - Data Status Report
+  - Ecological Report
+  - Package Overview
+  - Module Management
+  - Resources
+- Added example Data Status Report and Ecological Report galleries to the website.
+- Added links to full example HTML reports through the website.
+- Added a Resources page listing publications, workshops, and training activities related to `camtrapReport`.
+- Improved image paths and static assets used by pkgdown articles.
+- Regenerated the pkgdown website for the review version.
 
-# camtrapReport 1.0.45 (2026-08-05)
+## Example reports and example data
 
-- Initial official release.
+- Added example HTML reports to demonstrate package outputs.
+- Added example Data Status Reports for multiple camera-trap monitoring datasets.
+- Added example Ecological Reports for multiple monitoring projects.
+- Improved documentation around the bundled Leuven example dataset.
+- Documented the provenance, intended use, licence, and limitations of the example dataset.
+- Added guidance that example reports are intended to demonstrate reporting structure and functionality, not to replace dataset-specific ecological interpretation.
+
+## Data Status Report
+
+- Improved documentation of the Data Status Report workflow.
+- Clarified how the report can be used to assess dataset completeness, possible data issues, and readiness for ecological reporting.
[…]
