### 2026-08-26 fb67923
Address dependency handling for rOpenSci review
