+++ .github/CODE_OF_CONDUCT.md
@@ -1,132 +1,127 @@
-# Contributing to camtrapReport
+# Contributor Covenant Code of Conduct
 
-Thank you for your interest in contributing to `camtrapReport`. Bug reports, feature suggestions, documentation improvements, and new report modules are welcome.
+## Our Pledge
 
-## Reporting problems
+We as members, contributors, and leaders pledge to make participation in our
+community a harassment-free experience for everyone, regardless of age, body
+size, visible or invisible disability, ethnicity, sex characteristics, gender
+identity and expression, level of experience, education, socio-economic status,
+nationality, personal appearance, race, caste, color, religion, or sexual
+identity and orientation.
 
-Please report bugs through the [GitHub issue tracker](https://github.com/spatialecology/camtrapReport/issues).
+We pledge to act and interact in ways that contribute to an open, welcoming,
+diverse, inclusive, and healthy community.
 
-To help us investigate, please include:
+## Our Standards
 
-* a clear description of the problem;
-* the result you expected;
-* a minimal reproducible example;
-* the output of `sessionInfo()`; and
-* a small synthetic or openly shareable Camtrap DP dataset, where possible.
+Examples of behavior that contributes to a positive environment for our
+community include:
 
-If you cannot share the data because of privacy, confidentiality, or sensitivity concerns, please contact the maintainer [by email](mailto:eebrahimi.bio@gmail.com) to discuss an appropriate alternative.
+* Demonstrating empathy and kindness toward other people
+* Being respectful of differing opinions, viewpoints, and experiences
+* Giving and gracefully accepting constructive feedback
+* Accepting responsibility and apologizing to those affected by our mistakes,
+  and learning from the experience
+* Focusing on what is best not just for us as individuals, but for the overall
+  community
 
-Do not upload confidential camera-trap data, sensitive species locations, restricted images, personal information, credentials, or other protected material to a public GitHub issue.
+Examples of unacceptable behavior include:
 
-## Suggesting features
+* The use of sexualized language or imagery, and sexual attention or advances of
+  any kind
+* Trolling, insulting or derogatory comments, and personal or political attacks
+* Public or private harassment
+* Publishing others' private information, such as a physical or email address,
+  without their explicit permission
+* Other conduct which could reasonably be considered inappropriate in a
+  professional setting
 
-Feature requests and proposals for new ecological-report modules are welcome.
+## Enforcement Responsibilities
 
-For substantial changes, please open an issue first so that we can discuss how the proposal fits the package structure and existing workflow.
+Community leaders are responsible for clarifying and enforcing our standards of
+acceptable behavior and will take appropriate and fair corrective action in
+response to any behavior that they deem inappropriate, threatening, offensive,
+or harmful.
 
-## Pull requests
+Community leaders have the right and responsibility to remove, edit, or reject
+comments, commits, code, wiki edits, issues, and other contributions that are
+not aligned to this Code of Conduct, and will communicate reasons for moderation
+decisions when appropriate.
 
-Before submitting a pull request:
+## Scope
 
-1. Keep the proposed change focused.
-2. Add or update tests when behaviour changes.
-3. Update the roxygen2 documentation when necessary.
-4. Add an entry to [`NEWS.md`](https://github.com/spatialecology/camtrapReport/blob/main/NEWS.md) for user-facing changes.
-5. Install the package dependencies and run the development checks:
+This Code of Conduct applies within all community spaces, and also applies when
+an individual is officially representing the community in public spaces.
+Examples of representing our community include using an official e-mail address,
+posting via an official social media account, or acting as an appointed
+representative at an online or offline event.
 
-```r
-pak::pak("deps::.")
+## Enforcement
 
-devtools::document()
-devtools::test()
-devtools::check()
-```
[…]
