### 2026-08-26 4e2c9f7
Address dependency handling for rOpenSci review
