### 2026-08-26 c23036c
Fix report package-loader chunks
