+++ R/testSection.R
@@ -32,8 +32,7 @@
   # Reuse the same package-loader helper used by generateReport()
   .env$pkg_chunk <- .make_package_loader_chunk(
     pkgs = module_pkgs,
-    core = "knitr",
-    label = "camtrap-quick-test-packages"
+    core = "knitr"
   )
   
   rmd_template <- glue::glue(
@@ -107,8 +106,7 @@ output:
   # Reuse the same package-loader helper used by generateReport()
   .env$pkg_chunk <- .make_package_loader_chunk(
     pkgs = module_pkgs,
-    core = "knitr",
-    label = "camtrap-section-test-packages"
+    core = "knitr"
   )
   
   rmd_template <- glue::glue(
