### 2026-08-26 11462e0
Fix report rendering and regenerate pkgdown site
