### 2026-08-31 d480583
Update ecological report and rebuild pkgdown site
