+++ README.md
@@ -19,135 +19,80 @@ camtrapReport
 <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project status: active"></a>
 <a href="https://www.r-project.org/"><img src="https://img.shields.io/badge/R-%E2%89%A5%204.1.0-276DC3?logo=r&logoColor=white" alt="R version 4.1.0 or later"></a>
 <a href="https://github.com/spatialecology/camtrapReport/blob/main/LICENSE.md"><img src="https://img.shields.io/badge/license-MIT-yellow.svg" alt="MIT license"></a>
-<a href="https://spatialecology.github.io/camtrapReport/"><img src="https://img.shields.io/badge/docs-pkgdown-orange.svg" alt="pkgdown documentation"></a>
 </p>
-
-------------------------------------------------------------------------
 
 `camtrapReport` converts camera-trap data in [Camtrap
-DP](https://camtrap-dp.tdwg.org/) format into two coordinated HTML
+DP](https://camtrap-dp.tdwg.org/) format into two configurable HTML
 outputs:
 
-- a **Data Status Check**, which identifies issues in data completeness,
-  consistency, annotation, and spatiotemporal coverage; and
-- an **Ecological Report**, which combines selected ecological
-  summaries, analyses, figures, tables, maps, metadata, and explanatory
-  text in a configurable report.
+- a **Data Status Check** for data completeness, consistency,
+  annotation, and spatiotemporal coverage; and
+- an **Ecological Report** containing selected summaries, analyses,
+  figures, maps, tables, metadata, and explanatory text.
 
-The package focuses on the reporting workflow. It complements tools for
-camera-trap data management and specialised ecological modelling rather
-than replacing them.
-
-<p align="center">
-
-<img src="vignettes/figures/package-architecture.png"
-       width="900"
-       alt="Workflow from Camtrap DP input through the camReport object to the Data Status Check and Ecological Report"><br>
-<em>Overview of the camtrapReport workflow.</em>
-</p>
-
-## Use in practice
-
-Although `camtrapReport` has not yet been applied in a published
-ecological study, it has been developed and evaluated using real-world
-camera-trap monitoring datasets from multiple study sites across North
-America, Asia, Europe, Africa, and Australia. Evaluation involved both
-developer-led testing and independent users who applied the package to
-their own datasets and provided feedback. In Europe, this evaluation has
-drawn particularly on datasets available through camera-trap research
-networks, including the [European Observatory of Wildlife
-network](https://enetwild.com/), which spans diverse geographic settings
-and monitoring designs.
-
-Experience across these heterogeneous datasets has informed checks for
-metadata completeness and spatiotemporal coverage, improvements to
-report generation, and revisions based on user feedback. Related
-publications, presentations, and training activities are listed on the
-[Resources
-page](https://spatialecology.github.io/camtrapReport/articles/resources.html).
+<figure>
+<img src="man/figures/package-architecture.png"
+alt="Overview of the camtrapReport workflow." />
+<figcaption aria-hidden="true">Overview of the camtrapReport
+workflow.</figcaption>
+</figure>
 
 ## Installation
 
-`camtrapReport` requires R 4.1.0 or later. Install it from GitHub with
-[`pak`](https://pak.r-lib.org/):
+Install the development version from GitHub:
 
 ``` r
 if (!requireNamespace("pak", quietly = TRUE)) {
   install.packages("pak")
 }
 
-pak::pkg_install(
-  "spatialecology/camtrapReport"
-)
[…]
