+++ CITATION.cff
@@ -1,3 +1,4 @@
+# ------------------------------------------------
 # CITATION.cff file for camtrapReport
 # ------------------------------------------------
 
@@ -6,7 +7,7 @@ message: 'To cite package "camtrapReport" in publications, please use:'
 type: software
 license: MIT
 title: 'camtrapReport: Camera-Trap Report Generator'
-version: '1.0.47'
+version: '1.0.48'
 
 abstract: >-
   Provides an extensible workflow for processing camera-trap data in the
