+++ _pkgdown.yml
@@ -159,3 +159,4 @@ reference:
 
   - contents:
       - camReport-classes
+      

[…]
