+++ man/report.Rd
@@ -58,11 +58,8 @@ example_dataset <- system.file(
 )
 
 cm <- camData(example_dataset)
-
-# Use one lightweight section for the ecological report
 cm <- sections(cm, "introduction")
 
-# Generate an ecological report in a temporary directory
 report_stem <- tempfile("camtrapReport-report-")
 report_file <- report(
   cm,
@@ -71,24 +68,6 @@ report_file <- report(
 )
 
 file.exists(report_file)
-
-# Generate a data-status report in a temporary directory
-status_stem <- tempfile("camtrapReport-status-")
-status_file <- status(
-  cm,
-  filename = status_stem,
-  view = FALSE
-)
-
-file.exists(status_file)
-
-# Remove files generated by the examples
-unlink(c(
-  report_file,
-  paste0(report_stem, ".Rmd"),
-  status_file,
-  paste0(status_stem, ".Rmd")
-))
 \dontshow{\}) # examplesIf}
 }
 \seealso{
