### 2026-08-31 6a71c76
ropenSci adoptation
