### 2026-08-31 e975a22
Restore last passing package version
