+++ tests/testthat/test-modules.R
@@ -5,17 +5,17 @@ test_that("module tree helpers preserve parent-child ordering", {
     stringsAsFactors = FALSE
   )
 
-  expect_identical(camtrapReport:::.subtree_end(info, "methods"), 2L)
-  expect_identical(camtrapReport:::.get_descendants(info, "methods"), "sampling")
+  expect_identical(ct_internal(".subtree_end")(info, "methods"), 2L)
+  expect_identical(ct_internal(".get_descendants")(info, "methods"), "sampling")
   expect_identical(
-    camtrapReport:::.ancestor_chain(
+    ct_internal(".ancestor_chain")(
       "sampling",
       stats::setNames(info$parent, info$name)
     ),
     "methods"
   )
 
-  inserted <- camtrapReport:::.insert_module_info(
+  inserted <- ct_internal(".insert_module_info")(
     info,
     name = "effort",
     parent = "methods",
@@ -23,7 +23,7 @@ test_that("module tree helpers preserve parent-child ordering", {
   )
   expect_identical(inserted$name, c("methods", "sampling", "effort", "results"))
 
-  before <- camtrapReport:::.insert_module_info(
+  before <- ct_internal(".insert_module_info")(
     inserted,
     name = "locations",
     parent = "methods",
@@ -31,10 +31,15 @@ test_that("module tree helpers preserve parent-child ordering", {
   )
   expect_identical(before$name[2], "locations")
   expect_error(
-    camtrapReport:::.insert_module_info(before, "bad", before = "methods", after = "results"),
+    ct_internal(".insert_module_info")(
+      before,
+      "bad",
+      before = "methods",
+      after = "results"
+    ),
     "only one"
   )
-  expect_error(camtrapReport:::.subtree_end(info, "unknown"), "Unknown parent")
+  expect_error(ct_internal(".subtree_end")(info, "unknown"), "Unknown parent")
 })
 
 test_that("the public module listing supports tree and table views", {
@@ -47,30 +52,31 @@ test_that("the public module listing supports tree and table views", {
   expect_true(is.data.frame(full) || is.list(full))
 })
 
-test_that("modules can be added, deleted, restored, and purged in a temporary library", {
+test_that(
+  paste0(
+    "modules can be added, deleted, restored, and purged ",
+    "in a temporary library"
+  ),
+  {
   module_dir <- copy_camtrap_module_library()
   parent_file <- write_test_module(
-    .register_camtrap_test_path(
-      tempfile("test_parent_", fileext = ".yml")
-    ),
+    tempfile("test_parent_", fileext = ".yml"),
     "test_parent",
     parent = "results"
   )
   child_file <- write_test_module(
-    .register_camtrap_test_path(
-      tempfile("test_child_", fileext = ".yml")
-    ),
+    tempfile("test_child_", fileext = ".yml"),
     "test_child",
     parent = "test_parent"
   )
 
-  parent_added <- camtrapReport:::.add_Module(
+  parent_added <- ct_internal(".add_Module")(
     parent_file,
     after = "captures",
     dir = module_dir,
     test = FALSE
   )
-  child_added <- camtrapReport:::.add_Module(
+  child_added <- ct_internal(".add_Module")(
     child_file,
     dir = module_dir,
     test = FALSE
@@ -80,40 +86,44 @@ test_that("modules can be added, deleted, restored, and purged in a temporary li
   expect_true(file.exists(child_added$file))
   expect_identical(child_added$module@parent, "test_parent")
 
-  listed <- camtrapReport:::.list_Modules(dir = module_dir, validate = TRUE)
[…]
