+++ R/camtrapReport-package.R
@@ -19,15 +19,20 @@
 #' * Report bugs at <https://github.com/spatialecology/camtrapReport/issues>
 #'
 #' @keywords internal
-#' @import methods
-#' @importFrom graphics abline legend par points lines boxplot barplot arrows segments text axis image polygon
+#' @importFrom methods new setClass setGeneric setMethod
+#' @importFrom methods setReplaceMethod signature
+#' @importFrom graphics abline legend par points lines boxplot
+#' @importFrom graphics barplot arrows segments text axis image polygon
 #' @importFrom grDevices colorRampPalette
 #' @importFrom stats cor cov density quantile na.omit predict setNames time
-#' @importFrom utils read.csv install.packages unzip head remove.packages tail timestamp
+#' @importFrom utils read.csv install.packages unzip head
+#' @importFrom utils remove.packages tail timestamp
 #' @importFrom glue glue
 #' @importFrom data.table data.table
-#' @importFrom dplyr bind_rows distinct filter group_by mutate select slice left_join
-#' @importFrom terra hull crs expanse extract project rast readRDS unwrap vect wrap writeRaster
+#' @importFrom dplyr bind_rows distinct filter group_by mutate
+#' @importFrom dplyr select slice left_join
+#' @importFrom terra hull crs expanse extract project rast
+#' @importFrom terra readRDS unwrap vect wrap writeRaster
 #' @importFrom lubridate interval
 #' @export add_Module
 #' @export camData
@@ -36,7 +41,7 @@
 #' @export gui
 #' @export info
 #' @export info<-
-#' @export install_All
+#' @export install_all
 #' @export list_Modules
 #' @export listReportSections
 #' @export move_Module
@@ -50,4 +55,6 @@
 #' @export testSection
 #' @export updateReportSection
 #' @exportClass camReport
+#' @importFrom spatstat.geom owin ppp
+#' @importFrom spatstat.explore quadrat.test Kest
 "_PACKAGE"
