+++ tests/testthat/test-report-object-branches.R
@@ -76,8 +76,9 @@ test_that("root and nested report sections can be replaced", {
     
     tree <- api$objects()
     
-    expect_true(
-      is.list(tree$root)
+    expect_type(
+      tree$root,
+      "list"
     )
     
     expect_true(
@@ -94,7 +95,7 @@ test_that("root and nested report sections can be replaced", {
       "Replacement root text."
     )
     
-    expect_equal(
+    expect_identical(
       tree$root$root@headLevel,
       1
     )
@@ -120,7 +121,7 @@ test_that("root and nested report sections can be replaced", {
       "Replacement child text."
     )
     
-    expect_equal(
+    expect_identical(
       tree$root$child@headLevel,
       2
     )
@@ -170,12 +171,14 @@ test_that("third-level report sections can be added and replaced", {
     
     tree <- api$objects()
     
-    expect_true(
-      is.list(tree$root)
+    expect_type(
+      tree$root,
+      "list"
     )
     
-    expect_true(
-      is.list(tree$root$child)
+    expect_type(
+      tree$root$child,
+      "list"
     )
     
     expect_true(
@@ -192,7 +195,7 @@ test_that("third-level report sections can be added and replaced", {
       "Replacement grandchild text."
     )
     
-    expect_equal(
+    expect_identical(
       tree$root$child$grandchild@headLevel,
       3
     )
@@ -200,7 +203,12 @@ test_that("third-level report sections can be added and replaced", {
 })
 
 
-test_that("root report chunks can be added, converted to a list, and replaced", {
+test_that(
+  paste0(
+    "root report chunks can be added, converted to a list, ",
+    "and replaced"
+  ),
+  {
   for (status in c(FALSE, TRUE)) {
     api <- new_report_object_test_api(status)
     add <- api$add
@@ -264,8 +272,9 @@ test_that("root report chunks can be added, converted to a list, and replaced", 
     tree <- api$objects()
     chunks <- tree$root@Rchunk
     
-    expect_true(
-      is.list(chunks)
+    expect_type(
+      chunks,
+      "list"
     )
     
     expect_true(
@@ -291,8 +300,9 @@ test_that("root report chunks can be added, converted to a list, and replaced", 
     tree <- api$objects()
     chunks <- tree$root@Rchunk
     
-    expect_true(
-      is.list(chunks)
+    expect_type(
+      chunks,
+      "list"
     )
     
     expect_true(
@@ -375,8 +385,9 @@ test_that("second-level report chunks can be added and replaced", {
     tree <- api$objects()
     chunks <- tree$root$child@Rchunk
     
-    expect_true(
-      is.list(chunks)
+    expect_type(
+      chunks,
+      "list"
     )
     
     expect_true(
@@ -460,8 +471,9 @@ test_that("third-level report chunks can be added and replaced", {
     tree <- api$objects()
     chunks <- tree$root$child$grandchild@Rchunk
     
-    expect_true(
-      is.list(chunks)
+    expect_type(
+      chunks,
+      "list"
     )
     
     expect_true(
@@ -522,8 +534,8 @@ test_that("unsupported objects and chunks with unknown parents are ignored", {
     
     tree <- api$objects()
     
-    expect_identical(
-      names(tree),
+    expect_named(
+      tree,
       "root"
     )
     
