+++ tests/testthat/test-camR-render-helpers.R
@@ -1,5 +1,5 @@
 test_that("package normalization handles vectors and comma-separated entries", {
-  normalize_packages <- ct_internal(".normalize_packages")
+  normalize_packages <- camtrapReport:::.normalize_packages
   
   expect_identical(
     normalize_packages(NULL),
@@ -27,7 +27,7 @@ test_that("package normalization handles vectors and comma-separated entries", {
 
 
 test_that("module package collection handles single report chunks", {
-  collect_packages <- ct_internal(".collect_module_packages")
+  collect_packages <- camtrapReport:::.collect_module_packages
   
   section <- reportSection(
     name = "summary",
@@ -48,7 +48,7 @@ test_that("module package collection handles single report chunks", {
 
 
 test_that("module package collection handles nested sections", {
-  collect_packages <- ct_internal(".collect_module_packages")
+  collect_packages <- camtrapReport:::.collect_module_packages
   
   first <- reportSection(
     name = "first",
@@ -84,7 +84,7 @@ test_that("module package collection handles nested sections", {
 
 
 test_that("module package collection handles lists of chunks", {
-  collect_packages <- ct_internal(".collect_module_packages")
+  collect_packages <- camtrapReport:::.collect_module_packages
   
   section <- reportSection(
     name = "multi_chunk",
@@ -124,7 +124,7 @@ test_that("module package collection handles lists of chunks", {
 
 
 test_that("module package collection handles empty structures", {
-  collect_packages <- ct_internal(".collect_module_packages")
+  collect_packages <- camtrapReport:::.collect_module_packages
   
   expect_identical(
     collect_packages(NULL),
@@ -148,20 +148,19 @@ test_that("module package collection handles empty structures", {
 })
 
 
-test_that("package loader creates an empty uniquely labelled chunk", {
-  make_loader <- ct_internal(".make_package_loader_chunk")
+test_that("package loader creates an empty setup chunk", {
+  make_loader <- camtrapReport:::.make_package_loader_chunk
   
   loader <- make_loader(
     pkgs = NULL,
     core = character(),
-    attach = TRUE,
-    label = "test-packages"
+    attach = TRUE
   )
   
   expect_identical(
     loader,
     paste0(
-      "```{r test-packages, include=FALSE}\n",
+      "```{r setup, include=FALSE}\n",
       "# no extra packages\n",
       "```\n"
     )
@@ -170,7 +169,7 @@ test_that("package loader creates an empty uniquely labelled chunk", {
 
 
 test_that("package loader normalizes and deduplicates packages", {
-  make_loader <- ct_internal(".make_package_loader_chunk")
+  make_loader <- camtrapReport:::.make_package_loader_chunk
   
   loader <- make_loader(
     pkgs = c(
@@ -179,8 +178,7 @@ test_that("package loader normalizes and deduplicates packages", {
       "stats"
     ),
     core = c("knitr", " methods "),
-    attach = TRUE,
-    label = "test-packages"
+    attach = TRUE
   )
   
   expect_match(
@@ -206,7 +204,7 @@ test_that("package loader normalizes and deduplicates packages", {
     fixed = TRUE
   )
   
-  expect_identical(
+  expect_equal(
     lengths(
       regmatches(
         loader,
@@ -223,7 +221,7 @@ test_that("package loader normalizes and deduplicates packages", {
 
 
 test_that("package loader can check packages without attaching them", {
-  make_loader <- ct_internal(".make_package_loader_chunk")
+  make_loader <- camtrapReport:::.make_package_loader_chunk
   
   loader <- make_loader(
     pkgs = c("stats", "methods"),
[…]

[…]
