### 2026-08-31 8cf2a1d
update manuals and the website
