+++ man/camtrapReport-package.Rd
@@ -29,5 +29,10 @@ Useful links:
 \author{
 \strong{Maintainer}: Elham Ebrahimi \email{eebrahimi.bio@gmail.com} (\href{https://orcid.org/0000-0001-5191-9832}{ORCID})
 
+Authors:
+\itemize{
+  \item Elham Ebrahimi \email{eebrahimi.bio@gmail.com} (\href{https://orcid.org/0000-0001-5191-9832}{ORCID})
+}
+
 }
 \keyword{internal}
