+++ man/report.Rd
@@ -80,22 +80,10 @@ report_file <- report(
 
 file.exists(report_file)
 
-# Generate a data-status report in a temporary directory
-status_stem <- tempfile("camtrapReport-status-")
-status_file <- status(
-  cm,
-  filename = status_stem,
-  view = FALSE
-)
-
-file.exists(status_file)
-
 # Remove files generated by the examples
 unlink(c(
   report_file,
   paste0(report_stem, ".Rmd"),
-  status_file,
-  paste0(status_stem, ".Rmd"),
   example_dataset
 ), recursive = TRUE, force = TRUE)
 }
@@ -105,7 +93,7 @@ unlink(c(
 \code{\link[=camData]{camData()}}, \code{\link[=reportSection]{reportSection()}}, \code{\link[=updateReportSection]{updateReportSection()}},
 \code{\link[=section_names]{section_names()}}
 
-Other report generation: 
+Other report generation:
 \code{\link{gui,camReport-method}}
 }
 \concept{report generation}
