+++ man/testSection.Rd
@@ -33,7 +33,7 @@ The \code{object} argument is only needed when the R code in the section refers 
 a \code{camReport} object.
 }
 \examples{
-\dontshow{if (rmarkdown::pandoc_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (rmarkdown::pandoc_available()) withAutoprint(\{ # examplesIf}
 tx <- reportSection(
   name = "introduction",
   title = "Introduction",
@@ -59,9 +59,9 @@ unlink(test_file)
 \seealso{
 \code{\link[=camData]{camData()}}, \code{\link[=reportSection]{reportSection()}}, \code{\link[=updateReportSection]{updateReportSection()}}, \code{\link[=report]{report()}}
 
-Other report sections: 
+Other report sections:
 \code{\link{reportSection,character-method}},
-\code{\link{section_names}()},
-\code{\link{updateReportSection}()}
+\code{\link[=section_names]{section_names()}},
+\code{\link[=updateReportSection]{updateReportSection()}}
 }
 \concept{report sections}
