+++ man/section_names.Rd
@@ -82,9 +82,9 @@ unlink(example_dataset, recursive = TRUE, force = TRUE)
 \code{\link[=camData]{camData()}}, \code{\link[=report]{report()}}, \code{\link[=updateReportSection]{updateReportSection()}},
 \code{\link[=listReportSections]{listReportSections()}}
 
-Other report sections: 
+Other report sections:
 \code{\link{reportSection,character-method}},
 \code{\link{testSection,.textSection-method}},
-\code{\link{updateReportSection}()}
+\code{\link[=updateReportSection]{updateReportSection()}}
 }
 \concept{report sections}
