+++ R/camData.R
@@ -81,38 +81,10 @@
 }
 
 #--------
-
-.first_non_missing <- function(x) {
-  x <- unique(x[!is.na(x)])
-  if (length(x) == 0) return(NA)
-  x[1]
-}
-
-#--------
-
-.safe_min_time <- function(x, tz = "UTC") {
-  x <- x[!is.na(x)]
-  if (length(x) == 0) {
-    return(as.POSIXct(NA_real_, origin = "1970-01-01", tz = tz))
-  }
-  as.POSIXct(min(x), tz = tz)
-}
-
-#--------
-
-.safe_max_time <- function(x, tz = "UTC") {
-  x <- x[!is.na(x)]
-  if (length(x) == 0) {
-    return(as.POSIXct(NA_real_, origin = "1970-01-01", tz = tz))
-  }
-  as.POSIXct(max(x), tz = tz)
-}
-
-#--------
-
-
 .getSequences <- function(media) {
-  if (!.require('data.table')) stop('The data.table package is not installed...!')
+  if (!.require("data.table")) {
+    stop("The data.table package is not installed...!")
+  }
   
   sequences <- .eval('media |> 
     dplyr::distinct() |> 
@@ -264,21 +236,37 @@
     if (!is.null(path) && is.character(path)) {
       .path <- file.path(
         path.expand(path),
-        gsub(basename(file), pattern = ".zip", replacement = "", ignore.case = TRUE)
+        gsub(
+          basename(file),
+          pattern = ".zip",
+          replacement = "",
+          ignore.case = TRUE
+        )
       )
     } else {
-      .path <- gsub(basename(file), pattern = ".zip", replacement = "", ignore.case = TRUE)
+      .path <- gsub(
+        basename(file),
+        pattern = ".zip",
+        replacement = "",
+        ignore.case = TRUE
+      )
     }
     
     file <- utils::unzip(file, exdir = .path)
     
   } else if (dir.exists(file)) {
-    if (all(c("datapackage.json", "deployments.csv", "observations.csv") %in% tolower(dir(file)))) {
+    core_files <- c(
+      "datapackage.json",
+      "deployments.csv",
+      "observations.csv"
+    )
+
+    if (all(core_files %in% tolower(dir(file)))) {
       .path <- file
       file <- dir(file, full.names = TRUE)
     } else {
-      if (any(c("datapackage.json", "deployments.csv", "observations.csv") %in% tolower(dir(file)))) {
-        required_files <- c("datapackage.json", "deployments.csv", "observations.csv", "media.csv")
+      if (any(core_files %in% tolower(dir(file)))) {
+        required_files <- c(core_files, "media.csv")
         .w <- !required_files %in% tolower(dir(file))
         
         stop(paste0(
@@ -287,7 +275,9 @@
           ") are not available in the specified folder."
         ))
       } else {
-        stop("The specified folder does not have the standard Camtrap DP files.")
+        stop(
+          "The specified folder does not have the standard Camtrap DP files."
+        )
       }
     }
   } else {
@@ -323,11 +313,17 @@
   }
   
   if ("deploymentStart" %in% names(.d$deployments)) {
-    .d$deployments$deploymentStart <- .parse_cam_datetime(.d$deployments$deploymentStart, tz = tz)
+    .d$deployments$deploymentStart <- .parse_cam_datetime(
+      .d$deployments$deploymentStart,
+      tz = tz
+    )
   }
   
   if ("deploymentEnd" %in% names(.d$deployments)) {
-    .d$deployments$deploymentEnd <- .parse_cam_datetime(.d$deployments$deploymentEnd, tz = tz)
+    .d$deployments$deploymentEnd <- .parse_cam_datetime(
[…]
