+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: camtrapReport
 Type: Package
 Title: Camera-Trap Report Generator
-Version: 1.0.51
+Version: 1.0.52
 Authors@R: c(
     person(
       given = "Elham",
