+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: camtrapReport
 Type: Package
 Title: Camera-Trap Report Generator
-Version: 1.0.52
+Version: 1.0.53
 Authors@R: c(
     person(
       given = "Elham",
