+++ tests/testthat/test-supporting-functions.R
@@ -1,10 +1,10 @@
 test_that("species and class helpers summarise the bundled data", {
   cm <- camtrap_test_report()
 
-  species_counts <- camtrapReport:::.get_species(cm$data, count = TRUE)
-  species_names <- camtrapReport:::.get_species(cm$data, count = FALSE)
-  class_counts <- camtrapReport:::.get_classes(cm$data, count = TRUE)
-  class_names <- camtrapReport:::.get_classes(cm$data, count = FALSE)
+  species_counts <- .get_species(cm$data, count = TRUE)
+  species_names <- .get_species(cm$data, count = FALSE)
+  class_counts <- .get_classes(cm$data, count = TRUE)
+  class_names <- .get_classes(cm$data, count = FALSE)
 
   expect_s3_class(species_counts, "data.frame")
   expect_true(all(species_names %in% species_counts$scientificName))
@@ -15,7 +15,7 @@ test_that("species and class helpers summarise the bundled data", {
 
 test_that("merged data can retain nested media records", {
   cm <- camtrap_test_report()
-  merged <- camtrapReport:::.merge_data(cm$data, dropMedia = FALSE)
+  merged <- .merge_data(cm$data, dropMedia = FALSE)
 
   expect_s3_class(merged, "data.frame")
   expect_true("media" %in% names(merged))
@@ -25,20 +25,20 @@ test_that("merged data can retain nested media records", {
 
 test_that("effort helpers aggregate deployments and draw with base graphics", {
   cm <- camtrap_test_report()
-  total <- camtrapReport:::.calc_effort(cm$data, unit = "day")
-  by_location <- camtrapReport:::.calc_effort(
+  total <- .calc_effort(cm$data, unit = "day")
+  by_location <- .calc_effort(
     cm$data,
     by = "locationID",
     unit = "hour"
   )
-  step_table <- camtrapReport:::.effort_table(cm, startend = TRUE)
+  step_table <- .effort_table(cm, startend = TRUE)
 
   expect_gt(total$effort, 0)
   expect_identical(total$unit, "day")
   expect_true(all(c("locationID", "effort") %in% names(by_location)))
   expect_true(all(c("time", "nrCams") %in% names(step_table)))
   expect_error(
-    camtrapReport:::.calc_effort(cm$data, by = "missing_column"),
+    .calc_effort(cm$data, by = "missing_column"),
     "Grouping columns"
   )
 
@@ -51,7 +51,7 @@ test_that("effort helpers aggregate deployments and draw with base graphics", {
     },
     add = TRUE
   )
-  expect_null(camtrapReport:::.plot_effort(cm, dynamic = FALSE))
+  expect_null(.plot_effort(cm, dynamic = FALSE))
 })
 
 test_that("base left join handles equal and differently named keys", {
@@ -59,13 +59,13 @@ test_that("base left join handles equal and differently named keys", {
   right <- data.frame(id = 2, extra = "x")
   right2 <- data.frame(other_id = 2, extra = "x")
 
-  joined <- camtrapReport:::.left_join(left, right, "id")
-  joined2 <- camtrapReport:::.left_join(left, right2, c("id", "other_id"))
+  joined <- .left_join(left, right, "id")
+  joined2 <- .left_join(left, right2, c("id", "other_id"))
 
   expect_identical(nrow(joined), 2L)
   expect_identical(nrow(joined2), 2L)
   expect_true(is.na(joined$extra[joined$id == 1]))
-  expect_error(camtrapReport:::.left_join(left, right, "missing"))
+  expect_error(.left_join(left, right, "missing"))
 })
 
 test_that("the internal pivot helper supports tidy-style column specifications", {
[…]
