### 2026-09-02 de01f46
Checkpoint before automated lint cleanup
