+++ tests/testthat/test-report-output-paths.R
@@ -1,25 +1,32 @@
 library(testthat)
 
 capture_report_conditions <- function(expr) {
-  messages <- character()
-  warnings <- character()
+  captured <- new.env(parent = emptyenv())
+  captured$messages <- character()
+  captured$warnings <- character()
   
   value <- withCallingHandlers(
     expr,
     message = function(condition) {
-      messages <<- c(messages, conditionMessage(condition))
+      captured$messages <- c(
+        captured$messages,
+        conditionMessage(condition)
+      )
       invokeRestart("muffleMessage")
     },
     warning = function(condition) {
-      warnings <<- c(warnings, conditionMessage(condition))
+      captured$warnings <- c(
+        captured$warnings,
+        conditionMessage(condition)
+      )
       invokeRestart("muffleWarning")
     }
   )
   
   list(
     value = value,
-    messages = messages,
-    warnings = warnings
+    messages = captured$messages,
+    warnings = captured$warnings
   )
 }
 
@@ -128,11 +135,12 @@ test_that("report does not invoke viewer after rendering failure", {
   dir.create(output_dir)
   
   old_viewer <- getOption("viewer")
-  viewer_called <- FALSE
+  viewer_state <- new.env(parent = emptyenv())
+  viewer_state$called <- FALSE
   
   options(
     viewer = function(path) {
-      viewer_called <<- TRUE
+      viewer_state$called <- TRUE
     }
   )
   
@@ -157,5 +165,5 @@ test_that("report does not invoke viewer after rendering failure", {
   )
   
   expect_s3_class(captured$value, "try-error")
-  expect_false(viewer_called)
+  expect_false(viewer_state$called)
 })
