+++ R/camR_functions.R
@@ -1043,7 +1043,10 @@
   } else {
     
     loc <- NULL
-    land_extract <- data.frame(name = character())
+    land_extract <- data.frame(
+      name = character(),
+      stringsAsFactors = FALSE
+    )
     sea_outlier_status <- paste0(w, " No valid locations available for land/sea check.")
   }
   
@@ -3042,10 +3045,9 @@
   }
   
   # Pattern for likely organizations / non-person entries
-  org_pattern <- paste(
+  org_pattern <- paste0(
     "university|universiteit|institute|institution|center|centre|research|admin|",
-    "observatory|consortium|network|project|laboratory|lab|group|team",
-    sep = ""
+    "observatory|consortium|network|project|laboratory|lab|group|team"
   )
   
   title_values <- as.character(contributors_df[["title"]])
