### 2026-09-03 6f0b1bd
Finalize version 1.0.56 for rOpenSci review
