+++ R/update.R
@@ -110,10 +110,10 @@
   if (is.call(expr) && identical(expr[[1]], as.name("{"))) {
     lines <- vapply(
       as.list(expr)[-1],
-      function(e) {
-        deparse1(e, collapse = "\n", width.cutoff = 60L)
-      },
-      character(1)
+      deparse1,
+      character(1),
+      collapse = "\n",
+      width.cutoff = 60L
     )
     return(paste(lines, collapse = "\n"))
   }
@@ -519,10 +519,7 @@ setMethod("updateReportSection",signature(x = "camReport"),
     if (missing(append_text)) append_text <- FALSE
     
     if (missing(section) || !is.character(section) || length(section) != 1L) {
-      stop(paste0(
-        "'section' should be a single character string, either a section ",
-        "name or title."
-      ))
+      stop("'section' should be a single character string, either a section ", "name or title.")
     }
     
     catalog <- .reportSection_catalog(x$reportObjects)
