+++ tests/testthat/test-report-object-branches.R
@@ -92,10 +92,7 @@ test_that("root and nested report sections can be replaced", {
       "Replacement root text."
     )
     
-    expect_equal(
-      tree$root$root@headLevel,
-      1
-    )
+    expect_identical(tree$root$root@headLevel, 1)
     
     replacement_child <- reportSection(
       name = "child",
@@ -118,10 +115,7 @@ test_that("root and nested report sections can be replaced", {
       "Replacement child text."
     )
     
-    expect_equal(
-      tree$root$child@headLevel,
-      2
-    )
+    expect_identical(tree$root$child@headLevel, 2)
   }
 })
 
@@ -186,10 +180,7 @@ test_that("third-level report sections can be added and replaced", {
       "Replacement grandchild text."
     )
     
-    expect_equal(
-      tree$root$child$grandchild@headLevel,
-      3
-    )
+    expect_identical(tree$root$child$grandchild@headLevel, 3)
   }
 })
 
