+++ man/testSection.Rd
@@ -25,7 +25,8 @@ the section text and code can be rendered successfully.
 }
 \details{
 This function is mainly used to test custom report sections before they are
-added to a full \code{\link[=camReport-classes]{camReport}} report. It checks whether the
+added to a full \code{\link[=camReport-classes]{camReport}} report. It checks whether
+the
 section text, R Markdown chunk settings, required packages, and R code can be
 rendered successfully.
 
