### 2026-09-04 b9937c4
Resolve actionable goodpractice findings
