### 2026-09-04 9169c0e
Address editor feedback and automatic checks
