+++ _pkgdown.yml
@@ -44,7 +44,7 @@ navbar:
         - text: Package Overview
           href: articles/Package-Overview.html
 
-        - text: Data Status Check
+        - text: Data Status Report
           href: articles/data-status-report.html
 
         - text: Ecological Report
@@ -67,17 +67,8 @@ navbar:
         - text: Authors
           href: authors.html
 
-        - text: Citation
-          href: authors.html#citation
-
         - text: License
           href: LICENSE.html
-
-        - text: Contributing guide
-          href: CONTRIBUTING.html
-
-        - text: Code of conduct
-          href: CODE_OF_CONDUCT.html
 
     support:
       text: Support
@@ -125,7 +116,7 @@ reference:
 
   - title: Report generation
     desc: >
-      Generate the Data Status Check or Ecological Report, or use the
+      Generate the Data Status Report or Ecological Report, or use the
       graphical interface.
 
   - contents:
