### 2026-09-04 a5e219e
Fix README dependency guidance
