### 2026-09-04 5196b63
Address rOpenSci editor feedback and automatic checks (#20)
Address rOpenSci editor feedback, resolve Jarl and goodpractice findings, add executed README examples, and rebuild the pkgdown documentation.
