+++ R/info.R
@@ -39,7 +39,8 @@ setGeneric(
 #'   the default report-information fields are returned.
 #' @param value The new value to assign to the specified field.
 #'
-#' @return `info()` returns an object of class `camInfo` containing the requested
+#' @return `info()` returns an object of class `camInfo` containing the
+#' requested
 #'   fields. The replacement method returns the updated `camReport` object
 #'   invisibly.
 #'
@@ -109,9 +110,7 @@ setMethod(
       name <- name[name %in% names(camR$fields())]
       
       if (length(name) == 0) {
-        warning(
-          "The specified name(s) are not identified or available in the camReport object; the default fields are used."
-        )
+        warning("The specified name(s) are not identified or available in the ", "camReport object; the default fields are used.")
         name <- .default_cam_info_names
       }
     }
@@ -190,9 +189,7 @@ setReplaceMethod(
       name <- name[name %in% names(.f)]
       
       if (length(name) == 0) {
-        stop(
-          "The specified name is not identified or available in the camReport object."
-        )
+        stop("The specified name is not identified or available in the ", "camReport object.")
       }
       
       x[[name]] <- value
