+++ tests/testthat/test-camR-render-helpers.R
@@ -204,8 +204,7 @@ test_that("package loader normalizes and deduplicates packages", {
     fixed = TRUE
   )
   
-  expect_equal(
-    lengths(
+  expect_identical(lengths(
       regmatches(
         loader,
         gregexpr(
@@ -214,9 +213,7 @@ test_that("package loader normalizes and deduplicates packages", {
           fixed = TRUE
         )
       )
-    ),
-    1L
-  )
+    ), 1L)
 })
 
 
