+++ R/camtrapReport-package.R
@@ -19,7 +19,8 @@
 #' * Report bugs at <https://github.com/spatialecology/camtrapReport/issues>
 #'
 #' @keywords internal
-#' @import methods
+#' @importFrom methods new representation setClass setClassUnion setGeneric
+#'   setMethod setOldClass setRefClass setReplaceMethod signature
 #' @importFrom graphics abline legend par points lines boxplot barplot arrows
 #'   segments text axis image polygon
 #' @importFrom grDevices colorRampPalette

[…]
