### 2026-09-04 70ca1a5
Condense and rebuild package documentation
