+++ vignettes/Package-Overview.Rmd
@@ -151,21 +151,10 @@ guide.
 
 ## Reproducibility, interpretation, and scope
 
-`camtrapReport` supports reproducible reporting by applying a consistent
-workflow from data checking through ecological analysis to report generation.
-Analytical modules use established ecological and statistical methods, with
-default settings that provide a common starting point across camera-trap
-datasets.
+`camtrapReport` applies a consistent workflow from data checking to report generation. To reproduce a report, retain the input data, selected modules, settings, and relevant package versions. Results should be interpreted in relation to the sampling design and assumptions of the selected methods.
 
-Because the workflow is modular, users can select relevant analyses and adjust
-module-specific settings, species groups, and thresholds according to their
-data, study design, and analytical objectives. Results should therefore be
-interpreted in relation to the characteristics of the dataset, sampling design,
-and assumptions of the selected methods.
+For broader guidance, see the [Reproducible Research CRAN Task View](https://cran.r-project.org/view=ReproducibleResearch).
 
-General guidance on recording computational environments and managing
-reproducible research workflows is available from the
-[Reproducible Research CRAN Task View](https://cran.r-project.org/view=ReproducibleResearch).
 
 ## Relationship to other camera-trap packages
 
