### 2026-09-04 ba6a4c7
Make fallback datetime parsing robust
