### 2026-09-04 773c03c
Remove trailing whitespace
