+++ R/camData.R
@@ -5,7 +5,7 @@
 .parse_cam_datetime <- function(x, tz = "UTC") {
   if (inherits(x, "POSIXct")) return(x)
   if (inherits(x, "POSIXt")) return(as.POSIXct(x, tz = tz))
-  
+
   if (is.null(x)) {
     return(
       as.POSIXct(
@@ -15,19 +15,19 @@
       )
     )
   }
-  
+
   x_chr <- trimws(as.character(x))
-  
+
   x_chr[
     x_chr %in% c("", "NA", "NaN", "NULL", "null")
   ] <- NA_character_
-  
+
   out <- as.POSIXct(
     rep(NA_real_, length(x_chr)),
     origin = "1970-01-01",
     tz = tz
   )
-  
+
   x_try <- x_chr
   x_try <- gsub("Z$", "+0000", x_try)
   x_try <- gsub(
@@ -35,7 +35,7 @@
     "\\1\\2",
     x_try
   )
-  
+
   formats <- c(
     "%Y-%m-%dT%H:%M:%OS%z",
     "%Y-%m-%d %H:%M:%OS%z",
@@ -54,14 +54,14 @@
     "%Y-%m-%d",
     "%Y/%m/%d"
   )
-  
+
   for (fmt in formats) {
     missing_i <- is.na(out) & !is.na(x_try)
-    
+
     if (!any(missing_i)) {
       break
     }
-    
+
     parsed <- suppressWarnings(
       as.POSIXct(
         x_try[missing_i],
@@ -69,14 +69,14 @@
         tz = tz
       )
     )
-    
+
     ok <- !is.na(parsed)
-    
+
     out[which(missing_i)[ok]] <- parsed[ok]
   }
-  
+
   missing_i <- is.na(out) & !is.na(x_chr)
-  
+
   if (any(missing_i) && .require("lubridate")) {
     fallback_orders <- c(
       "ymd HMS z", "ymd HMS",
@@ -86,16 +86,16 @@
       "ymdT HMS z", "ymdT HMS",
       "ymdT HM z",  "ymdT HM"
     )
-    
+
     # Trying the orders separately prevents lubridate from constructing
     # one excessively large regular expression.
     for (order_spec in fallback_orders) {
       missing_i <- is.na(out) & !is.na(x_chr)
-      
+
       if (!any(missing_i)) {
         break
       }
-      
+
       parsed <- tryCatch(
         suppressWarnings(
           lubridate::parse_date_time(
@@ -113,9 +113,9 @@
           )
         }
       )
-      
+
       ok <- !is.na(parsed)
-      
+
       if (any(ok)) {
         out[which(missing_i)[ok]] <- as.POSIXct(
           parsed[ok],
@@ -124,7 +124,7 @@
       }
     }
   }
-  
+
   out
 }
 
@@ -135,14 +135,14 @@
   if (!.require("data.table")) {
     stop("The data.table package is not installed...!")
   }
-  
-  sequences <- .eval('media |> 
-    dplyr::distinct() |> 
-    dplyr::select(deploymentID, sequenceID, timestamp, captureMethod) |> 
+
+  sequences <- .eval('media |>
+    dplyr::distinct() |>
+    dplyr::select(deploymentID, sequenceID, timestamp, captureMethod) |>
     data.table::data.table(key = "sequenceID")',environment())
-  
+
   sequences <- sequences[!is.na(sequences$sequenceID),]
-  
+
   # summarize per key
   sequences <- .eval("sequences[, list(deploymentID = unique(deploymentID),
                                 captureMethod = unique(captureMethod),
@@ -150,15 +150,15 @@
                                 end = max(timestamp),
                                 nrphotos = length(timestamp)),
[…]
