+++ README.Rmd
@@ -6,7 +6,7 @@ output:
 
 <!-- README.md is generated from README.Rmd. Edit README.Rmd, not README.md. -->
 
-```{r readme-setup, include=FALSE}
+```{r setup, include=FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
@@ -17,19 +17,22 @@ knitr::opts_chunk$set(
 options(width = 80)
 ```
 
+
 <h1 align="center">camtrapReport</h1>
 
 <p align="center">
   <em>Reproducible reporting for camera-trap monitoring data</em>
 </p>
 
-[![R CMD check status](https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml)
-[![rOpenSci package check status](https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml/badge.svg)](https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml)
-[![Test coverage](https://codecov.io/github/spatialecology/camtrapReport/graph/badge.svg?token=9VBXAR9XOD)](https://app.codecov.io/github/spatialecology/camtrapReport)
-[![Project status: active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![R version](https://img.shields.io/badge/R-%E2%89%A5%204.1.0-276DC3?logo=r&logoColor=white)](https://www.r-project.org/)
-[![MIT license](https://img.shields.io/badge/license-MIT-yellow.svg)](https://github.com/spatialecology/camtrapReport/blob/main/LICENSE.md)
-[![pkgdown documentation](https://img.shields.io/badge/docs-pkgdown-orange.svg)](https://spatialecology.github.io/camtrapReport/)
+<p align="center">
+  <a href="https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml"><img src="https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml/badge.svg" alt="R CMD check status"></a>
+  <a href="https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml"><img src="https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml/badge.svg" alt="rOpenSci package check status"></a>
+  <a href="https://app.codecov.io/github/spatialecology/camtrapReport"><img src="https://codecov.io/github/spatialecology/camtrapReport/graph/badge.svg?token=9VBXAR9XOD" alt="Test coverage"></a>
+  <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project status: active"></a>
+  <a href="https://www.r-project.org/"><img src="https://img.shields.io/badge/R-%E2%89%A5%204.1.0-276DC3?logo=r&logoColor=white" alt="R version 4.1.0 or later"></a>
+  <a href="https://github.com/spatialecology/camtrapReport/blob/main/LICENSE.md"><img src="https://img.shields.io/badge/license-MIT-yellow.svg" alt="MIT license"></a>
+  <a href="https://spatialecology.github.io/camtrapReport/"><img src="https://img.shields.io/badge/docs-pkgdown-orange.svg" alt="pkgdown documentation"></a>
+</p>
 
 ---
 
@@ -48,7 +51,7 @@ camera-trap data management and specialised ecological modelling rather than
 replacing them.
 
 <p align="center">
-  <img src="man/figures/package-architecture.png"
[…]
