+++ README.md
@@ -11,20 +11,16 @@ camtrapReport
 <em>Reproducible reporting for camera-trap monitoring data</em>
 </p>
 
-[![R CMD check
-status](https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml)
-[![rOpenSci package check
-status](https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml/badge.svg)](https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml)
-[![Test
-coverage](https://codecov.io/github/spatialecology/camtrapReport/graph/badge.svg?token=9VBXAR9XOD)](https://app.codecov.io/github/spatialecology/camtrapReport)
-[![Project status:
-active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![R
-version](https://img.shields.io/badge/R-%E2%89%A5%204.1.0-276DC3?logo=r&logoColor=white)](https://www.r-project.org/)
-[![MIT
-license](https://img.shields.io/badge/license-MIT-yellow.svg)](https://github.com/spatialecology/camtrapReport/blob/main/LICENSE.md)
-[![pkgdown
-documentation](https://img.shields.io/badge/docs-pkgdown-orange.svg)](https://spatialecology.github.io/camtrapReport/)
+<p align="center">
+
+<a href="https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml"><img src="https://github.com/spatialecology/camtrapReport/actions/workflows/R-CMD-check.yaml/badge.svg" alt="R CMD check status"></a>
+<a href="https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml"><img src="https://github.com/spatialecology/camtrapReport/actions/workflows/pkgcheck.yaml/badge.svg" alt="rOpenSci package check status"></a>
+<a href="https://app.codecov.io/github/spatialecology/camtrapReport"><img src="https://codecov.io/github/spatialecology/camtrapReport/graph/badge.svg?token=9VBXAR9XOD" alt="Test coverage"></a>
+<a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project status: active"></a>
+<a href="https://www.r-project.org/"><img src="https://img.shields.io/badge/R-%E2%89%A5%204.1.0-276DC3?logo=r&logoColor=white" alt="R version 4.1.0 or later"></a>
+<a href="https://github.com/spatialecology/camtrapReport/blob/main/LICENSE.md"><img src="https://img.shields.io/badge/license-MIT-yellow.svg" alt="MIT license"></a>
+<a href="https://spatialecology.github.io/camtrapReport/"><img src="https://img.shields.io/badge/docs-pkgdown-orange.svg" alt="pkgdown documentation"></a>
+</p>
 
 ------------------------------------------------------------------------
 
@@ -44,7 +40,7 @@ than replacing them.
 
 <p align="center">
 
-<img src="man/figures/package-architecture.png"
+<img src="vignettes/figures/package-architecture.png"
        width="900"
        alt="Workflow from Camtrap DP input through the camReport object to the Data Status Check and Ecological Report"><br>
 <em>Overview of the camtrapReport workflow.</em>
@@ -65,7 +61,10 @@ and monitoring designs.
 
 Experience across these heterogeneous datasets has informed checks for
 metadata completeness and spatiotemporal coverage, improvements to
-report generation, and revisions based on user feedback.
[…]
