### 2026-09-05 c72d912
Execute README quick-start examples
