### 2026-09-05 7519bcf
Document coding-assistance use
