### 2026-09-05 #799 @spatialecology
Hi @maelle,

Thank you for this detailed and constructive assessment. I have addressed each of the points raised below.

**MEE submission**
Yes. An earlier version of the package and its software manuscript received an initial editorial assessment at Methods in Ecology and Evolution. The editors recommended completing rOpenSci software peer review before resubmission. The manuscript has therefore not entered external peer review at MEE. The comment in the GitHub issue was informal feedback from colleagues who had read the editor’s email and wanted to help address it. It was not part of the formal MEE editorial assessment.

**Use of the package**
To the best of my knowledge, camtrapReport has not yet been used in a published ecological study. It has been developed and evaluated using real camera-trap datasets from study sites worldwide, including multiple datasets from North America, Asia, Europe, Africa, and Australia. Independent researchers have also applied the package to their own camera-trap datasets and provided feedback. I now state this explicitly in the [Use in practice section of the README](https://github.com/spatialecology/camtrapReport#use-in-practice), without presenting this evaluation as published use.

**Executed examples and getting started**
I revised the [README Quick start](https://github.com/spatialecology/camtrapReport#quick-start) and the [Get started page](https://spatialecology.github.io/camtrapReport/#quick-start) to provide a runnable example using a bundled, relationship-preserving subset of the [GMU8_LEUVEN Camera Trap Data Package dataset](https://www.gbif.org/dataset/c9cbc586-660e-4d89-ba14-0000c5770de1/download), together with bundled habitat data and a study-area boundary. All paths to the bundled package data are obtained using `system.file()` rather than placeholders.

The example is executed when the README is generated. It covers loading the data, creating the camReport object, and generating the Data Status Report and Ecological Report. Hidden validation checks confirm that both report files were created successfully, and the temporary example directory is removed afterward. The provenance and intended use of the example data are documented separately.

**Automatic checks**
I removed the duplicate definitions and addressed the actionable jarl and goodpractice findings, including replacing the broad methods import with explicit imports. Both jarl check and the duplicate-function check now pass.
I reviewed the remaining recommendations individually and made changes where these could be implemented safely. I retained the existing code where refactoring could affect dynamic YAML modules, optional dependencies, report generation, or backward compatibility. The remaining function-length, line-length, coverage, and stylistic messages are advisory rather than check failures.
All 959 tests pass. The current The current [GitHub Actions workflows](https://github.com/spatialecology/camtrapReport/actions) are also passing, including jarl, pkgcheck, test coverage, R CMD check across the five configured R and operating-system combinations, and the documentation build and deployment. R CMD check --as-cran completes with 0 errors, 0 warnings, and 0 notes.

**Dependencies**
Thank you for raising this. I reviewed the package dependency structure and documented the role of each dependency. I retained both data.table and dplyr. data.table provides a substantial advantage for efficiently reading and aggregating large Camtrap DP tables. Where equivalent functionality can be implemented clearly using base R, I generally prefer to avoid adding dependencies solely for convenience. However, I retained dplyr because it is already used within the package and is also commonly used in user-developed modules, which are part of the modular design of camtrapReport.

tidyr is not used as a core package dependency. The remaining suggested packages are used for documentation, testing, code coverage, or functionality that is only required in specific contexts. Packages required by individual report modules are not included as core package dependencies; these are handled separately, as described in my response below.

_**install_All()**_

I retained install_All() because it serves a package-specific purpose within the modular architecture of camtrapReport. Report modules can declare their own package dependencies, including modules developed and added by users. install_All() identifies these module-level dependencies and provides a convenient way to install any that are missing.
 The function is optional and is not called when camtrapReport is loaded or when a report is generated. I have also revised the installation procedure used by install_All() to use pak.

The architecture of camtrapReport is based on S4 classes and generics, together with reference classes, to support its modular design and associated functionality. For consistency with this architecture, the main package functions, including install_All(), are implemented as S4 generics.

No coding assistant was used to design the package architecture, modular system, scientific methodology, or core functionality. Coding-assistance tools were used only for specific supporting tasks during later stages of development. These uses are documented in the [Use of coding-assistance tools](https://github.com/spatialecology/camtrapReport/blob/main/.github/CONTRIBUTING.md#use-of-coding-assistance-tools) section of the contributing guide.

**Design**
Thank you for this suggestion. I removed discussion of Reference Classes from the README and Package Overview and revised the diagram to present the user-facing workflow without describing the central object as a container class. The camReport object is still shown because it is the object returned by camData() and passed through the public workflow.

I chose a Reference Class because the camReport object represents mutable state throughout the reporting workflow. It holds the imported data, metadata, user settings, selected report sections, and results generated or updated by individual modules. As the analysis proceeds, different functions and modules need to modify this shared state. Reference semantics allow these updates to be made in place while preserving a single consistent object, rather than requiring each step to reconstruct and return an updated copy of a potentially large and complex nested object.

In this sense, the camReport object also serves internally as the central container for the information required across the modular workflow. This is an implementation detail, however, rather than something users need to understand in order to use the package.

I have therefore moved this architectural rationale to the contributing guide, where it is more relevant to developers and potential contributors. Low-level class documentation remains available in the function reference but is no longer emphasised in the user-facing introductions.

**Interface**
Thank you for the suggestion. I considered changing the exported function names to snake_case, but I decided to retain the existing naming convention.

camtrapReport was developed using camelCase for its public API, and these function names are already used in existing analysis workflows, report modules, documentation, examples, and external scripts that use the package. Renaming the exported functions at this stage would therefore introduce a breaking API change and could affect the reproducibility of existing workflows.

I agree that snake_case is a common convention in contemporary R packages, but in this case I think maintaining the stability and continuity of the existing public API outweighs the benefit of changing the naming style. I have therefore retained the established function names and will keep the naming convention consistent within the package rather than introducing a parallel set of aliases.
Comparison with other packages
I condensed the comparison, removed the vague column, and revised the text to explain how `camtrapReport` complements specific camera-trap packages rather than comparing it with an undefined group of tools. The revised comparison is available in the [Relationship to other camera-trap packages](https://github.com/spatialecology/camtrapReport#relationship-to-other-camera-trap-packages) section of the README.

**Reproducibility**
I removed the discussion of `sessionInfo()` and shortened this section to focus on guidance specific to reproducing and interpreting `camtrapReport` outputs. For broader guidance on reproducible workflows in R, the section now links to the [Reproducible Research CRAN Task View](https://cran.r-project.org/web/views/ReproducibleResearch.html).

**Documentation**
I removed the repetitive section from the Package Overview and condensed the remaining content. I added @keywords internal to the package-level documentation and moved the camReport class reference to the final section of the reference index, after the user-facing functions.
Public materials are not available for the listed workshops, so the [Resources page](https://spatialecology.github.io/camtrapReport/articles/resources.html) now states this and presents them as records of completed events rather than as inactive links. I also removed the separate index.md; pkgdown now builds the landing page from the README, including the author, GitHub and support links, installation instructions, and worked example.
I also split internal utilities by responsibility. Spatial and plotting helpers are in R/utils-spatial.R, taxonomy helpers in R/utils-taxonomy.R, and module evaluation and rendering helpers in R/utils-render.R. The remaining R/utils.R contains general utilities shared across the package.

**Code and module evaluation**
The reviewer was correct that .eval() was unnecessary for the lubridate::parse_date_time() call. I replaced it with a direct namespaced call.
The essential use of .eval() comes from the modular architecture. Report modules are YAML files that may contain executable R code as text. .eval() is a small internal wrapper around parse() and eval() that evaluates this code in the environment explicitly supplied by the caller, rather than in the user's global environment. It is also reused for some predefined internal expressions and optional-package calls that need local symbols without attaching namespaces.
User-authored modules are executable code by design and must therefore be treated as trusted code. This is documented in the contributing guide and in the comment above .eval() in [R/utils-render.R](https://github.com/spatialecology/camtrapReport/blob/main/R/utils-render.R).
I removed the duplicate definitions and split spatial, taxonomy, and rendering or module-evaluation helpers into the three topic-specific files described above. R/utils.R remains substantial because it now contains the general string, date and time, file, formatting, and package helpers shared by several parts of the package..make_render_env() builds the environment in which YAML module code is rendered. It exposes the camReport object under the historical aliases used by existing modules, copies the required object fields and formatting helpers, and keeps module assignments outside the user's global environment. I documented this role in the contributing guide and added an explanatory comment above the function in R/utils-render.R.

**Temporary files and cache location**
Thank you for pointing this out. I audited the uses of `tempdir()`, `tempfile()`, and package-created temporary directories in the package code and tests.
I revised the internal temporary-file handling so that temporary working files and directories are removed after use, including when an error occurs. This includes the files used for internal report-section testing, the temporary directory used when testing report modules after a rendering failure, temporary HTML-widget files, and files copied during a GUI session. GUI temporary files are now scoped to the Shiny session and removed when the session ends. I also added focused tests for the relevant cleanup behaviour.
I also reviewed the distinction between temporary files and caching. There is no package cache intended to persist across R sessions in `tempdir()`, so I did not introduce a persistent cache under `tools::R_user_dir()`. The `__camReport_Object.rds` file created by `camData()` is different: it is dataset-specific persisted state and is deliberately stored with the extracted Camtrap DP dataset so that the same dataset can be reopened without recreating the object. I retained this dataset-local behaviour.
The temporary HTML returned by `testSection()` is also retained because returning that file is part of the documented behaviour of the function; its intermediate files are removed.

**Tests**
I replaced the three assertions with expect_no_match(), and jarl now passes. I did not force a one-to-one mapping between test files and files in R/. Some source files contain several broad components, so tests are divided by behavior or component to keep each test file focused and manageable.
Yes. [tests/testthat/test-gui.R](https://github.com/spatialecology/camtrapReport/blob/main/tests/testthat/test-gui.R) tests application construction, reactive server outputs, editing and saving settings, loading an existing camReport object, and launching the application. The server logic is tested with shiny::testServer(); browser-based visual regression testing is not currently included.

Thank you again for your helpful feedback.

Best wishes,
Elham
