### 2026-09-06 62e353c
Update README quick-start output
