### 2026-09-06 6b53941
Update README and rebuild pkgdown site
