### 2026-09-07 f65a682
Clarify README and module documentation following reviewer feedback
