+++ tests/testthat/test-supporting-functions.R
@@ -76,7 +76,9 @@ test_that("base left join handles equal and differently named keys", {
   expect_error(.left_join(left, right, "missing"))
 })
 
-test_that("the internal pivot helper supports tidy-style column specifications", {
+test_that(
+  "the internal pivot helper supports tidy-style column specifications",
+  {
   data <- data.frame(
     location = c("A", "A", "B"),
     species = c("fox", "hare", "fox"),
@@ -109,4 +111,5 @@ test_that("the internal pivot helper supports tidy-style column specifications",
     .pivot_wider(data, location, unknown, count),
     "Unknown column"
   )
-})
+  }
+)
