+++ R/info.R
@@ -12,7 +12,6 @@
 )
 
 
-
 setGeneric(
   "info",
   function(x, name) {
@@ -94,36 +93,30 @@ setMethod(
   "info",
   signature(x = "camReport"),
   function(x, name) {
-    
     if (missing(name)) {
       name <- NULL
     }
-    
+
     .inf <- list()
-    
+
     if (is.null(name)) {
-      
       name <- .default_cam_info_names
-      
     } else {
-      
       name <- name[name %in% names(camR$fields())]
-      
+
       if (length(name) == 0) {
         warning(
-          paste0(
-            "The specified name(s) are not identified or available in the ",
-            "camReport object; the default fields are used."
-          )
+          "The specified name(s) are not identified or available in the ",
+          "camReport object; the default fields are used."
         )
         name <- .default_cam_info_names
       }
     }
-    
+
     for (n in name) {
       .inf[[n]] <- x[[n]]
     }
-    
+
     class(.inf) <- "camInfo"
     .inf
   }
@@ -141,67 +134,64 @@ setReplaceMethod(
   "info",
   signature(x = "camReport"),
   function(x, name, value) {
-    
     if (
       missing(name) ||
-      length(name) != 1L ||
-      is.na(name) ||
-      !nzchar(trimws(name))
+        length(name) != 1L ||
+        is.na(name) ||
+        !nzchar(trimws(name))
     ) {
       stop(
         "'name' must be one non-empty character string.",
         call. = FALSE
       )
     }
-    
+
     name <- trimws(as.character(name))
     name_lower <- tolower(name)
-    
-    if (name_lower %in% c(
-      "introduction",
-      "study area",
-      "image processing",
-      "sampling efforts",
-      "sampling effort",
-      "acknowledgements",
-      "acknowledgement"
-    )) {
-      
+
+    if (
+      name_lower %in%
+        c(
+          "introduction",
+          "study area",
+          "image processing",
+          "sampling efforts",
+          "sampling effort",
+          "acknowledgements",
+          "acknowledgement"
+        )
+    ) {
       if (name_lower == "study area") {
         x$description <- value
         return(invisible(x))
       }
-      
+
       if (name_lower == "introduction") {
         return(updateReportSection(x, "introduction", text = value))
       }
-      
+
       if (name_lower == "image processing") {
         return(updateReportSection(x, "image processing", text = value))
       }
-      
+
       if (name_lower %in% c("sampling efforts", "sampling effort")) {
         return(updateReportSection(x, "sampling efforts", text = value))
       }
-      
+
       if (name_lower %in% c("acknowledgements", "acknowledgement")) {
         return(updateReportSection(x, "acknowledgements", text = value))
       }
-      
     } else {
-      
       .f <- camR$fields()
       name <- name[name %in% names(.f)]
-      
+
       if (length(name) == 0) {
         stop(
-          paste0(
-            "The specified name is not identified or available in the ",
-            "camReport object."
-          )
+          "The specified name is not identified or available in the ",
+          "camReport object."
         )
       }
-      
+
       x[[name]] <- value
       return(invisible(x))
     }
