+++ tests/testthat/test-coverage-regressions.R
@@ -1,0 +1,502 @@
+make_sampling_text_report <- function(
+  sampling_design = NULL,
+  camera_model = NULL,
+  bait_use = NULL,
+  camera_height = NULL,
+  capture_method = NULL,
+  individual_animals = NULL,
+  is_eow = FALSE
+) {
+  object <- camR$new()
+  object$info <- list(
+    is.EOW = is_eow,
+    json = list(
+      project = list(
+        samplingDesign = sampling_design,
+        captureMethod = capture_method,
+        individualAnimals = individual_animals
+      )
+    )
+  )
+  object$data <- list(
+    deployments = data.frame(
+      cameraModel = camera_model,
+      baitUse = bait_use,
+      cameraHeight = camera_height,
+      stringsAsFactors = FALSE
+    )
+  )
+  object$reportTextElements <- list()
+  object
+}
+
+
+test_that("sampling text describes supported survey configurations", {
+  mixed <- make_sampling_text_report(
+    sampling_design = c("simpleRandom", "targeted", "customDesign"),
+    camera_model = c("A", "B", "C", "D"),
+    bait_use = c(TRUE, FALSE, TRUE, FALSE),
+    camera_height = c(0.5, 1, 1.5, 2),
+    capture_method = c("motionDetection", "timeLapse", "audio"),
+    individual_animals = c(TRUE, FALSE),
+    is_eow = TRUE
+  )
+
+  .get_sampling_text(mixed)
+  expect_match(mixed$reportTextElements$sampling, "EOW camera-trap protocol")
+  expect_match(mixed$reportTextElements$sampling, "combines")
+  expect_match(mixed$reportTextElements$sampling, "Multiple camera models")
+  expect_match(mixed$reportTextElements$sampling, "mixture of baited")
+  expect_match(mixed$reportTextElements$sampling, "ranging from 0.5 to 2")
+  expect_match(mixed$reportTextElements$sampling, "motion detection")
+  expect_match(mixed$reportTextElements$sampling, "both the identification")
+
+  single <- make_sampling_text_report(
+    sampling_design = "simpleRandom",
+    camera_model = "ToyCam",
+    bait_use = FALSE,
+    camera_height = 0.75,
+    capture_method = "activityDetection",
+    individual_animals = FALSE
+  )
+
+  .get_sampling_text(single)
+  expect_match(single$reportTextElements$sampling, "simple random")
+  expect_match(single$reportTextElements$sampling, "ToyCam")
+  expect_match(single$reportTextElements$sampling, "No bait")
+  expect_match(single$reportTextElements$sampling, "0.75 m")
+  expect_match(single$reportTextElements$sampling, "activity detection")
+  expect_match(single$reportTextElements$sampling, "not specifically")
+
+  paired <- make_sampling_text_report(
+    sampling_design = c("experimental", "opportunistic"),
+    camera_model = c("A", "B"),
+    bait_use = TRUE,
+    camera_height = c(0.5, 1),
+    capture_method = "audio",
+    individual_animals = TRUE
+  )
+
+  .get_sampling_text(paired)
+  expect_match(paired$reportTextElements$sampling, "experimental and")
+  expect_match(paired$reportTextElements$sampling, "Bait was used")
+  expect_match(paired$reportTextElements$sampling, "audio recording")
+  expect_match(paired$reportTextElements$sampling, "designed to support")
+
+  unknown <- make_sampling_text_report(
+    sampling_design = "customDesign"
+  )
+  unknown$data$deployments <- data.frame(dummy = 1)
+
+  .get_sampling_text(unknown)
+  expect_match(unknown$reportTextElements$sampling, "customDesign")
+})
+
+
[…]
