+++ R/install_all.R
@@ -5,10 +5,10 @@
 .detachPackage <- function(n, unload = TRUE, force = TRUE) {
   n <- unique(as.character(n))
   n <- n[!is.na(n) & nzchar(n)]
-
+  
   for (pkg in n) {
     package_name <- paste0("package:", pkg)
-
+    
     if (package_name %in% search()) {
       try(
         # Detaching is required before pak can update an attached namespace.
@@ -24,7 +24,7 @@
       )
     }
   }
-
+  
   invisible(NULL)
 }
 
@@ -34,9 +34,9 @@
   if (length(n) == 0L) {
     return(setNames(logical(0), character(0)))
   }
-
+  
   n <- as.character(n)
-
+  
   installed <- vapply(
     n,
     function(pkg) {
@@ -46,7 +46,7 @@
     },
     logical(1)
   )
-
+  
   names(installed) <- n
   installed
 }
@@ -69,24 +69,24 @@
 
 .getPackageList <- function() {
   packages <- .get_module_packages()
-
+  
   config_file <- system.file(
     "external",
     "camtrapReportConfig.rds",
     package = "camtrapReport"
   )
-
+  
   if (nzchar(config_file) && file.exists(config_file)) {
     config <- readRDS(config_file)
-
+    
     if (!is.null(config$packages)) {
       packages <- unique(c(packages, config$packages))
     }
   }
-
+  
   packages <- as.character(packages)
   packages <- packages[!is.na(packages) & nzchar(packages)]
-
+  
   unique(packages)
 }
 
@@ -98,17 +98,17 @@
     "camtrapReportConfig.rds",
     package = "camtrapReport"
   )
-
+  
   if (!nzchar(config_file) || !file.exists(config_file)) {
     return(character(0))
   }
-
+  
   config <- readRDS(config_file)
-
+  
   if (is.null(config$github)) {
     return(character(0))
   }
-
+  
   unlist(config$github, use.names = TRUE)
 }
 
@@ -120,17 +120,17 @@
     "camtrapReportConfig.rds",
     package = "camtrapReport"
   )
-
+  
   if (!nzchar(config_file) || !file.exists(config_file)) {
     return(character(0))
   }
-
+  
   config <- readRDS(config_file)
-
+  
   if (is.null(config$gitlab)) {
     return(character(0))
   }
-
+  
   unlist(config$gitlab, use.names = TRUE)
 }
 
@@ -139,18 +139,18 @@
 
 .pakRemoteReferences <- function(repositories, source) {
   source <- match.arg(source, c("github", "gitlab"))
-
+  
   if (length(repositories) == 0L) {
     return(setNames(character(0), character(0)))
   }
-
+  
   package_names <- names(repositories)
   repositories <- as.character(repositories)
-
+  
   if (
     is.null(package_names) ||
-      anyNA(package_names) ||
-      !all(nzchar(package_names))
+    anyNA(package_names) ||
+    !all(nzchar(package_names))
   ) {
     stop(
       sprintf(
@@ -160,9 +160,9 @@
       call. = FALSE
     )
   }
-
+  
   repositories <- trimws(repositories)
-
+  
   if (anyNA(repositories) || !all(nzchar(repositories))) {
     stop(
       sprintf(
@@ -172,7 +172,7 @@
       call. = FALSE
     )
   }
-
+  
   references <- paste0(
     package_names,
     "=",
@@ -180,7 +180,7 @@
     "::",
     repositories
   )
-
+  
   names(references) <- package_names
   references
 }
@@ -191,19 +191,19 @@
   if (length(references) == 0L) {
     return(references)
   }
-
+  
   separator <- ifelse(
     grepl("?", references, fixed = TRUE),
     "&",
[…]
