+++ NAMESPACE
@@ -8,6 +8,7 @@ export(empty_trash)
 export(gui)
 export(info)
 export(install_All)
+export(install_all)
 export(listReportSections)
 export(list_Modules)
 export(move_Module)

[…]
