### 2026-09-08 fdc9681
Rename install_All to install_all and improve package checks
