### 2026-09-08 429f064
Clarify package architecture and evaluation design
