### 2026-09-09 11d468c
Clarify report-module architecture
