### 2026-09-10 e1b2170
Add rOpenSci review badge and update NEWS
